% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoID.R
\name{autoID}
\alias{autoID}
\title{Automatically define measurements' window based on an auxiliary file}
\usage{
autoID(
  inputfile,
  auxfile = NULL,
  obs.length = NULL,
  deadband = 0,
  crop.end = 0,
  shoulder = 60,
  drop = TRUE
)
}
\arguments{
\item{inputfile}{data.frame; output from import or align functions.}

\item{auxfile}{data.frame; auxiliary data frame containing the columns
\code{start.time} and \code{UniqueID}. \code{start.time} must
contain a date and be in POSIXct format. The time zone must be
the same as the POSIX.time in \code{inputfile}. The default
time zone for the import functions is "UTC". A data frame from
the Smart Chamber (LI-8200) can be used as an auxiliary file.
In that case, \code{chamID} will be used instead of \code{UniqueID},
if \code{UniqueID} cannot be found.}

\item{obs.length}{numerical; chamber closure time (seconds). Default is NULL.
If \code{obs.length} is not provided, a column \code{obs.length}
should be contained in \code{auxfile} or \code{inputfile}.
Alternatively, \code{obs.length} will be calculated from
\code{start.time} and \code{cham.open} or \code{end.time}
if found in \code{auxfile} or \code{inputfile}.}

\item{deadband}{numerical; time (seconds) at the start of the measurement,
after chamber closure, to be removed from the time stamps
used to estimate the flux.}

\item{crop.end}{numerical; time (seconds) at the end of the measurement,
before chamber opening, to be removed from the time stamps
used to estimate the flux.}

\item{shoulder}{numerical; time before and after measurement in observation
window (seconds). Default is 60 seconds. If the time between
two measurements is smaller than \code{shoulder}, the time
stamps are duplicated.}

\item{drop}{logical; if \code{drop = TRUE} (default), time stamps without an
assigned \code{UniqueID} are removed from the final data frame.}
}
\value{
A merged data.frame between the \code{inputfile} and the \code{auxfile},
         with the additional columns \code{flag}, \code{Etime} and \code{obs.length}.
}
\description{
Uses an auxiliary file to identify start and end time of measurements, add
a \code{UniqueID} per measurement, and merge other auxiliary data required
with the \code{\link[goFlux]{goFlux}} function.
}
\details{
Note that, if the time between two measurements is smaller than \code{shoulder},
the time stamps are duplicated, which can result in an output data frame
larger (more rows) than the \code{inputfile}. Conversely, if \code{drop = TRUE},
time stamps without an assigned \code{UniqueID} are removed from the final data
frame, which can result in an output data frame smaller (less rows) than the
\code{inputfile}.
}
\examples{
# How to use in multiple situations:
library(dplyr)

## with a LGR instrument and an auxiliary file (.txt)
aux.path <- system.file("extdata", "aux_UGGA/aux_UGGA.txt", package = "goFlux")
auxfile <- read.delim(aux.path) \%>\%
  mutate(start.time = as.POSIXct(start.time, tz = "UTC"))
data(imp.UGGA)
autoID.UGGA <- autoID(inputfile = imp.UGGA, auxfile = auxfile, obs.length = 180)

## with a LI-COR instrument and the Smart Chamber as auxiliary file
data(imp.LI8200)
data(imp.LI7820)
autoID.LI7820 <- autoID(inputfile = imp.LI7820, deadband = 20, auxfile = imp.LI8200)

## with the LI-6400 and no auxiliary file
data(imp.LI6400)
ow.LI6400 <- obs.win(inputfile = imp.LI6400, shoulder = 0)

}
\seealso{
For a manual identification of start and end of gas measurements,
see the functions \code{\link[goFlux]{obs.win}} and
\code{\link[goFlux]{click.peak2}}.
}
