% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.meas.R
\name{crop.meas}
\alias{crop.meas}
\title{Manual editing of observation window of flux measurements}
\usage{
crop.meas(
  dataframe,
  auxfile = NULL,
  crop_start = 0,
  crop_end = 0,
  deadband = 0,
  max.obs.length = NULL,
  cham.close = "start.time",
  cham.open = "end.time"
)
}
\arguments{
\item{dataframe}{a dataframe containing gas measurements, output from
import or ID functions (click.peak2 or manID). Must contain
the columns \code{UniqueID}, \code{start.time},
\code{end.time} and \code{Etime}.}

\item{auxfile}{data.frame; auxiliary data frame containing the columns
\code{UniqueID}, \code{crop_start}, \code{crop_end},
\code{deadband} and/or \code{max.obs.length}.}

\item{crop_start}{numerical; number of observations to remove at the start
of the measurement. To provide a unique value per
UniqueID, the column \code{crop_start} must be present
in the auxfile and the argument must be set to
\code{crop_start = "aux"}.}

\item{crop_end}{numerical; number of observations to remove at the
end of the measurement. To provide a unique value per
UniqueID, the column \code{crop_end} must be present
in the auxfile and the argument must be set to
\code{crop_end = "aux"}.}

\item{deadband}{numerical; define a deadband at the start of measurements
(seconds). To provide a unique value per UniqueID, the column
\code{deadband} must be present in the auxfile and the
argument must be set to \code{deadband = "aux"}.}

\item{max.obs.length}{numerical; maximum length of measurement (seconds) to
keep after \code{cham.close}. To provide a unique value
per UniqueID, the column \code{max.obs.length} must be
present in the auxfile and the argument must be set to
\code{max.obs.length = "aux"}.}

\item{cham.close}{character; if \code{cham.close = "start.time"} (default), the
original \code{start.time} before cropping will be saved into
a new column \code{cham.close}. Set \code{cham.close = NULL}
to drop the original \code{start.time}.}

\item{cham.open}{character; if \code{cham.open = "end.time"} (default), the
original \code{end.time} before cropping will be saved into
a new column \code{cham.open}. Set \code{cham.open = NULL}
to drop the original \code{end.time}.}
}
\value{
a data frame, with corrected Etime and flag, based on newly defined
         start.time, end.time and deadband.
}
\description{
Crop measurements from the start, from the end, or adjust the deadband.
Individual parameters can be provided for each measurement, in an auxiliary
file, or the same parameters can be applied to all measurements.
}
\details{
Note that \code{crop_start} and \code{deadband} "crop" the start of the
measurements in the same way.

Note that if \code{max.obs.length} is defined, \code{end.time} is calculated
based on \code{cham.close} (original \code{start.time} in \code{dataframe}, if
\code{cham.close = "start.time"}) and \code{max.obs.length}, regardless of other
defined values in the function. If \code{cham.close = NULL}, \code{end.time}
is calculated based on \code{start.time} and \code{max.obs.length}.

If \code{cham.close} and \code{cham.open} are originally present in
\code{dataframe}, set arguments \code{cham.close} and \code{cham.open}
to \code{NULL}.

Note that the arguments \code{crop_start} and \code{crop_end} accept negative
values and can be used to include more data points before and after the
previously defined \code{start.time} and \code{end.time}.
}
\examples{
# Load example data
data(manID.G2201i)

# Create an auxfile to define different crop_start and crop_end per UniqueID
auxfile <- cbind.data.frame(UniqueID = c("meas1", "meas2"),
                            crop_start = c(2,10),
                            crop_end = c(5,0))

# The arguments "crop_start" and "crop_end" must be set to "aux" to use
# unique values from the auxfile
crop.G2201i <- crop.meas(dataframe = manID.G2201i, auxfile = auxfile,
                         crop_start = "aux", crop_end = "aux")

# Alternatively, cropping can be applied equally across all measurements
crop.G2201i <- crop.meas(dataframe = manID.G2201i, auxfile = auxfile,
                         crop_start = 5, crop_end = 5)
}
