% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGM5_import.R
\name{EGM5_import}
\alias{EGM5_import}
\title{Import function for the PP-Systems EGM-5 Portable
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} Gas Analyzer}
\usage{
EGM5_import(
  inputfile,
  date.format = "dmy",
  timezone = "UTC",
  save = FALSE,
  keep_all = FALSE,
  prec = c(3, 1, 500)
)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .TXT}

\item{date.format}{character string; specifies the date format found in the
raw data file. Choose one of the following: "dmy", "ymd",
or "mdy". Default is "dmy", as it is the date format from
the example data file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if \code{save = TRUE}, saves the file as an .RData file
in a RData folder in the current working directory. If
\code{save = FALSE}, returns the file in the Console, or load in
the Environment if assigned to an object.}

\item{keep_all}{logical; if \code{keep_all = TRUE}, keep all columns from the raw
file. The default is \code{keep_all = FALSE}, and columns that
are not necessary for gas flux calculation are removed.}

\item{prec}{numerical vector; the precision of the instrument for each gas,
in the following order: "CO2dry_ppm", "O2dry_pct" and H2O_ppm".
The default is \code{prec = c(3, 1, 500)}.}
}
\value{
A data frame containing raw data from the PP-Systems EGM-5 Portable
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} Gas Analyzer.
}
\description{
Imports single raw gas measurement files from the PP-Systems EGM-5 Portable
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} Gas Analyzer, with
the extension .TXT
}
\details{
In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For the instrument EGM5, the date
is found in the column "Date".

Note that this function was designed for the following units in the raw file:
\itemize{
  \item ppm for \ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}}
  \item \% for \ifelse{html}{\out{O<sub>2</sub>}}{\eqn{O[2]}{ASCII}} (pct)
  \item mb for \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}
  \item mb for pressure
  \item Celsius for temperature}
If your PP-Systems EGM-5 uses different units, either convert the units after
import, change the settings on your instrument, or contact the maintainer of
this package for support.

The precision of the instrument is needed to restrict kappa-max
(\code{\link[GoFluxYourself]{k.max}}) in the non-linear flux calculation
(\code{\link[GoFluxYourself]{HM.flux}}). Kappa-max is inversely proportional to
instrument precision. If the precision of your instrument is unknown, it is
better to use a low value (e.g. 1 ppm for
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}, or 1\% for
\ifelse{html}{\out{O<sub>2</sub>}}{\eqn{O[2]}{ASCII}}) to allow for more
curvature, especially for water vapor fluxes, or very long measurements, that
are normally curved. The default values given for instrument precision are
the ones provided by the manufacturer upon request, at the time of the
creation of this package (11-2023).
}
\examples{
# Examples on how to use:
file.path <- system.file("extdata", "EGM5/EGM5.TXT", package = "GoFluxYourself")

EGM5_imp <- EGM5_import(inputfile = file.path)

}
\seealso{
Use the wrapper function \code{\link[GoFluxYourself]{import2RData}}
         to import multiple files from the same folder path using any instrument.

See also, import functions for other instruments:
         \code{\link[GoFluxYourself]{DX4015_import}},
         \code{\link[GoFluxYourself]{G2508_import}},
         \code{\link[GoFluxYourself]{G4301_import}},
         \code{\link[GoFluxYourself]{GAIA_import}},
         \code{\link[GoFluxYourself]{LGR_import}},
         \code{\link[GoFluxYourself]{LI6400_import}},
         \code{\link[GoFluxYourself]{LI7810_import}},
         \code{\link[GoFluxYourself]{LI7820_import}},
         \code{\link[GoFluxYourself]{LI8100_import}},
         \code{\link[GoFluxYourself]{LI8200_import}},
         \code{\link[GoFluxYourself]{N2OM1_import}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
