% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_LI7810_imp-data.R
\docType{data}
\name{example_LI7810_imp}
\alias{example_LI7810_imp}
\title{Example gas measurements data from LI-COR instruments}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(example_LI7810_imp)
}
\description{
Raw data from the LI-7810 greenhouse gas analyzer that has
been processed by the function LI7810_import of this package.
}
\examples{
data(example_LI7810_imp)
head(example_LI7810_imp)

}
\keyword{datasets}
