% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_LI8200_imp-data.R
\docType{data}
\name{example_LI8200_imp}
\alias{example_LI8200_imp}
\title{Example gas measurements data from LI-COR instruments}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(example_LI8200_imp)
}
\description{
Raw data from the LI-8200 greenhouse gas analyzer that has
been processed by the function \code{\link[GoFluxYourself]{LI8200_import}}
of this package.
}
\examples{
data(example_LI8200_imp)
head(example_LI8200_imp)

}
\keyword{datasets}
