% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G4301_import.R
\name{G4301_import}
\alias{G4301_import}
\title{Import function for the Picarro
\ifelse{html}{\out{GasScouter<sup>TM</sup>}}{\eqn{GasScouter^{TM}}{ASCII}}
G4301 Mobile Gas Concentration Analyzer}
\usage{
G4301_import(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  save = FALSE,
  keep_all = FALSE,
  prec = c(0.025, 0.1, 10)
)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .dat}

\item{date.format}{character string; specifies the date format found in the
raw data file. Choose one of the following: "dmy", "ymd",
or "mdy". Default is "ymd", as it is the date format from
the example data file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if \code{save = TRUE}, saves the file as RData in a
RData folder in the current working directory. If
\code{save = FALSE}, returns the file in the Console, or load in
the Environment if assigned to an object.}

\item{keep_all}{logical; if \code{keep_all = TRUE}, keep all columns from raw
file. The default is \code{keep_all = FALSE}, and columns that
are not necessary for gas flux calculation are removed.}

\item{prec}{numerical vector; the precision of the instrument for each gas,
in the following order: "CO2dry_ppm", CH4dry_ppb" and H2O_ppm".
The default is \code{prec = c(0.025, 0.1, 10)}.}
}
\value{
A data frame containing raw data from Picarro G4301 GHG analyzer.
}
\description{
Imports single raw gas measurement files from the Picarro G4301 with the
extension .dat (\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}},
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}} GHG analyzer)
}
\details{
In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For the instrument G4301, the date
is found in the column "DATE".

Note that this function was designed for the following units in the raw file:
\itemize{
  \item ppm for \ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}}
  \item ppb for \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}
  \item mmol/mol for \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}
  \item Torr for pressure
  \item Celsius for temperature}
If your instrument uses different units, either convert the units after import,
change the settings on your instrument, or contact the maintainer of this
package for support.

The precision of the instrument is needed to restrict kappa-max
(\code{\link[GoFluxYourself]{k.max}}) in the non-linear flux calculation
(\code{\link[GoFluxYourself]{HM.flux}}). Kappa-max is inversely proportional to
instrument precision. If the precision of your instrument is unknown, it is
better to use a low value (e.g. 1 ppm for
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}, or 1 ppb for
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}) to allow for more
curvature, especially for water vapor fluxes, or very long measurements, that
are normally curved. The default values given for instrument precision are
the ones found for the latest model of this instrument, available at the
time of the creation of this package (11-2023).
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "G4301/07_2022/G4301.dat", package = "GoFluxYourself")

# Run function
G4301_imp <- G4301_import(inputfile = file.path)

}
\seealso{
Use the wrapper function \code{\link[GoFluxYourself]{import2RData}}
         to import multiple files from the same folder path using any instrument.

See also, import functions for other instruments:
         \code{\link[GoFluxYourself]{DX4015_import}},
         \code{\link[GoFluxYourself]{EGM5_import}},
         \code{\link[GoFluxYourself]{G2508_import}},
         \code{\link[GoFluxYourself]{GAIA_import}},
         \code{\link[GoFluxYourself]{LGR_import}},
         \code{\link[GoFluxYourself]{LI6400_import}},
         \code{\link[GoFluxYourself]{LI7810_import}},
         \code{\link[GoFluxYourself]{LI7820_import}},
         \code{\link[GoFluxYourself]{LI8100_import}},
         \code{\link[GoFluxYourself]{LI8200_import}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
