% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAIA_import.R
\name{GAIA_import}
\alias{GAIA_import}
\title{Import function for GAIA synced with multiple chambers and instruments}
\usage{
GAIA_import(inputfile, date.format = "ymd", timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{the name of a file with the extension .csv}

\item{date.format}{Date format. Chose one of the following: "dmy", "ymd", or "mdy".
Default is "ymd" as it is the date format from the example
data file provided.}

\item{timezone}{a time zone in which to import the data to POSIXct format.
Default is "UTC". Note about time zone: I recommend using
the time zone "UTC" to avoid any issue related to summer
time and winter time changes.}

\item{save}{logical. If save = TRUE, save the file as Rdata in the current
working directory. If save = FALSE, return the file in the Console,
or load in the Environment if assigned to an object.}
}
\value{
a data frame
}
\description{
Imports single raw gas measurement files from the automated chamber
ECOFlux (GAIA2TECH) with the extension .csv
(LI-7810: CO2, CH4 and H2O / LI7820: N2O and H2O)
}
\examples{
# Examples on how to use:
file.path <- system.file("extdata", "GAIA/example_GAIA.csv", package = "GoFluxYourself")

# 1. If you want to import the data and work with it directly
GAIA.data <- GAIA_import(inputfile = file.path)

# 2. Or if you want to import the data and save it as Rdata
GAIA_import(inputfile = file.path, save = TRUE)
}
