% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goFlux.nokmax.R
\name{goFlux.nokmax}
\alias{goFlux.nokmax}
\title{goFlux (without kappa-max): a user-friendly GHG fluxes calculation tool}
\usage{
goFlux.nokmax(
  dataframe,
  gastype,
  H2O_col = "H2O_ppm",
  prec = NULL,
  Area = NULL,
  offset = NULL,
  Vtot = NULL,
  Vcham = NULL,
  Pcham = NULL,
  Tcham = NULL,
  warn.length = 60
)
}
\arguments{
\item{dataframe}{a data.frame containing gas measurements (see \code{gastype}
below), water vapor measurements (see \code{H2O_col} below)
and the following columns: \code{UniqueID}, \code{Etime},
\code{Vtot}, \code{Area}, \code{Pcham}, \code{Tcham} and
\code{flag} (see the parameters \code{Vtot}, \code{Area},
\code{Pcham} and \code{Tcham} below for more details).
\code{chamID} may be used instead of \code{UniqueID}.}

\item{gastype}{character string; specifies which column should be used for the
flux calculations. Must be one of the following: "CO2dry_ppm",
"CH4dry_ppb", "COdry_ppb", "N2Odry_ppb", "NH3dry_ppb" or "H2O_ppm".}

\item{H2O_col}{character string; specifies which column should be used to
subtract the effect of water vapor in the chamber space.
Default is \code{H2O_col = "H2O_ppm"}.}

\item{prec}{numerical value; precision of the instruments. Units must be the
same as \code{gastype}. With the default \code{prec = NULL},
instrument precision for each gas must be provided in
\code{dataframe}.}

\item{Area}{numerical value; area of the soil surface inside the chamber
\ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}}.
Alternatively, provide the column \code{Area} in \code{dataframe}
if \code{Area} is different between samples.}

\item{offset}{(optional) numerical value; height between the soil surface and
the chamber (cm). Alternatively, provide the column \code{offset}
in \code{dataframe} if \code{offset} is different between samples.
\code{offset} is only used if \code{Vtot} is missing.}

\item{Vtot}{numerical value; total volume inside the chamber, tubes, instruments,
etc. (L). Alternatively, provide the column \code{Vtot} in
\code{dataframe} if \code{Vtot} is different between samples. If
\code{Vtot} is missing, the function will calculate it from
\code{Area}, \code{Vcham} and \code{offset}.}

\item{Vcham}{(optional) numerical value; volume inside the chamber, tubes and
instruments (L). Alternatively, provide the column \code{Vcham}
in \code{dataframe} if \code{Vcham} is different between samples.
\code{Vhcam} is only used if \code{Vtot} is missing.}

\item{Pcham}{numerical value; pressure inside the chamber (kPa).
Alternatively, provide the column \code{Pcham} in \code{dataframe}
if \code{Pcham} is different between samples. If \code{Pcham} is
not provided, normal atmospheric pressure (101.325 kPa) is used.}

\item{Tcham}{numerical value; temperature inside the chamber (Celsius).
Alternatively, provide the column \code{Tcham} in \code{dataframe}
if \code{Tcham} is different between samples. If \code{Tcham} is
not provided, 15°C is used as default.}

\item{warn.length}{numerical value; limit under which a measurement is
flagged for being too short (\code{nb.obs < warn.length}).}
}
\value{
Returns a data frame with 28 columns: a \code{UniqueID} per
measurement, 11 columns for the linear model results (linear flux estimate
(\code{\link[goFlux]{LM.flux}}), initial gas concentration
(\code{LM.C0}), final gas concentration (\code{LM.Ct}), slope of linear
regression (\code{LM.slope}), mean absolute error (\code{LM.MAE}), root mean
square error (\code{LM.RMSE}), Akaike's information criterion corrected for
small sample size (\code{LM.AICc}), standard error (\code{LM.SE}), relative
standard error (\code{LM.se.rel}), coefficient of determination (\code{LM.r2}),
and \emph{p-value} (\code{LM.p.val})), 11 columns for the non-linear model
results (non-linear flux estimate (\code{\link[goFlux]{HM.flux}}),
initial gas concentration (\code{HM.C0}), the assumed concentration of
constant gas source below the surface (\code{HM.Ci}), slope at \code{t=0}
(\code{HM.slope}), mean absolute error (\code{HM.MAE}), root mean square error
(\code{HM.RMSE}), Akaike's information criterion corrected for small sample
size (\code{HM.AICc}), standard error (\code{HM.SE}), relative standard error
(\code{HM.se.rel}), coefficient of determination (\code{HM.r2}), and curvature
(kappa; \code{HM.k}), as well as the minimal detectable flux
(\code{\link[goFlux]{MDF}}), the precision of the instrument
(\code{prec}), the flux term (\code{\link[goFlux]{flux.term}}),
kappa-max (\code{\link[goFlux]{k.max}}), the g factor (g.fact;
\code{\link[goFlux]{g.factor}}), the number of observations used
(\code{nb.obs}) and the true initial gas concentration (\code{C0}) and final
gas concentration (\code{Ct}).
}
\description{
A wrapper function to calculate GHG fluxes from static chamber measurements.
Calculates linear (\code{\link[goFlux]{LM.flux}}) and non-linear
fluxes (Hutchinson and Mosier model; \code{\link[goFlux]{HM.flux}}),
from a variety of greenhouse gasses (
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}},
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}},
\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}},
\ifelse{html}{\out{NH<sub>3</sub>}}{\eqn{NH[3]}{ASCII}}, CO, and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}).
}
\details{
This function differs from the original version
\code{\link[goFlux]{goFlux}}. In this version, the non-linear model
is calculated without kappa-max (\code{\link[goFlux]{k.max}}). In the
original version, the kappa-max is estimated using the
\code{\link[goFlux]{MDF}}.

Flux estimate units are
\ifelse{html}{\out{µmol m<sup>-2</sup>s<sup>-1</sup>}}{\eqn{µmol m^{-2}s^{-1}}{ASCII}}
(if initial concentration is ppm, e.g. CO2dry_ppm) and
\ifelse{html}{\out{nmol m<sup>-2</sup>s<sup>-1</sup>}}{\eqn{nmol m^{-2}s^{-1}}{ASCII}}
(if initial concentration is ppb, e.g. CH4dry_ppb).

The \code{\link[goFlux]{goFlux}} function calculates flux estimates
from the linear model (LM) and the Hutchinson and Mosier model (HM). The HM
model is a non-linear model, whose curvature is controlled by the parameter
kappa. A large kappa returns a strong curvature. A maximum threshold for this
parameter, kappa-max (\code{\link[goFlux]{k.max}}), can be calculated
from the linear flux estimate (\code{\link[goFlux]{LM.flux}}), the
minimal detectable flux (\code{\link[goFlux]{MDF}}) and the time of
chamber closure. This limit of kappa-max is included in the
\code{\link[goFlux]{goFlux}} function, so that the non-linear flux
estimate cannot exceed this maximum curvature.

All flux estimates, including the \code{\link[goFlux]{MDF}}, are
obtained from the multiplication of the slope and the
\code{\link[goFlux]{flux.term}}, which is used to correct the
pressure inside the chamber (due to water vapor), as well as convert the
units to obtain a term in nmol or
\ifelse{html}{\out{µmol m<sup>-2</sup>s<sup>-1</sup>}}{\eqn{µmol m^{-2}s^{-1}}{ASCII}}.

The argument \code{Area} is in \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}},
but the output units from \code{\link[goFlux]{goFlux}} are in
\ifelse{html}{\out{(m<sup>2</sup>)}}{\eqn{(m^2)}{ASCII}}. This is due to the
conversion from \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}}
to \ifelse{html}{\out{(m<sup>2</sup>)}}{\eqn{(m^2)}{ASCII}} within the
function. This means that there is a factor of 10,000 to convert from
\ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}}
to \ifelse{html}{\out{(m<sup>2</sup>)}}{\eqn{(m^2)}{ASCII}}. This is important
to take into account if one would provide something else than an \code{Area}
in \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}} to the function.
For example, with incubated soil samples, one may provide an amount of soil
(kg) instead of an area in the column \code{Area}. To get the right units in
that case, multiply the kilograms of soil by 10,000 to remove the conversion
from \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}} to
\ifelse{html}{\out{(m<sup>2</sup>)}}{\eqn{(m^2)}{ASCII}}.

In \code{gastype}, the gas species listed are the ones for which this package
has been adapted. Please write to the maintainer of this package for
adaptation of additional gases.

#' \code{warn.length} is the limit below which the chamber closure time is
flagged for being too short (\code{nb.obs < warn.length}). Portable
greenhouse gas analyzers typically measure at a frequency of 1 Hz. Therefore,
for the default setting of \code{warn.length = 60}, the chamber closure time
should be approximately one minute (60 seconds). If the number of
observations is smaller than the threshold, a warning is printed: e.g. "Low
number of observations: UniqueID X has 59 observations".
}
\examples{
data(manID.UGGA)
CO2_flux <- goFlux.nokmax(manID.UGGA, "CO2dry_ppm")
CH4_flux <- goFlux.nokmax(manID.UGGA, "CH4dry_ppb")
H2O_flux <- goFlux.nokmax(manID.UGGA, "H2O_ppm")

}
\references{
Hüppi et al. (2018). Restricting the nonlinearity parameter in
soil greenhouse gas flux calculation for more reliable flux estimates.
\emph{PloS one}, 13(7), e0200876.

Hutchinson and Mosier (1981). Improved soil cover method for
field measurement of nitrous oxide fluxes.
\emph{Soil Science Society of America Journal}, 45(2), 311-316.
}
\seealso{
Look up the functions \code{\link[goFlux]{MDF}},
         \code{\link[goFlux]{flux.term}},
         \code{\link[goFlux]{g.factor}},
         \code{\link[goFlux]{k.max}},
         \code{\link[goFlux]{HM.flux}} and
         \code{\link[goFlux]{LM.flux}} of this package for more
         information about these parameters.

See also the import functions to find the
         default instrument precision for each instrument:
         \code{\link[goFlux]{import.DX4015}},
         \code{\link[goFlux]{import.EGM5}},
         \code{\link[goFlux]{import.G2508}},
         \code{\link[goFlux]{import.G4301}},
         \code{\link[goFlux]{import.GAIA}},
         \code{\link[goFlux]{import.LI6400}},
         \code{\link[goFlux]{import.LI7810}},
         \code{\link[goFlux]{import.LI7820}},
         \code{\link[goFlux]{import.LI8100}},
         \code{\link[goFlux]{import.LI8200}},
         \code{\link[goFlux]{import.N2OM1}},
         \code{\link[goFlux]{import.uCH4}},
         \code{\link[goFlux]{import.uN2O}},
         \code{\link[goFlux]{import.UGGA}}
}
