% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goFlux_nokmax.R
\name{goFlux_nokmax}
\alias{goFlux_nokmax}
\title{goFlux (without kappa-max): a user-friendly GHG fluxes calculation tool}
\usage{
goFlux_nokmax(
  dataframe,
  gastype,
  H2O_col = "H2O_ppm",
  prec = NULL,
  Area = NULL,
  offset = NULL,
  Vtot = NULL,
  Vcham = NULL,
  Pcham = NULL,
  Tcham = NULL,
  warn.length = 60
)
}
\arguments{
\item{dataframe}{a data.frame containing gas measurements (see \code{gastype}
below), water vapor measurements (see \code{H2O_col} below)
and the following columns: \code{UniqueID}, \code{Etime},
\code{Vtot}, \code{Area}, \code{Pcham}, \code{Tcham} and
\code{flag} (see the parameters \code{Vtot}, \code{Area},
\code{Pcham} and \code{Tcham} below for more details).
\code{chamID} may be used instead of \code{UniqueID}.}

\item{gastype}{character string; specifies which column should be used for the
flux calculations. Must be one of the following: "CO2dry_ppm",
"CH4dry_ppb", "COdry_ppb", "N2Odry_ppb", "NH3dry_ppb" or "H2O_ppm".}

\item{H2O_col}{character string; specifies which column should be used to
subtract the effect of water vapor in the chamber space.}

\item{prec}{numerical value; precision of the instruments. Units must be the
same as \code{gastype}. With the default \code{prec = NULL},
instrument precision for each gas must be provided in
\code{dataframe}.}

\item{Area}{numerical value; area of the soil surface inside the chamber
\ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}}.
Alternatively, provide the column \code{Area} in \code{dataframe}
if \code{Area} is different between samples.}

\item{offset}{(optional) numerical value; height between the soil surface and
the chamber (cm). Alternatively, provide the column \code{offset}
in \code{dataframe} if \code{offset} is different between samples.
\code{offset} is only used if \code{Vtot} is missing.}

\item{Vtot}{numerical value; total volume inside the chamber, tubes, instruments,
etc. (L). Alternatively, provide the column \code{Vtot} in
\code{dataframe} if \code{Vtot} is different between samples. If
\code{Vtot} is missing, the function will calculate it from
\code{Area}, \code{Vcham} and \code{offset}.}

\item{Vcham}{(optional) numerical value; volume inside the chamber, tubes and
instruments (L). Alternatively, provide the column \code{Vcham}
in \code{dataframe} if \code{Vcham} is different between samples.
\code{Vhcam} is only used if \code{Vtot} is missing.}

\item{Pcham}{numerical value; pressure inside the chamber (kPa).
Alternatively, provide the column \code{Pcham} in \code{dataframe}
if \code{Pcham} is different between samples. If \code{Pcham} is
not provided, normal atmospheric pressure (101.325 kPa) is used.}

\item{Tcham}{numerical value; temperature inside the chamber (Celsius).
Alternatively, provide the column \code{Tcham} in \code{dataframe}
if \code{Tcham} is different between samples. If \code{Tcham} is
not provided, normal air temperature (15°C) is used.}

\item{warn.length}{numerical; minimum amount of observations accepted (number
of data points). With nowadays portable greenhouse gas
analyzers, the frequency of measurement is usually one
measurement per second. Therefore, for a default setting
of \code{warn.length = 60}, the chamber closure time
should be approximately one minute (60 seconds).}
}
\value{
Returns a data frame with 26 columns: a UniqueID per measurement,
         10 columns for the linear model results (linear flux estimate
         (\code{LM.flux}), initial gas concentration (\code{LM.C0}), final
         gas concentration (\code{LM.Ct}), slope of linear regression
         (\code{LM.slope}), mean absolute error ((\code{LM.MAE})), root mean
         square error ((\code{LM.RMSE})), standard error ((\code{LM.se})),
         relative se ((\code{LM.se.rel})),
         \ifelse{html}{\out{r<sup>2</sup>}}{\eqn{r^2}{ASCII}}, and p-value
         ((\code{LM.p.val}))), 10 columns for the non-linear model results
         (non-linear flux estimate (\code{HM.flux}), initial gas concentration
         (\code{HM.C0}), final gas concentration (\code{HM.Ci}), slope
         at \code{t=0} (\code{HM.slope}), mean absolute error (\code{HM.MAE}),
         root mean square error (\code{HM.RMSE}), standard error (\code{HM.se}),
         relative se (\code{HM.se.rel}),
         \ifelse{html}{\out{r<sup>2</sup>}}{\eqn{r^2}{ASCII}}, and curvature
         (kappa; \code{HM.k}), as well as the minimal detectable flux
         (\code{\link[GoFluxYourself]{MDF}}), the precision of the instrument
         (\code{prec}), the flux term (\code{\link[GoFluxYourself]{flux.term}}),
         kappa-max (\code{\link[GoFluxYourself]{k.max}}), the g factor
         (\code{\link[GoFluxYourself]{g.factor}}), the number of observations
         used (\code{nb.obs}) and the true initial gas concentration
         (\code{C0}) and final gas concentration (\code{Ct}).
}
\description{
A wrapper function to calculate GHG fluxes from static chamber measurements.
Calculates linear (\code{\link[GoFluxYourself]{LM.flux}}) and non-linear fluxes
(Hutchinson and Mosier model; \code{\link[GoFluxYourself]{HM.flux}}).
}
\details{
This function differs from the original version
\code{\link[GoFluxYourself]{goFlux}}. In this version, the non-linear model
is calculated without kappa-max (\code{\link[GoFluxYourself]{k.max}}). In the
original version, the kappa-max is estimated using the
\code{\link[GoFluxYourself]{MDF}}.

Flux estimate units are
\ifelse{html}{\out{µmol/m<sup>2</sup>s}}{\eqn{µmol/m^{2}s}{ASCII}}
(if initial concentration is ppm, e.g. CO2dry_ppm) and
\ifelse{html}{\out{nmol/m<sup>2</sup>s}}{\eqn{nmol/m^{2}s}{ASCII}}
(if initial concentration is ppb, e.g. CH4dry_ppb).

The function \code{\link[GoFluxYourself]{k.max}} calculates the maximal
curvature (kappa) of the non-linear model (Hutchinson and Mosier;
\code{\link[GoFluxYourself]{HM.flux}}) allowed for each flux measurements.
kappa-max is calculated based on the minimal detectable flux
(\code{\link[GoFluxYourself]{MDF}}), the linear flux estimate
\code{\link[GoFluxYourself]{LM.flux}}
and the measurement time. The unit of kappa-max is
\ifelse{html}{\out{s<sup>-1</sup>}}{\eqn{s^{-1}}{ASCII}}.

The function \code{\link[GoFluxYourself]{MDF}} calculates the minimal detectable
flux (MDF) based on instrument precision and measurement time.

The argument \code{Area} is in \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}},
but the output units from \code{goFlux()} are in
\ifelse{html}{\out{(m<sup>2</sup>)}}{\eqn{(m^2)}{ASCII}}. This means that there is a factor
of 10000 to convert from \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}}
to \ifelse{html}{\out{(m<sup>2</sup>)}}{\eqn{(m^2)}{ASCII}}. This is important
to take into account if one would provide something else than an Area in
\ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}} to the function.
For example, with incubated soil samples, one may provide an amount of soil
(kg) instead of an Area. To get the right units in that case, multiply the
kilograms of soil by 10000 to remove the conversion from
\ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}} to
\ifelse{html}{\out{(m<sup>2</sup>)}}{\eqn{(m^2)}{ASCII}}.
}
\examples{
data(LGR_manID)
CO2_flux <- goFlux_nokmax(LGR_manID, "CO2dry_ppm")
CH4_flux <- goFlux_nokmax(LGR_manID, "CH4dry_ppb")
H2O_flux <- goFlux_nokmax(LGR_manID, "H2O_ppm")

}
\seealso{
Look up the functions \code{\link[GoFluxYourself]{MDF}},
         \code{\link[GoFluxYourself]{flux.term}},
         \code{\link[GoFluxYourself]{g.factor}},
         \code{\link[GoFluxYourself]{k.max}},
         \code{\link[GoFluxYourself]{HM.flux}} and
         \code{\link[GoFluxYourself]{LM.flux}} of this package for more
         information about these parameters.

See also the import functions to find the
         default instrument precision for each instrument:
         \code{\link[GoFluxYourself]{DX4015_import}},
         \code{\link[GoFluxYourself]{G2508_import}},
         \code{\link[GoFluxYourself]{GAIA_import}},
         \code{\link[GoFluxYourself]{LGR_import}},
         \code{\link[GoFluxYourself]{LI6400_import}},
         \code{\link[GoFluxYourself]{LI7810_import}},
         \code{\link[GoFluxYourself]{LI7820_import}},
         \code{\link[GoFluxYourself]{LI8100_import}},
         \code{\link[GoFluxYourself]{LI8200_import}}
}
