% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.G2508-data.R
\docType{data}
\name{imp.G2508}
\alias{imp.G2508}
\title{Example gas measurements data from the Picarro G2508 gas analyzer}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(imp.G2508)
}
\description{
Raw data from the Picarro G2508 greenhouse gas analyzer that
has been processed by the function \code{\link[goFlux]{import.G2508}}
of this package.
}
\examples{
data(imp.G2508)
head(imp.G2508)

}
\keyword{datasets}
