% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.LI6400-data.R
\docType{data}
\name{imp.LI6400}
\alias{imp.LI6400}
\title{Example gas measurements data from the LI-COR GHG analyzer LI-6400}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(imp.LI6400)
}
\description{
Raw data from the LI-6400 greenhouse gas analyzer that has
been processed by the function \code{\link[goFlux]{import.LI6400}}
of this package.
}
\examples{
data(imp.LI6400)
head(imp.LI6400)

}
\keyword{datasets}
