% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.LI7820-data.R
\docType{data}
\name{imp.LI7820}
\alias{imp.LI7820}
\title{Example gas measurements data from the LI-COR GHG analyzer LI-7820}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(imp.LI7820)
}
\description{
Raw data from the LI-7820 greenhouse gas analyzer that has
been processed by the function \code{\link[goFlux]{import.LI7820}}
of this package.
}
\examples{
data(imp.LI7820)
head(imp.LI7820)

}
\keyword{datasets}
