% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.LI8200-data.R
\docType{data}
\name{imp.LI8200}
\alias{imp.LI8200}
\title{Example gas measurements data from the LI-COR Smart Chamber}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(imp.LI8200)
}
\description{
Raw data from the LI-8200 Smart Chamber that has
been processed by the function \code{\link[goFlux]{import.LI8200}}
of this package.
}
\examples{
data(imp.LI8200)
head(imp.LI8200)

}
\keyword{datasets}
