% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.UGGA-data.R
\docType{data}
\name{imp.UGGA}
\alias{imp.UGGA}
\title{Example gas measurements data from the LGR GHG analyzer UGGA}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(imp.UGGA)
}
\description{
Raw data from the micro-GGA Los Gatos Research greenhouse gas analyzer that
has been processed by the function \code{\link[goFlux]{import.UGGA}}
of this package.
}
\examples{
data(imp.UGGA)
head(imp.UGGA)

}
\keyword{datasets}
