% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.CR1000X.R
\name{import.CR1000X}
\alias{import.CR1000X}
\title{Import function for data from several gas analyzers (LI7810, LI7820, LI8100) saved on the Campbell Scientific CR1000X data logger}
\usage{
import.CR1000X(inputfile, timezone = "UTC", save = FALSE, keep_all = FALSE)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension
.data or .txt}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if \code{save = TRUE}, saves the file as an .RData file
in a RData folder in the current working directory. If
\code{save = FALSE}, returns the file in the Console, or load in
the Environment if assigned to an object.}

\item{keep_all}{logical; if \code{keep_all = TRUE}, keep all columns from the raw
file. The default is \code{keep_all = FALSE}, and columns that
are not necessary for gas flux calculation are removed.}
}
\value{
List of data frames containing raw data from each of the gas analyzers.
}
\description{
Imports single raw gas measurement files from the Campbell Scientific CR1000X data logger
(\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{CO[2]}{ASCII}} and
(\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{CH[4]}{ASCII}} and
(\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}})
}
\details{
Note that this function was designed for the following units in the raw file:
\itemize{
  \item ppm for \ifelse{html}{\out{CO<sub>2<}}{\eqn{CO[2]}{ASCII}}
  \item ppb for \ifelse{html}{\out{CH<sub>4<}}{\eqn{CH[4]}{ASCII}}
  \item ppb for \ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}}
  \item ppm for \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}
  \item kPa for pressure
  \item Celsius for temperature}
If your gas analyzers uses different units, either convert the units after
import, change the settings on your instrument, or contact the maintainer of
this package for support.

The precision of the instrument is needed to restrict kappa-max
(\code{\link[goFlux]{k.max}}) in the non-linear flux calculation
(\code{\link[goFlux]{HM.flux}}). Kappa-max is inversely proportional to
instrument precision. If the precision of your instrument is unknown, it is
better to use a low value (e.g. 1 ppm for
#' \ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}, or 1 ppb for
#' \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}) and
\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}}) to allow for more
curvature, especially for water vapor fluxes, or very long measurements, that
are normally curved. The default values given for instrument precision are
the ones found \href{https://www.licor.com/products/trace-gas/LI-7820}{online}
for the latest model of this instrument available at the time of the
creation of this function (11-2023).
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "CR1000X/CR1000X.data", package = "goFlux")

# Run function
imp.CR1000X <- import.CR1000X(inputfile = file.path)
}
\seealso{
Use the wrapper function \code{\link[goFlux]{import2RData}}
         to import multiple files from the same folder path using any instrument.

See also, import functions for other instruments:
         \code{\link[goFlux]{import.DX4015}},
         \code{\link[goFlux]{import.eosMX12}},
         \code{\link[goFlux]{import.EGM5}},
         \code{\link[goFlux]{import.G2201i}},
         \code{\link[goFlux]{import.G2508}},
         \code{\link[goFlux]{import.G4301}},
         \code{\link[goFlux]{import.GAIA}},
         \code{\link[goFlux]{import.GasmetPD}},
         \code{\link[goFlux]{import.GT5000}},
         \code{\link[goFlux]{import.HT8850}},
         \code{\link[goFlux]{import.LI6400}},
         \code{\link[goFlux]{import.LI7810}},
         \code{\link[goFlux]{import.LI8100}},
         \code{\link[goFlux]{import.LI8150}},
         \code{\link[goFlux]{import.LI8200}},
         \code{\link[goFlux]{import.LI8250}},
         \code{\link[goFlux]{import.N2OM1}},
         \code{\link[goFlux]{import.N2Oi2}},
         \code{\link[goFlux]{import.skyline}},
         \code{\link[goFlux]{import.uCH4}},
         \code{\link[goFlux]{import.uN2O}},
         \code{\link[goFlux]{import.UGGA}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
