% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.eosMX12.R
\name{import.eosMX12}
\alias{import.eosMX12}
\title{Import function for the eosMX 12-Channel Multiplexer}
\usage{
import.eosMX12(inputfile, timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .log}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if \code{save = TRUE}, saves the file as an .RData file
in a RData folder in the current working directory. If
\code{save = FALSE}, returns the file in the Console, or load in
the Environment if assigned to an object.}
}
\value{
A data frame containing raw data from the Eosense 12-Channel
Autochamber Multiplexer FRMonitor Files. Acts as auxiliary data sheet
recording chamber port, status, closures, openings, temperature, pressure,
and readings from 5 AUX sensors.
}
\description{
Imports single raw data files from the Eosense 12-Channel Autochamber Multiplexer
}
\details{
Note that this function was designed for the following units in the raw files:
\itemize{
  \item Epoch (unix) time for \code{date.format}
  \item kPa for atmospheric pressure
  \item Voltage for temperature
  \item Voltage for other AUX sensors}
If your eosMX instrument uses different units, either convert the units after
import, change the settings on your instrument, or contact the maintainer of
this package for support.
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "eosMX12/eosMX12.log", package = "goFlux")

# Import FRMonitor file as auxfile
imp.eosMX12 <- import.eosMX12(inputfile = file.path)

# Match the multiplexer file with gas data using the function autoID
data(imp.G2508)
autoID.eosMX12 <- autoID(imp.G2508, imp.eosMX12)
}
\seealso{
Use the wrapper function \code{\link[goFlux]{import2RData}}
         to import multiple files from the same folder path using any instrument.

See also, import functions for other instruments:
         \code{\link[goFlux]{import.DX4015}},
         \code{\link[goFlux]{import.EGM5}},
         \code{\link[goFlux]{import.G2201i}},
         \code{\link[goFlux]{import.G2508}},
         \code{\link[goFlux]{import.G4301}},
         \code{\link[goFlux]{import.GAIA}},
         \code{\link[goFlux]{import.GasmetPD}},
         \code{\link[goFlux]{import.GT5000}},
         \code{\link[goFlux]{import.HT8850}},
         \code{\link[goFlux]{import.LI6400}},
         \code{\link[goFlux]{import.LI7810}},
         \code{\link[goFlux]{import.LI7820}},
         \code{\link[goFlux]{import.LI8100}},
         \code{\link[goFlux]{import.LI8150}},
         \code{\link[goFlux]{import.LI8200}},
         \code{\link[goFlux]{import.LI8250}},
         \code{\link[goFlux]{import.N2OM1}},
         \code{\link[goFlux]{import.N2Oi2}},
         \code{\link[goFlux]{import.skyline}},
         \code{\link[goFlux]{import.uCH4}},
         \code{\link[goFlux]{import.uN2O}},
         \code{\link[goFlux]{import.UGGA}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
