% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.G2201i.R
\name{import.G2201i}
\alias{import.G2201i}
\title{Import function for the Picarro G2201-i Isotopic Analyzer}
\usage{
import.G2201i(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  save = FALSE,
  keep_all = FALSE,
  prec = c(0.2, 50, 100),
  CH4 = "HP_12CH4_dry",
  CO2 = "12CO2_dry",
  sum = TRUE,
  range = 10
)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .dat}

\item{date.format}{character string; specifies the date format found in the
raw data file. Choose one of the following: "dmy", "ymd",
or "mdy". Default is "ymd", as it is the date format from
the example data file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if \code{save = TRUE}, saves the file as an .RData file
in a RData folder in the current working directory. If
\code{save = FALSE}, returns the file in the Console, or load in
the Environment if assigned to an object.}

\item{keep_all}{logical; if \code{keep_all = TRUE}, keep all columns from the raw
file. The default is \code{keep_all = FALSE}, and columns that
are not necessary for gas flux calculation are removed.}

\item{prec}{numerical vector; the precision of the instrument for each gas,
in the following order: "CO2dry_ppm", "CH4dry_ppb" and "H2O_ppm".
The default is \code{prec = c(0.2, 50, 100)}.}

\item{CH4}{character string; which column should be used for
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} concentration.
By default, \code{CH4 = "HP_12CH4_dry"}.}

\item{CO2}{character string; which column should be used for
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} concentration.
By default, \code{CO2 = "12CO2_dry"}.}

\item{sum}{logical; if \code{sum = TRUE} (default), the arguments \code{CH4} and
\code{CO2} are ignored and, instead, both fractions (carbon isotope
12 and 13) of each gas, are summed after correction for water vapor,
to obtain the total gas concentration. See the argument \code{range}
for the selection between the HP or the HR measurement of
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} concentration.}

\item{range}{numerical value; at which threshold should the function select
between high precision (HP) and high resolution (HR) of
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}
concentration? The default is \code{range = 10}. If the methane
concentration recorded in the column \code{HP_12CH4_dry} is above
10 ppm, the concentration from the columns \code{HR_12CH4_dry}
and \code{HR_13CH4} are used for the sum instead of HP.}
}
\value{
A data frame containing raw data from Picarro G2201-i isotopic analyzer.
}
\description{
Imports single raw gas measurement files from the Picarro G2201-i with the
extension .dat (\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} and
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} isotopic analyzer)
}
\details{
In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For the instrument G2201-i, the date
is found in the column "DATE".

The arguments \code{CH4} and \code{CO2} require the user to specify which
columns contain the two gases of interest among the following for \code{CH4}:
\itemize{
  \item HP_12CH4
  \item HP_12CH4_dry
  \item HP_13CH4
  \item HR_12CH4
  \item HR_12CH4_dry
  \item HR_13CH4}
And among the following for \code{CO2}:
\itemize{
  \item 12CO2
  \item 12CO2_dry
  \item 13CO2
  \item 13CO2_dry}
Note that, HP stands for \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}
concentration measured with strong line for high precision but limited range
and HR stands for \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}
concentration measured with high range (between 10 and 1000 ppm). In addition,
_dry stands for dry mole fraction of gases, corrected for water vapor. Finally,
12 and 13 indicate each carbon isotope contained in each gas. If one selects
one of the non-dry mole fraction, it will be corrected for water vapor during
the import process.

Note that this function was designed for the following units in the raw file:
\itemize{
  \item ppm for \ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} and
        \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}
  \item percent (\%) for \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}
  \item Torr for pressure
  \item Celsius for temperature}
If your Picarro G2201-i uses different units, either convert the units after
import, change the settings on your instrument, or contact the maintainer of
this package for support.

The precision of the instrument is needed to restrict kappa-max
(\code{\link[goFlux]{k.max}}) in the non-linear flux calculation
(\code{\link[goFlux]{HM.flux}}). Kappa-max is inversely proportional to
instrument precision. If the precision of your instrument is unknown, it is
better to use a low value (e.g. 1 ppm) to allow for more curvature,
especially for water vapor fluxes, or very long measurements, that are
normally curved. The default values given for instrument precision are the
ones found \href{https://www.picarro.com/environmental/products/g2201i_isotopic_analyzer}{online}
for the latest model of this instrument, available at the time of the
creation of this function (05-2025).
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "G2201i/2025/05/13/G2201i.dat", package = "goFlux")

# Run function
imp.G2201i <- import.G2201i(inputfile = file.path)
}
\seealso{
Use the wrapper function \code{\link[goFlux]{import2RData}}
         to import multiple files from the same folder path using any instrument.

See also, import functions for other instruments:
         \code{\link[goFlux]{import.DX4015}},
         \code{\link[goFlux]{import.EGM5}},
         \code{\link[goFlux]{import.G2508}},
         \code{\link[goFlux]{import.G4301}},
         \code{\link[goFlux]{import.GAIA}},
         \code{\link[goFlux]{import.GasmetPD}},
         \code{\link[goFlux]{import.GT5000}},
         \code{\link[goFlux]{import.LI6400}},
         \code{\link[goFlux]{import.LI7810}},
         \code{\link[goFlux]{import.LI7820}},
         \code{\link[goFlux]{import.LI8100}},
         \code{\link[goFlux]{import.LI8200}},
         \code{\link[goFlux]{import.LI8250}},
         \code{\link[goFlux]{import.N2OM1}},
         \code{\link[goFlux]{import.uCH4}},
         \code{\link[goFlux]{import.uN2O}},
         \code{\link[goFlux]{import.UGGA}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
