% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.LI8200.R
\name{import.LI8200}
\alias{import.LI8200}
\title{Import function for the LI-COR Smart Chamber (LI-8200)}
\usage{
import.LI8200(inputfile, date.format = "ymd", timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .json}

\item{date.format}{character string; specifies the date format found in the
raw data file. Choose one of the following: "dmy", "ymd",
or "mdy". Default is "ymd", as it is the date format from
the example data file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if \code{save = TRUE}, saves the file as an .RData file
in a RData folder in the current working directory. If
\code{save = FALSE}, returns the file in the Console, or load in
the Environment if assigned to an object.}
}
\value{
A data frame containing raw data from the LI-COR Smart Chamber (LI-8200).
}
\description{
Imports single raw data files from the LI-COR Smart Chamber (LI-8200)
}
\details{
In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For the instrument LI-8200, the
date is found under one of the measurements, next to '"Date":'.

Note that this function was designed for the following units in the raw file:
\itemize{
  \item seconds for DeadBand
  \item cm for Offset (collar height)
  \item \ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}} for Area
  \item \ifelse{html}{\out{(cm<sup>3</sup>)}}{\eqn{(cm^3)}{ASCII}} for volumes
  \item mmol/mol for \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}
  \item kPa for pressure
  \item Celsius for temperature}
If your Smart Chamber uses different units, either convert the units after
import, change the settings on your instrument, or contact the maintainer of
this package for support.

Important note about the total volume: it is calculated from the addition
of the Area x Offset and the Chamber Volume. If a chamber volume is specified
("ChamVolume" in the raw file), that value is used, otherwise, the default
value is 4244.1\ifelse{html}{\out{(cm<sup>3</sup>)}}{\eqn{(cm^3)}{ASCII}}.
For more precise results, remember to add the volume inside your GHG analyzer
(IRGA volume; e.g., 6.4 \ifelse{html}{\out{(cm<sup>3</sup>)}}{\eqn{(cm^3)}{ASCII}}
for the LI-7810 and the LI-7820), plus the volume inside the tubing connecting
your instruments. For more details about volume estimation, download the
example auxiliary file from
\href{https://qepanna.github.io/goFlux/example.html#create-an-auxiliary-file}{the goFlux webpage}.

In addition, soil temperature and moisture are read with a HydraProbe
connected to the Smart Chamber. If you are using a different setup, please
contact the maintainer of this package for support.

As opposed to the other import functions, there is no option to "keep_all"
with the Smart Chamber. If you would like to import additional data using
this function, please contact the maintainer of this package for support.
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "LI8200/LI8200.json", package = "goFlux")

# Run function
imp.LI8200 <- import.LI8200(inputfile = file.path)
}
\seealso{
Use the wrapper function \code{\link[goFlux]{import2RData}}
         to import multiple files from the same folder path using any instrument.

See also, import functions for other instruments:
         \code{\link[goFlux]{import.DX4015}},
         \code{\link[goFlux]{import.eosMX12}},
         \code{\link[goFlux]{import.EGM5}},
         \code{\link[goFlux]{import.G2201i}},
         \code{\link[goFlux]{import.G2508}},
         \code{\link[goFlux]{import.G4301}},
         \code{\link[goFlux]{import.GAIA}},
         \code{\link[goFlux]{import.GasmetPD}},
         \code{\link[goFlux]{import.GT5000}},
         \code{\link[goFlux]{import.HT8850}},
         \code{\link[goFlux]{import.LI6400}},
         \code{\link[goFlux]{import.LI7810}},
         \code{\link[goFlux]{import.LI7820}},
         \code{\link[goFlux]{import.LI8100}},
         \code{\link[goFlux]{import.LI8150}},
         \code{\link[goFlux]{import.LI8250}},
         \code{\link[goFlux]{import.N2OM1}},
         \code{\link[goFlux]{import.N2Oi2}},
         \code{\link[goFlux]{import.skyline}},
         \code{\link[goFlux]{import.uCH4}},
         \code{\link[goFlux]{import.uN2O}},
         \code{\link[goFlux]{import.UGGA}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
