% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.LI8250.R
\name{import.LI8250}
\alias{import.LI8250}
\title{Import function for LI-COR Multiplexer (LI-8250)}
\usage{
import.LI8250(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  save = FALSE,
  keep_all = FALSE,
  inst1,
  inst2 = NULL,
  gas1,
  gas2 = NULL,
  prec1,
  prec2 = NULL,
  dry1 = T,
  dry2 = T,
  SWC.col = NULL,
  Tsoil.col = NULL
)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .82z}

\item{date.format}{character string; specifies the date format found in the
raw data file. Choose one of the following: "dmy", "ymd",
or "mdy". Default is "ymd", as it is the date format from
the example data file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if \code{save = TRUE}, saves the file as an .RData file
in a RData folder in the current working directory. If
\code{save = FALSE}, returns the file in the Console, or load in
the Environment if assigned to an object.}

\item{keep_all}{logical; if \code{keep_all = TRUE}, keep all columns from the
raw file. The default is \code{keep_all = FALSE}, and columns
that are not necessary for gas flux calculation are removed.}

\item{inst1, inst2}{character strings; a pattern to match the columns
containing the name of each instrument. For example, with a LI-COR
LI-7810, \code{inst1 = "LI-7810"} and \code{inst2 = NULL}, which
matches the example data file provided with the package.}

\item{gas1, gas2}{character vectors; a pattern to match the columns
containing each gas measurement. For example, with a LI-COR LI-7810
\code{gas1 = c("CO2_DRY", "CH4_DRY", "H2O")} and \code{gas2 = NULL},
which matches the example data file provided with the package.}

\item{prec1, prec2}{numerical vectors; the precision of the instrument
for each gas mentioned in \code{gas1} and \code{gas2}. For example,
with a LI-COR LI-7810, \code{prec1 = c(3.5, 0.6, 45)} and
\code{prec2 = NULL}, which matches the example data file provided with
the package. Note that the order in the arguments \code{precX} must
match the order of the arguments in \code{gasX}.}

\item{dry1, dry2}{logical; are the gas measurements compensated for water
vapor (dry fraction)? If \code{dryX = TRUE} (default), the gases are
compensated for water vapor, and will be named accordingly. For
example, the column "CO2_DRY" will become "CO2dry_ppm".}

\item{Tsoil.col, SWC.col}{character string; a pattern to match the columns
that fit the corresponding parameter.}
}
\value{
A data frame containing raw data from the LI-COR Multiplexer LI-8250.
}
\description{
Imports single raw data files from the LI-COR Multiplexer (LI-8250)
}
\details{
Files with the format .82z are zip archives that include all measurements
that occurred at a given port for one observation. Each .82z file includes
two files: a data.csv file and a metadata.json file. Raw files names for
.82z files are formatted with the instrument serial number and a time stamp
(serial-number-YYYYMMDDHHMMSS.82z). In \code{date.format}, the date format
refers to the date format in the file name. Alternatively, the date format
can be found in the data.csv file extracted from the .82z file, under the
column "Date".

Note that this function was designed for the following units in the raw file,
according to the data dictionary of the LI-8250 Multiplexer
\href{https://www.licor.com/env/support/LI-8250/topics/files-page.html#Datadictionary}{online}.
\itemize{
  \item \% v/v for soil water content (SWC)
  \item litters for volumes
  \item kPa for pressure
  \item Celsius for temperature}
For each gas, the units are taken from the third row in the data.csv file.
Currently, the function is adapted for use with LI-COR instruments (LI-870
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}}/
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}} Analyzer or LI-78xx
Trace Gas Analyzers). If you are using the function with a non-LI-COR
instrument, please contact the maintainer of this package for support.

Regarding the parameters \code{dryX}, in case of uncertainty, either contact
your technical support or assume that gases are compensated for water vapor,
which is normally the case.

The precision of the instrument is needed to restrict kappa-max
(\code{\link[goFlux]{k.max}}) in the non-linear flux calculation
(\code{\link[goFlux]{HM.flux}}). Kappa-max is inversely proportional to
instrument precision. If the precision of your instrument is unknown, it is
better to use a low value (e.g. 1 ppm for
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}}, or 1 ppb for
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}) to allow for more
curvature, especially for water vapor fluxes, or very long measurements, that
are normally curved.

According to the data dictionary of the LI-8250 Multiplexer
\href{https://www.licor.com/env/support/LI-8250/topics/files-page.html#Datadictionary}{online},
the column STATE indicates the chamber state, where, for example, 1 = closing
and 5 = closed.
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "LI8250/LI8250.82z", package = "goFlux")

# Run function
imp.LI8250 <- import.LI8250(inputfile = file.path, inst1 = "LI-7810",
                            gas1 = c("CO2_DRY", "CH4_DRY", "H2O"),
                            prec1 = c(3.5, 0.6, 45), SWC.col = "SWC_1",
                            Tsoil.col = "TS_1")
}
\seealso{
Use the wrapper function \code{\link[goFlux]{import2RData}}
         to import multiple files from the same folder path using any instrument.

See also, import functions for other instruments:
         \code{\link[goFlux]{import.DX4015}},
         \code{\link[goFlux]{import.EGM5}},
         \code{\link[goFlux]{import.G2201i}},
         \code{\link[goFlux]{import.G2508}},
         \code{\link[goFlux]{import.G4301}},
         \code{\link[goFlux]{import.GAIA}},
         \code{\link[goFlux]{import.GasmetPD}},
         \code{\link[goFlux]{import.GT5000}},
         \code{\link[goFlux]{import.LI6400}},
         \code{\link[goFlux]{import.LI7810}},
         \code{\link[goFlux]{import.LI7820}},
         \code{\link[goFlux]{import.LI8100}},
         \code{\link[goFlux]{import.LI8200}},
         \code{\link[goFlux]{import.N2OM1}},
         \code{\link[goFlux]{import.uCH4}},
         \code{\link[goFlux]{import.uN2O}},
         \code{\link[goFlux]{import.UGGA}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
