% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import2RData.R
\name{import2RData}
\alias{import2RData}
\title{Looping function for quick import of multiple raw gas measurement files}
\usage{
import2RData(path, instrument, date.format, timezone = "UTC")
}
\arguments{
\item{path}{character string; a folder path containing all files to be imported.
Ideally, to avoid any errors, the folder should only contain the
files to be imported.}

\item{instrument}{character string; specifies which instrument was used to
generate the files contained in the folder path. Chose one
of the following: "DX4015", "LGR", "G2508", "GAIA", "LI-6400",
"LI-7810", "LI-7820", "LI-8100", or "LI-8200". For more
information about an instrument, see the section "See also"
below.}

\item{date.format}{date format; the date format used in the raw data file.
Chose one of the following: "dmy", "ymd", or "mdy".}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}
}
\value{
a data frame saved as RData in a newly created folder, RData, into
the working directory.
}
\description{
Imports all raw gas measurement files contained in a folder.
Adapted to multiple greenhouse gas analyzers and other instruments:
\itemize{
\item LI-COR: LI-6400, LI-7810, LI-7820, LI-8100, LI-8200 (smart chamber)
\item Los Gatos Research instruments: (e.g. UGGA and m-GGA)
\item GAIA2TECH (DMR) automated chamber ECOFlux
\item Picarro: G2508
\item Gasmet: DX4015
}
}
\details{
In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For most instruments, the date is
one of the columns in the raw data file:
\itemize{
\item DX4015: column Date
\item G2508: column DATE
\item GAIA: column Titles:
\item LGR: column Time
\item LI-7810: column DATE
\item LI-7820: column DATE
\item LI-8100: column Date
}
For the instrument LI-6400, the date is found in one of the first lines in
a format containing abbreviations, for example "Thr Aug 6 2020", which would
be the date format "mdy".
}
\examples{
# Examples on how to use it with all instruments.
# The default time zone "UTC" is used in all cases.

# with the Gasmet instrument DX4015
import2RData(path = "inst/extdata/DX4015", instrument = "DX4015", date.format = "ymd")

# with the Picarro instrument G2508
import2RData(path = "inst/extdata/G2508", instrument = "G2508", date.format = "ymd")

# with the automated chamber ECOFlux (GAIA2TECH)
import2RData(path = "inst/extdata/GAIA", instrument = "GAIA", date.format = "ymd")

# with Los Gatos Research instruments (e.g. UGGA or m-GGA)
import2RData(path = "inst/extdata/LGR", instrument = "LGR", date.format = "dmy")

# with LI-COR instruments
import2RData(path = "inst/extdata/LI6400", instrument = "LI-6400", date.format = "mdy")
import2RData(path = "inst/extdata/LI7810", instrument = "LI-7810", date.format = "ymd")
import2RData(path = "inst/extdata/LI7820", instrument = "LI-7820", date.format = "ymd")
import2RData(path = "inst/extdata/LI8100", instrument = "LI-8100", date.format = "ymd")

# with the LI-COR smart chamber (LI-8200)
# with this instrument, date.format is not a useful parameter.
import2RData(path = "inst/extdata/LI8200", instrument = "LI-8200")

}
\seealso{
Use the wraper function \code{\link[GoFluxYourself]{import2RData}}
to import multiple files from the same folder path using any instrument.

Import functions for individual instruments:
\code{\link[GoFluxYourself]{DX4015_import}},
\code{\link[GoFluxYourself]{G2508_import}},
\code{\link[GoFluxYourself]{GAIA_import}},
\code{\link[GoFluxYourself]{LGR_import}},
\code{\link[GoFluxYourself]{LI6400_import}},
\code{\link[GoFluxYourself]{LI7810_import}},
\code{\link[GoFluxYourself]{LI7820_import}},
\code{\link[GoFluxYourself]{LI8100_import}},
\code{\link[GoFluxYourself]{LI8200_import}}
}
