% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.max.R
\name{k.max}
\alias{k.max}
\title{KAPPA max}
\usage{
k.max(MDF, LM.flux, t)
}
\arguments{
\item{MDF}{numerical; minimal detectable flux, calculated with the function MDF}

\item{LM.flux}{numerical; flux estimate from a linear flux calculation model}

\item{t}{measurement time (enclosure time) (seconds)}
}
\value{
a numerical value
}
\description{
The parameter kappa determines the curvature of the non-linear flux in the
Hutchinson and Mosier model. A maximum threshold for this parameter, kappa-max
can be calculated based on the minimal detectable flux (MDF), the linear
flux estimate and the measurement time. The units of the kappa-max is s-1
}
\keyword{internal}
