% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI6400_imp-data.R
\docType{data}
\name{LI6400_imp}
\alias{LI6400_imp}
\title{Example gas measurements data from LI-COR instruments}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(LI6400_imp)
}
\description{
Raw data from the LI-6400 greenhouse gas analyzer that has
been processed by the function \code{\link[goFlux]{LI6400_import}}
of this package.
}
\examples{
data(LI6400_imp)
head(LI6400_imp)

}
\keyword{datasets}
