% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI6400_import.R
\name{LI6400_import}
\alias{LI6400_import}
\title{Import function for LI-COR GHG analyzer LI-6400}
\usage{
LI6400_import(inputfile, date.format = "mdy", timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .txt}

\item{date.format}{date format; the date format used in the raw data file.
Chose one of the following: "dmy", "ymd", or "mdy". Default
is "mdy", as it is the date format from the example data
file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if save = TRUE, saves the file as RData in a RData folder
in the current working directory. If save = FALSE, returns the file
in the Console, or load in the Environment if assigned to an object.}
}
\value{
a data frame containing raw data from LI-COR GHG analyzer LI-6400.
}
\description{
Imports single raw gas measurement files from the LI-COR 6400
(CO2 and H2O GHG analyzer)
}
\details{
In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For the instrument LI-6400, the
date is found in one of the first lines in a format containing abbreviations,
for example "Thr Aug 6 2020", which would be the date format "mdy".
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "LI6400/example_LI6400.txt", package = "GoFluxYourself")

# Run function
LI6400.data <- LI6400_import(inputfile = file.path)

}
\seealso{
Use the wrapper function \code{\link[GoFluxYourself]{import2RData}}
         to import multiple files from the same folder path using any instrument.

Import functions for individual instruments:
         \code{\link[GoFluxYourself]{G2508_import}},
         \code{\link[GoFluxYourself]{GAIA_import}},
         \code{\link[GoFluxYourself]{LGR_import}},
         \code{\link[GoFluxYourself]{LI7810_import}},
         \code{\link[GoFluxYourself]{LI7820_import}},
         \code{\link[GoFluxYourself]{LI8100_import}},
         \code{\link[GoFluxYourself]{LI8200_import}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
