% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI7810_imp-data.R
\docType{data}
\name{LI7810_imp}
\alias{LI7810_imp}
\title{Example gas measurements data from LI-COR instruments}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(LI7810_imp)
}
\description{
Raw data from the LI-7810 greenhouse gas analyzer that has
been processed by the function \code{\link[GoFluxYourself]{LI7810_import}}
of this package.
}
\examples{
data(LI7810_imp)
head(LI7810_imp)

}
\keyword{datasets}
