% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI7820_imp-data.R
\docType{data}
\name{LI7820_imp}
\alias{LI7820_imp}
\title{Example gas measurements data from LI-COR instruments}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(LI7820_imp)
}
\description{
Raw data from the LI-7820 greenhouse gas analyzer that has
been processed by the function \code{\link[goFlux]{LI7820_import}}
of this package.
}
\examples{
data(LI7820_imp)
head(LI7820_imp)

}
\keyword{datasets}
