% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI7820_import.R
\name{LI7820_import}
\alias{LI7820_import}
\title{Import function for LI-COR GHG analyzer LI-7820}
\usage{
LI7820_import(inputfile, date.format = "ymd", timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{the name of a file with the extension .data}

\item{date.format}{Date format. Chose one of the following: "dmy", "ymd", or "mdy".
Default is "ymd" as it is the date format from the example
data file provided.}

\item{timezone}{a time zone in which to import the data to POSIXct format.
Default is "UTC". Note about time zone: I recommend using
the time zone "UTC" to avoid any issue related to summer
time and winter time changes.}

\item{save}{logical. If save = TRUE, save the file as Rdata in the current
working directory. If save = FALSE, return the file in the Console,
or load in the Environment if assigned to an object.}
}
\value{
a data frame
}
\description{
Imports single raw gas measurement files from the LI-COR 7820
(N2O and H2O GHG analyzer)
}
\examples{
# Examples on how to use:
file.path <- system.file("extdata", "LI7820/example_LI7820.data", package = "GoFluxYourself")

# 1. If you want to import the data and work with it directly
LI7820.data <- LI7820_import(inputfile = file.path)

# 2. Or if you want to import the data and save it as Rdata
LI7820_import(inputfile = file.path, save = TRUE)
}
