% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI8200_import.R
\name{LI8200_import}
\alias{LI8200_import}
\title{Import function for LI-COR smart chamber (LI-8200)}
\usage{
LI8200_import(inputfile, save = FALSE)
}
\arguments{
\item{inputfile}{the name of a file with the extension .json}

\item{save}{logical. If save = TRUE, save the file as Rdata in the current
working directory. If save = FALSE, return the file in the Console,
or load in the Environment if assigned to an object.}
}
\value{
a data frame
}
\description{
Imports single raw data files from the LI-COR smart chamber (LI-8200)
}
\examples{
# Examples on how to use:
file.path <- system.file("extdata", "LI8200/example_LI8200.json", package = "GoFluxYourself")

# 1. If you want to import the data and work with it directly
LI8200.data <- LI8200_import(inputfile = file.path)

# 2. Or if you want to import the data and save it as Rdata
LI8200_import(inputfile = file.path, save = TRUE)
}
