% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manID.G2201i-data.R
\docType{data}
\name{manID.G2201i}
\alias{manID.G2201i}
\title{Example gas measurements data from the Picarro G2201-i Isotopic Analyzer}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(manID.G2201i)
}
\description{
Processed data from the Picarro G2201-i Isotopic Analyzer after manual
identification of measurements with the functions \code{\link[goFlux]{obs.win}}
and \code{\link[goFlux]{click.peak2}} of this package.
}
\examples{
data(manID.G2201i)
head(manID.G2201i)

}
\keyword{datasets}
