% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manID.UGGA-data.R
\docType{data}
\name{manID.UGGA}
\alias{manID.UGGA}
\title{Example gas measurements data from the LGR analyzer UGGA}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(manID.UGGA)
}
\description{
Processed data from the micro-GGA Los Gatos Research greenhouse gas
analyzer after manual identification of measurements with the functions
\code{\link[goFlux]{obs.win}} and
\code{\link[goFlux]{click.peak2}} of this package.
}
\examples{
data(manID.UGGA)
head(manID.UGGA)

}
\keyword{datasets}
