% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv.dir.R
\name{mv.dir}
\alias{mv.dir}
\title{Move files in folder to another folder}
\usage{
mv.dir(in_path, out_path, pattern = NULL, recursive = FALSE)
}
\arguments{
\item{in_path}{a character vector of full path name. Input folder containing files to move}

\item{out_path}{a character vector of full path name. Output folder to move files into}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will be returned.}

\item{recursive}{logical. Should directories be deleted recursively?}
}
\value{
None
}
\description{
Wrapper function for moving all files contained in a folder to another folder.
}
\examples{
# create a dummy in_path folder
if(dir.exists("in_path") == FALSE){dir.create("in_path")}

# write a dummy file
write("dummy", file = "in_path/dummy.txt")

# move dummy file to out_path
# when recursive = TRUE, folder "in_path" is erased
mv.dir("in_path", "out_path", pattern = ".txt", recursive = TRUE)

# erase out_path and dummy file
unlink("out_path", recursive = TRUE)

}
