% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.win.R
\name{obs.win}
\alias{obs.win}
\title{Define measurements' window for manual identification}
\usage{
obs.win(
  inputfile,
  auxfile = NULL,
  gastype = deprecated(),
  obs.length = NULL,
  shoulder = 120
)
}
\arguments{
\item{inputfile}{data.frame; output from import or align functions.}

\item{auxfile}{data.frame; auxiliary data frame containing the columns
\code{start.time} and \code{UniqueID}. \code{start.time} must
contain a date and be in POSIXct format. The time zone must be
the same as the POSIX.time in \code{inputfile}. The default
time zone for the import functions is "UTC". A data frame from
the Smart Chamber (LI-8200) can be used as an auxiliary file.
In that case, \code{chamID} will be used instead of \code{UniqueID},
if \code{UniqueID} cannot be found.}

\item{gastype}{deprecated}

\item{obs.length}{numerical; chamber closure time (seconds). Default is NULL.
If \code{obs.length} is not provided, a column \code{obs.length}
should be contained in \code{auxfile} or \code{inputfile}.
Alternatively, \code{obs.length} will be calculated from
\code{start.time} and \code{cham.open} or \code{end.time}
if found in \code{auxfile} or \code{inputfile}.}

\item{shoulder}{numerical; time before and after measurement in observation
window (seconds). Default is 120 seconds.}
}
\value{
a list of data frames, split by UniqueID, merging \code{inputfile},
         and \code{auxfile}. Additionally, adds some time (shoulder) before
         and after the chamber closure time to help identify the best window
         of measurement with the function \code{\link[goFlux]{click.peak2}}.
}
\description{
Split gas data into one observation window per gas measurement
and create a list of data frames (one data frame per UniqueID).
}
\details{
In \code{gastype}, the gas species listed are the ones for which this 
         package has been adapted. Please write to the maintainer of this package 
         for adaptation of additional gases.
}
\examples{
# How to use in multiple situations:
library(dplyr)

## with a LGR instrument and an auxiliary file (.txt)
aux.path <- system.file("extdata", "aux_UGGA/aux_UGGA.txt", package = "goFlux")
auxfile <- read.delim(aux.path) \%>\%
  mutate(start.time = as.POSIXct(start.time, tz = "UTC"))
data(imp.UGGA)
ow.UGGA <- obs.win(inputfile = imp.UGGA, auxfile = auxfile,
                   obs.length = 180, shoulder = 60)

## with a LI-COR instrument and the Smart Chamber as auxiliary file
data(imp.LI8200)
data(imp.LI7820)
ow.LI7820 <- obs.win(inputfile = imp.LI7820, auxfile = imp.LI8200, shoulder = 60)

## with the LI-6400 and no auxiliary file
data(imp.LI6400)
ow.LI6400 <- obs.win(inputfile = imp.LI6400, shoulder = 0)

}
\seealso{
After defining the observation window with the function \code{obs.win},
         Use the function \code{\link[goFlux]{click.peak2}}.

For an automatic identification of gas measurements, see the function
\code{\link[goFlux]{autoID}}.
}
