% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.rep.R
\name{seq.rep}
\alias{seq.rep}
\title{Sequence repeat}
\usage{
\method{seq}{rep}(from, by, n.rep, length.seq, rep.seq = F)
}
\arguments{
\item{from}{numerical; inherited from seq().}

\item{by}{numerical; inherited from seq().}

\item{n.rep}{numerical; number of time the sequence should be repeated.}

\item{length.seq}{numerical; inherited from seq(..., length.out = length.seq).}

\item{rep.seq}{logical; instead of repeating a sequence, (1,2,3,1,2,3),
sequence a repetition (1,1,2,2,3,3). Default is rep.seq = F.}
}
\value{
a numerical sequence
}
\description{
A wrapper function that merges seq() and rep().
}
\keyword{internal}
