% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supress_message.R
\name{supress_message}
\alias{supress_message}
\title{Suppress messages}
\usage{
supress_message(..., messages = FALSE)
}
\arguments{
\item{...}{objects evaluated by eval()}

\item{messages}{logical.}
}
\value{
removes messages from evaluated statements
}
\description{
Function for suppressing messages from other functions.
}
\keyword{internal}
