% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uCH4_import.R
\name{uCH4_import}
\alias{uCH4_import}
\title{Import function for the MIRA Ultra
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}/
\ifelse{html}{\out{C<sub>2</sub>H<sub>6</sub>}}{\eqn{C[2]H[6]}{ASCII}} analyzer
from Aeris Technologies}
\usage{
uCH4_import(
  inputfile,
  date.format = "mdy",
  timezone = "UTC",
  save = FALSE,
  keep_all = FALSE,
  prec = c(1, 0.5, 1)
)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .txt}

\item{date.format}{character string; specifies the date format found in the
raw data file. Choose one of the following: "dmy", "ymd",
or "mdy". Default is "mdy", as it is the date format from
the example data file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if \code{save = TRUE}, saves the file as an .RData file
in a RData folder in the current working directory. If
\code{save = FALSE}, returns the file in the Console, or load in
the Environment if assigned to an object.}

\item{keep_all}{logical; if \code{keep_all = TRUE}, keep all columns from the raw
file. The default is \code{keep_all = FALSE}, and columns that
are not necessary for gas flux calculation are removed.}

\item{prec}{numerical vector; the precision of the instrument for each gas,
in the following order: "CH4_ppb", "C2H6dry_ppb" and "H2O_ppm".
The default is \code{prec = c(1, 0.5, 1)}. The precision for
water vapor is currently unknown, and a default of 1 ppm is used.}
}
\value{
A data frame containing raw data from the MIRA Ultra
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}/
\ifelse{html}{\out{C<sub>2</sub>H<sub>6</sub>}}{\eqn{C[2]H[6]}{ASCII}} analyzer
from Aeris Technologies
}
\description{
Imports single raw gas measurement files from the MIRA Ultra
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}/
\ifelse{html}{\out{C<sub>2</sub>H<sub>6</sub>}}{\eqn{C[2]H[6]}{ASCII}}
Natural Gas Leak Detection System (with GPS) from Aeris Technologies
(\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}},
\ifelse{html}{\out{C<sub>2</sub>H<sub>6</sub>}}{\eqn{C[2]H[6]}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}) with the extension .txt
}
\details{
In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For these instruments, the date
is found in the column "Time Stamp".

Note that this function was designed for the following units in the raw file:
\itemize{
  \item ppm for \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} and
  \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}
  \item ppb for \ifelse{html}{\out{C<sub>2</sub>H<sub>6</sub>}}{\eqn{C[2]H[6]}{ASCII}}
  \item mbars for pressure
  \item Celsius for temperature}
If your Aeris instrument (MIRA Ultra
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}/
\ifelse{html}{\out{C<sub>2</sub>H<sub>6</sub>}}{\eqn{C[2]H[6]}{ASCII}} analyzer) uses
different units, either convert the units after import, change the settings
on your instrument, or contact the maintainer of this package for support.

The precision of the instrument is needed to restrict kappa-max
(\code{\link[goFlux]{k.max}}) in the non-linear flux calculation
(\code{\link[goFlux]{HM.flux}}). Kappa-max is inversely proportional to
instrument precision. If the precision of your instrument is unknown, it is
better to use a low value (e.g. 1 ppm for
\ifelse{html}{\out{C<sub>2</sub>H<sub>6</sub>}}{\eqn{C[2]H[6]}{ASCII}} and
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}, or 1 ppb for
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}) to allow for more
curvature, especially for water vapor fluxes, or very long measurements, that
are normally curved. The default values given for instrument precision are
the ones found \href{https://aerissensors.com/mira-pico-mobile-lds/s}{online}
for the latest model of this instrument available at the
time of the creation of this function (01-2024). The precision for water
vapor is currently unknown, and a default of 1 ppm is used.
}
\examples{
# Examples on how to use:
file.path <- system.file("extdata", "uCH4/uCH4.txt", package = "goFlux")

uCH4_imp <- uCH4_import(inputfile = file.path)

}
\seealso{
Use the wrapper function \code{\link[goFlux]{import2RData}}
         to import multiple files from the same folder path using any instrument.

See also, import functions for other instruments:
         \code{\link[goFlux]{DX4015_import}},
         \code{\link[goFlux]{EGM5_import}},
         \code{\link[goFlux]{G2508_import}},
         \code{\link[goFlux]{G4301_import}},
         \code{\link[goFlux]{GAIA_import}},
         \code{\link[goFlux]{LGR_import}},
         \code{\link[goFlux]{LI6400_import}},
         \code{\link[goFlux]{LI7810_import}},
         \code{\link[goFlux]{LI7820_import}},
         \code{\link[goFlux]{LI8100_import}},
         \code{\link[goFlux]{LI8200_import}},
         \code{\link[goFlux]{N2OM1_import}},
         \code{\link[goFlux]{uN2O_import}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
