% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.flux.R
\name{best.flux}
\alias{best.flux}
\title{Automatic selection of best flux estimate}
\usage{
best.flux(
  flux.result,
  criteria = c("MAE", "RMSE", "AICc", "SErel", "g.factor", "kappa", "MDF", "nb.obs",
    "intercept", "p-value"),
  intercept.lim = NULL,
  SErel = 5,
  g.limit = 2,
  p.val = 0.05,
  k.ratio = 1,
  warn.length = 60
)
}
\arguments{
\item{flux.result}{data.frame; output from the function
\code{\link[GoFluxYourself]{goFlux}}.}

\item{criteria}{character vector; criteria used to asses the goodness of fit
of the linear and non-linear flux estimates. Must be at least
one the following: "MAE", "RMSE", "AICc", "SErel", "g.factor",
"kappa", "MDF", "nb.obs", "p-value", or "intercept". The
default is all of them.}

\item{intercept.lim}{numerical vector of length 2; inferior and superior
limits of the intercept (initial concentration;
\ifelse{html}{\out{C<sub>0</sub>}}{\eqn{C[0]}{ASCII}}).
Must be the same units as \code{gastype} in the
\code{\link[GoFluxYourself]{goFlux}} function. If
\code{intercept.lim = NULL}, the limits are calculated
from the true values of
\ifelse{html}{\out{C<sub>0</sub>}}{\eqn{C[0]}{ASCII}} and
\ifelse{html}{\out{C<sub>i</sub>}}{\eqn{C[i]}{ASCII}}
for each measurement.}

\item{SErel}{numerical value; maximal relative standard error accepted (\%)
on flux estimate. The default is 5\%.}

\item{g.limit}{numerical value; maximal limit of the the g-factor, which is the
ratio between the non-linear flux estimate
(\code{\link[GoFluxYourself]{HM.flux}}) and the linear flux
estimate (\code{\link[GoFluxYourself]{LM.flux}}). Recommended
thresholds for the g-factor are 4 (flexible), 2 (medium), or
1.25 (conservative). The default limit is \code{g.limit = 2}.}

\item{p.val}{numerical value; a limit for a statistically non-zero flux.
The default threshold is p-value < 0.05.}

\item{k.ratio}{numerical value; maximal limit of the ratio between kappa and
the kappa-max. kappa-max is the maximal curvature (kappa)
of the non-linear regression (Hutchinson and Mosier model)
allowed for a each flux measurements. With \code{k.ratio}, one
can chose to select the LM flux estimate instead of the
non-linear flux estimate by giving a percentage of kappa-max.
Default is \code{k.ratio = 1}.}

\item{warn.length}{numerical; minimum amount of observations accepted (nb.obs).
With nowadays portable greenhouse gas analyzers, the
frequency of measurement is usually one measurement per
second. Therefore, for a default setting of
\code{warn.length = 60}, the chamber closure time
should be approximately one minute (60 seconds).}
}
\value{
a data.frame identical to the input \code{flux.result} with the
         additional columns \code{HM.diagnose}, \code{LM.diagnose},
         \code{best.flux}, \code{model} and \code{quality.check}. For each
         criteria selected, an additional column is also added to specify
         the limits used for those criteria (e.g. \code{RMSE.lim},
         \code{p.val.lim}, etc.)
}
\description{
This automatic selection of the best flux estimate (linear or non-linear) is
based on objective criteria and non-arbitrary thresholds.
}
\details{
In \code{criteria}, the indices of model fit "MAE", "RMSE" and "SErel" all
have a threshold. For MAE and RMSE, the threshold is instrument precision.
For SErel, the threshold is given as an argument, where the default limit is
5\%. These indices also compare the two models (linear, LM, and non-linear,
HM). The selection of the best model regarding indices of model fit ("MAE",
"RMSE", "AICc" and "SErel") is based on a scoring system. Each model starts
with a score of 0. Each time one of the model has a worst fit than the other
model, based on "MAE", "RMSE", "AICc" or "SErel", that model's score is
increased by 1. Whichever model has the lowest score wins. In case of a tie,
the non-linear model wins.

Another criteria, which is calculated automatically with the function
\code{\link[GoFluxYourself]{best.flux}}, is the g-factor
(\code{\link[GoFluxYourself]{g.factor}}), which is the ratio between the HM
flux estimate and the LM flux estimate. Recommended thresholds for the g-factor
are 4 (flexible), 2 (medium), or 1.25 (conservative).

The function \code{\link[GoFluxYourself]{k.max}} calculates the maximal
curvature (kappa) of the non-linear model (Hutchinson and Mosier) allowed for
each flux measurement. k.max is calculated based on the minimal detectable
flux (\code{\link[GoFluxYourself]{MDF}}), the linear flux estimate
(\code{LM.flux}) and the measurement time. The unit of kappa-max is
\ifelse{html}{\out{s<sup>-1</sup>}}{\eqn{s^{-1}}{ASCII}}.

The function \code{\link[GoFluxYourself]{MDF}} calculates the minimal
detectable flux (MDF) based on instrument precision and measurement time.
}
\examples{
data(LGR_manID)
LGR_flux <- goFlux(LGR_manID, "CO2dry_ppm")
LGR_res <- best.flux(LGR_flux)

}
\seealso{
Look up the functions \code{\link[GoFluxYourself]{MDF}},
         \code{\link[GoFluxYourself]{flux.term}},
         \code{\link[GoFluxYourself]{g.factor}},
         \code{\link[GoFluxYourself]{k.max}},
         \code{\link[GoFluxYourself]{HM.flux}} and
         \code{\link[GoFluxYourself]{LM.flux}}
         of this package for more information about these parameters.

See also the function \code{\link[GoFluxYourself]{goFlux}}
         for more information about usage.
}
