% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.flux.R
\name{best.flux}
\alias{best.flux}
\title{Automatic selection of best flux estimate}
\usage{
best.flux(
  flux.result,
  gastype,
  criteria = c("g.factor", "kappa", "MDF", "p-value", "R2", "intercept", "SE.rel"),
  intercept.lim = NULL,
  SE.rel = 5,
  g.limit = 2,
  p.val = 0.05,
  r2 = 0.6
)
}
\arguments{
\item{flux.result}{data.frame; output from goFlux()}

\item{gastype}{character string; same as used in goFlux. Must be one of the
following: "CO2dry_ppm", "CH4dry_ppb", "N2Odry_ppb" or "H2O_ppm".}

\item{criteria}{character vector; criteria used to asses the goodness of fit
of the linear and non-linear flux estimates. Must be at least
one the following: criteria = c("g.factor", "kappa", "MDF",
"p-value", "R2", "intercept", "SE.rel"). Default is all of them.}

\item{intercept.lim}{numerical vector of length 2; inferior and superior
limits of the intercept (initial concentration). Same
units as gastype.}

\item{SE.rel}{numerical value; maximal standard error accepted (\%) on flux
estimate. The default setting is 5\%.}

\item{g.limit}{numerical value; maximal limit of the the G-factor: the ratio
between the result of a non-linear flux calculation model
(e.g. Hutchinson and Mosier; HM) and the result of a linear
flux calculation model. The threshold for the G-factor should
be 4 (flexible), 2 (medium), or 1.25 (conservative). Default
limit is set to 2.}

\item{p.val}{numerical value; a limit for a statistically non-zero flux.
The default threshold is p-value < 0.05.}

\item{r2}{numerical value; a maximal limit under which a warning is issued
for quality checking. The default value is r2 < 0.6.}
}
\value{
a data.frame
}
\description{
This automatic selection of the best flux estimate (linear or non-linear) is
based on objective criteria and non-arbitrary thresholds.
}
\examples{
data(example_LGR_manID)
example_LGR_flux <- goFlux(example_LGR_manID, "CO2dry_ppm")
criteria <- c("g.factor", "kappa", "MDF", "R2", "SE.rel")
example_LGR_res <- best.flux(example_LGR_flux, "CO2dry_ppm", criteria)

}
\seealso{
[goFlux()]

[flux.plot()]

[flux2pdf()]
}
