% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.flux.R
\name{best.flux}
\alias{best.flux}
\title{Automatic selection of best flux estimate}
\usage{
best.flux(
  flux.result,
  criteria = c("MAE", "g.factor", "kappa", "MDF", "nb.obs", "p-value", "intercept",
    "SE.rel"),
  intercept.lim = NULL,
  SE.rel = 5,
  g.limit = 2,
  p.val = 0.05,
  k.ratio = 1,
  RMSE = FALSE,
  warn.length = 60
)
}
\arguments{
\item{flux.result}{data.frame; output from the function \code{\link[GoFluxYourself]{goFlux}}.}

\item{criteria}{character vector; criteria used to asses the goodness of fit of the
linear and non-linear flux estimates. Must be at least one the following:
\code{criteria = c("MAE", "g.factor", "kappa", "MDF", "nb.obs", "p-value", "intercept", "SE.rel")}.
Default is all of them.}

\item{intercept.lim}{numerical vector of length 2; inferior and superior
limits of the intercept (initial concentration; C0).
Must be the same units as \code{gastype}.}

\item{SE.rel}{numerical value; maximal relative standard error accepted (\%)
on flux estimate. The default is 5\%.}

\item{g.limit}{numerical value; maximal limit of the the g-factor, which is the
ratio between the non-linear flux estimate
(\code{\link[GoFluxYourself]{HM.flux}}) and the linear flux
estimate (\code{\link[GoFluxYourself]{LM.flux}}). Recommended
thresholds for the g-factor are 4 (flexible), 2 (medium), or
1.25 (conservative). The default limit is \code{g.limit = 2}.}

\item{p.val}{numerical value; a limit for a statistically non-zero flux.
The default threshold is p-value < 0.05.}

\item{k.ratio}{numerical value; maximal limit of the ratio between kappa and
the kappa-max. kappa-max is the maximal curvature (kappa)
of the non-linear regression (Hutchinson and Mosier model)
allowed for a each flux measurements. With \code{k.ratio}, one
can chose to select the LM flux estimate instead of the
non-linear flux estimate by giving a percentage of kappa-max.
Default is \code{k.ratio = 1}.}

\item{RMSE}{logical; If \code{RMSE = TRUE}, use Root Mean Square Error (RMSE)
instead of Mean Absolute Error (MAE).}

\item{warn.length}{numerical; minimum amount of observations accepted (nb.obs).
With nowadays portable greenhouse gas analyzers, the
frequency of measurement is usually one measurement per
second. Therefore, for a default setting of
\code{warn.length = 60}, the chamber closure time
should be approximately one minute (60 seconds).}
}
\value{
a data.frame identical to the input \code{flux.result} with the
         additional columns \code{HM.diagnose}, \code{LM.diagnose},
         \code{best.flux}, \code{model} and \code{quality.check}.
}
\description{
This automatic selection of the best flux estimate (linear or non-linear) is
based on objective criteria and non-arbitrary thresholds.
}
\details{
The function \code{\link[GoFluxYourself]{k.max}} calculates the maximal
curvature (kappa) of the non-linear model (Hutchinson and Mosier) allowed for
each flux measurements. k.max is calculated based on the minimal detectable
flux (MDF), the linear flux estimate (LM.flux) and the measurement time. The
unit of kappa-max is \ifelse{html}{\out{s<sup>-1</sup>}}{\eqn{s^{-1}}{ASCII}}.

The function \code{\link[GoFluxYourself]{MDF}} calculates the minimal
detectable flux (MDF) based on instrument precision and measurements time.

Another criteria, which is calculated automatically with the function
\code{\link[GoFluxYourself]{best.flux}}, is the g-factor
(\code{\link[GoFluxYourself]{g.factor}}), which is the ratio between the HM
flux estimate and the LM flux estimate. The threshold for the g-factor
should be 4 (flexible), 2 (medium), or 1.25 (conservative).
}
\examples{
data(example_LGR_manID)
example_LGR_flux <- goFlux(example_LGR_manID, "CO2dry_ppm")
criteria <- c("g.factor", "kappa", "MDF", "SE.rel")
example_LGR_res <- best.flux(example_LGR_flux, "CO2dry_ppm", criteria)

}
\seealso{
Look up the functions \code{\link[GoFluxYourself]{MDF}},
         \code{\link[GoFluxYourself]{flux.term}},
         \code{\link[GoFluxYourself]{g.factor}},
         \code{\link[GoFluxYourself]{k.max}},
         \code{\link[GoFluxYourself]{HM.flux}} and
         \code{\link[GoFluxYourself]{LM.flux}}
         of this package for more information about these parameters.

See also the function \code{\link[GoFluxYourself]{goFlux}}
         for more information about usage.
}
