% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click.peak.loop.R
\name{click.peak.loop}
\alias{click.peak.loop}
\title{Looped manual identification of start and end of gas measurements}
\usage{
click.peak.loop(
  x,
  flux.unique,
  gastype = "CO2dry_ppm",
  sleep = 3,
  plot.lim = c(380, 1000),
  warn.length = 60
)
}
\arguments{
\item{x}{a numerical sequence that indicates objects in a list. Must be used
with \code{\link[base]{lapply}} after the function
\code{\link[GoFluxYourself]{obs.win}}. See examples below to see
how to use with \code{\link[base]{lapply}}.}

\item{flux.unique}{data.frame; output from the function
\code{\link[GoFluxYourself]{obs.win}}.
Must contain the columns \code{gastype} (see below),
\code{POSIX.time} and \code{UniqueID}.}

\item{gastype}{character string; specifies which gas should be displayed on
the plot to manually select start time and end time of
measurements. Must be one of the following: "CO2dry_ppm",
"COdry_ppb", "CH4dry_ppb", "N2Odry_ppb", "NH3dry_ppb" or
"H2O_ppm". Default is "CO2dry_ppm".}

\item{sleep}{numerical value; delay before closing the resulting plot. When
used with the function \code{\link[GoFluxYourself]{click.peak.loop}},
grants a delay between measurements to visually inspect the
output before processing the next measurement. Sleep must be
shorter than 10 seconds.}

\item{plot.lim}{numerical vector of length 2; sets the Y axis limits in the
plots. Default values are set for a typical gas measurement
of "CO2dry_ppm" from soils: \code{plot.lim = c(380,1000)}.}

\item{warn.length}{numerical value; limit under which a measurement is flagged
for being too short (\code{nb.obs < warn.length}). Default
value is \code{warn.length = 60}.}
}
\value{
A list of data.frame, split by \code{UniqueID}, identical to the
         input \code{flux.unique}, with the additional columns \code{flag},
         \code{Etime}, \code{start.time_corr}, \code{end.time_corr} and
         \code{obs.length_corr}.
}
\description{
Identify the start and the end of a measurement by clicking on them in a
scatter plot. To use in a loop with multiple measurements, first use the
function \code{\link[GoFluxYourself]{obs.win}} to identify the observation
window of each measurement and then use the wrapper function
\code{\link[GoFluxYourself]{click.peak.loop}} with
\code{\link[base]{lapply}} (see example below).
}
\details{
The argument \code{plot.lim} is used to remove any data points below and
above the plot limits for a better view of the scatter plot. If the gas
measurements are larger than the minimum or smaller than the maximum plot
limit values, then the plot will automatically zoom in and adjust to those
values. The default plot limits are set for a typical gas measurement of
"CO2dry_ppm" from a soil respiration measurement: \code{plot.lim = c(380,1000)},
where 380 ppm is the minimum plotted concentration, which should be close to
atmospheric concentration, and 1000 ppm is the maximum plotted concentration,
which correspond to a maximal accumulated concentration in the chamber before
considering it an outlier (e.g. caused by breath or gas bubble). For other
gasses, the user must specify the plot limits themselves. Here are some
suggestions of plot limits for the other gases:
\itemize{
  \item "CH4dry_ppb": \code{plot.lim = c(2200, 1800)}
  \item "N2Odry_ppb": \code{plot.lim = c(250, 500)}
  \item "NH3dry_ppb": \code{plot.lim = c(0, 200)}
  \item "COdry_ppb": \code{plot.lim = c(0, 200)}
  \item "H2O_ppm": \code{plot.lim = c(10000, 20000)}
}
These values will vary depending on ecosystem type and chamber application scheme.

\code{warn.length} is the limit below which the chamber closure time is
flagged for being too short (\code{nb.obs < warn.length}). Portable
greenhouse gas analyzers typically measure at a frequency of 1 Hz. Therefore,
for the default setting of \code{warn.length = 60}, the chamber closure time
should be approximately one minute (60 seconds). If the number of
observations is smaller than the threshold, a warning is printed after the
loop: "Number of observations for UniqueID: 'UniqueID' is X observations".

In \code{gastype}, the gas species listed are the ones for which this package
has been adapted. Please write to the maintainer of this package for
adaptation of additional gases.
}
\examples{
\dontrun{
# How to use in multiple situations:
# Note that gastype = "CO2dry_ppm" is the default setting
library(dplyr)
library(purrr)

## with a LGR instrument and an auxiliary file (.txt)
aux.path <- system.file("extdata", "LGR_aux/LGR_aux.txt", package = "GoFluxYourself")
auxfile <- read.delim(aux.path) \%>\%
  mutate(start.time = as.POSIXct(start.time, tz = "UTC"))
data(LGR_imp)
LGR_ow <- obs.win(inputfile = LGR_imp, auxfile = auxfile,
                  obs.length = 180, shoulder = 60)
LGR_manID <- lapply(seq(1,3), click.peak.loop, flux.unique = LGR_ow) \%>\%
  map_df(., ~as.data.frame(.x))

## with a LI-COR instrument and the Smart Chamber as auxiliary file
data(LI8200_imp)
data(LI7820_imp)
LI7820_ow <- obs.win(inputfile = LI7820_imp, auxfile = LI8200_imp, shoulder = 60)
LI7810_manID <- lapply(seq(1,3), click.peak.loop, flux.unique = LI7820_ow,
                       gastype = "N2Odry_ppb", plot.lim = c(250,500)) \%>\%
  map_df(., ~as.data.frame(.x))

## with the LI-6400 and no auxiliary file
data(LI6400_imp)
LI6400_ow <- obs.win(inputfile = LI6400_imp, shoulder = 0)
LI6400_manID <- lapply(seq(1,3), click.peak.loop, flux.unique = LI6400_ow) \%>\%
  map_df(., ~as.data.frame(.x))
}
}
\seealso{
Use the function \code{\link[GoFluxYourself]{click.peak}} for
         individual measurements, instead of using in a loop with
         \code{\link[GoFluxYourself]{click.peak.loop}}. See also
         \code{\link[GoFluxYourself]{obs.win}} to prepare a list of data.frame.
}
