% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_LGR_manID-data.R
\docType{data}
\name{example_LGR_manID}
\alias{example_LGR_manID}
\title{Example gas measurements data from LGR instruments}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(example_LGR_manID)
}
\description{
Processed data from the micro-GGA Los Gastos Reasearch greenhouse gas
analyzer after manual indentification of measurements with the functions
obs.win() and click.peak.loop() of this package.
}
\examples{
data(example_LGR_manID)
head(example_LGR_manID)

}
\keyword{datasets}
