% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.plot.R
\name{flux.plot}
\alias{flux.plot}
\title{Plots for quality checking of GHG flux measurements}
\usage{
flux.plot(flux.results, dataframe, gastype, shoulder = 30)
}
\arguments{
\item{flux.results}{a data.frame; output from the function goFlux().}

\item{dataframe}{a data.frame containing gas measurements (see gastype below),
and the following: UniqueID, Etime and flag (same dataframe
as used with the function goFlux()).}

\item{gastype}{character string; specify which column was used for the
flux calculations. Must be one of the following: "CO2dry_ppm",
"CH4dry_ppb", "N2Odry_ppb" or "H2O_ppm".}

\item{shoulder}{numerical value; time before and after measurement in observation
window (seconds). Default is 30 seconds.}
}
\value{
a list of plots, one per UniqueID
}
\description{
Description
}
\examples{
data(example_LGR_manID)
example_LGR_flux <- goFlux(example_LGR_manID, "CO2dry_ppm")
criteria <- c("g.factor", "kappa", "MDF", "R2", "SE.rel")
example_LGR_res <- best.flux(example_LGR_flux, "CO2dry_ppm", criteria)
example_LGR_plots <- flux.plot(example_LGR_res, example_LGR_manID, "CO2dry_ppm")

}
\seealso{
[goFlux()]

[best.flux()]

[flux2pdf()]
}
