% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.plot.R
\name{flux.plot}
\alias{flux.plot}
\title{Plots for quality checking of GHG flux measurements}
\usage{
flux.plot(
  flux.results,
  dataframe,
  gastype,
  shoulder = 30,
  plot.legend = c("MAE", "RMSE"),
  plot.display = c("MDF", "prec"),
  quality.check = TRUE,
  flux.unit = NULL,
  flux.term.unit = NULL,
  best.model = TRUE,
  p.val.disp = "round",
  side = "left"
)
}
\arguments{
\item{flux.results}{a data.frame; output from the function
\code{\link[GoFluxYourself]{best.flux}}}

\item{dataframe}{a data.frame containing gas measurements (see \code{gastype}
below) and the following columns: \code{UniqueID}, \code{Etime}
and \code{flag} (same \code{dataframe} as used with the function
\code{\link[GoFluxYourself]{goFlux}}).}

\item{gastype}{character string; specifies which column was used for the
flux calculations. Must be one of the following: "CO2dry_ppm",
"COdry_ppb", "CH4dry_ppb", "N2Odry_ppb", "NH3dry_ppb" or "H2O_ppm".}

\item{shoulder}{numerical value; time before and after measurement in observation
window (seconds). Default is 30 seconds.}

\item{plot.legend}{character vector; specifies which parameters should be
displayed in a legend above each plot. "flux" is always
displayed. A maximum of 6 parameters can be displayed in
the legend (including "flux"). Chose up to five extra
parameters from the following: "MAE", "RMSE", "SErel",
"SE", "r2", "LM.p.val", "HM.k", "k.max", "k.ratio" and
"g.factor". Default is \code{plot.legend = c("MAE", "RMSE")}.}

\item{plot.display}{character vector; specifies which parameters should be
displayed on the plot. Chose from the following: "C0",
"Ci", "cham.close", "cham.open", "crop", "MDF", "nb.obs",
"flux.term" and "prec". Default is
\code{plot.display = c("MDF", "prec")}.}

\item{quality.check}{logical; if \code{quality.check = TRUE}, the column
\code{quality.check} (output from the function
\code{\link[GoFluxYourself]{best.flux}}) is displayed
below the plot.}

\item{flux.unit}{character string; flux units to be displayed on the plots.
By default, the units are
\ifelse{html}{\out{µmol m<sup>-2</sup>s<sup>-1</sup>}}{\eqn{µmol m^{-2}s^{-1}}{ASCII}}
(if initial concentration is ppm, e.g. CO2dry_ppm) and
\ifelse{html}{\out{nmol m<sup>-2</sup>s<sup>-1</sup>}}{\eqn{nmol m^{-2}s^{-1}}{ASCII}}
(if initial concentration is ppb, e.g. CH4dry_ppb). \cr
For example, one may want to use
\ifelse{html}{\out{nmol kg<sup>-1</sup>h<sup>-1</sup>}}{\eqn{µmol kg^{-1}h^{-1}}{ASCII}}
for incubated soil samples. In such a case, write
\code{flux.unit = "nmol~kg^-1*h^-1"}.}

\item{flux.term.unit}{character string; units for the flux term to be displayed
on the plots. By default, the units are
\ifelse{html}{\out{mol m<sup>-2</sup>}}{\eqn{mol m^{-2}}{ASCII}}:
\code{flux.term.unit = "mol~m^-2"}}

\item{best.model}{logical; if \code{best.model = TRUE}, display a star sign
next to the best model selected by the function
\code{\link[GoFluxYourself]{best.flux}}.}

\item{p.val.disp}{character string; how should the p-value be displayed in the
legend above the plot. Chose one of the following: "star",
"round", "value".}

\item{side}{character string; chose a side to display
\ifelse{html}{\out{C<sub>0</sub>}}{\eqn{mol C[0]}{ASCII}} and
\ifelse{html}{\out{C<sub>i</sub>}}{\eqn{mol C[i]}{ASCII}} values.
By default, they are displayed on the left side of the plot.}
}
\value{
a list of plots, one per UniqueID, drawn from flux restults (output
from the functions #' \code{\link[GoFluxYourself]{goFlux}} and
\code{\link[GoFluxYourself]{best.flux}}).
}
\description{
Returns a list plots, drawn from flux restults (output from the functions
\code{\link[GoFluxYourself]{goFlux}} and \code{\link[GoFluxYourself]{best.flux}}).
The plots are customizable.
}
\details{
In \code{plot.legend}, one may chose to display up to 5 additional parameters
in a legend above the plots. Some parameters are displayed for both the linear
model (\code{\link[GoFluxYourself]{HM.flux}}) and the non-linear model
(\code{\link[GoFluxYourself]{HM.flux}}): Mean Absolute Error (\code{MAE}),
Root Mean Square Error (\code{RMSE}), Standard Error (\code{SE}), relative
Standard Error (\code{SErel}), and coefficient of determination (\code{r2}).
The p-value (\code{LM.p.val}) is displayed for the linear model only. The kappa
(\code{HM.k}), kappa-max (\code{\link[GoFluxYourself]{k.max}}), kappa ratio
(\code{k.ratio}) and g-factor (\code{\link[GoFluxYourself]{g.factor}}) are
displayed for the Hutchinson and Mosier model only. One may chose to display
no additional parameter with \code{plot.legend = NULL}.

In \code{plot.display}, one may chose to display some parameters on the plot:
The initial and final gas concentrations (\code{C0} and \code{Ci}) for both
models, the number of observations (\code{nb.obs}) flagged, the Minimal
Detectable Flux (\code{\link[GoFluxYourself]{MDF}}), the flux term
(\code{\link[GoFluxYourself]{flux.term}}), the instrument precision (prec),
the chamber closure (\code{cham.close}) and opening (\code{cham.open})
(indicated with a green star), and the data points between chamber closure
and opening that have been removed (\code{crop}) (indicated in light red).
For manual chamber measurements, because there is no automatic chamber closure
and opening, no green stars can be displayed. In addition, \code{crop} is
only relevant if data points have been removed with the function
\code{crop.meas()} (this function is not available yet).
One may chose to display none of these parameters with \code{plot.display = NULL}.
The order in which \code{prec}, \code{flux.term}, \code{MDF} and \code{nb.obs}
are put in \code{plot.display = c()} decides the order in which they are
displayed at the bottom of the plot.

In \code{flux.unit}, remember to multiply the flux results with an appropriate
factor to convert the results from a unit to another. If kilograms of soil
were used to calculate the fluxes (see the details section of the function
\code{\link[GoFluxYourself]{goFlux}}), the units would be
\ifelse{html}{\out{µmol kg<sup>-1</sup>h<sup>-1</sup>}}{\eqn{µmol kg^{-1}s^{-1}}{ASCII}}.
To convert the units to
\ifelse{html}{\out{µmol kg<sup>-1</sup>h<sup>-1</sup>}}{\eqn{µmol kg^{-1}h^{-1}}{ASCII}}
instead, one would need to multiply the flux results by 3600 to convert from
seconds to hours. To print non-ASCII characters use Unicode. For example, to
print the Greek letter "mu" (\eqn{µ}), use the Unicode \code{\\u03BC}:
\code{flux.unit = "\\u03BCmol~kg^-1*h^-1"}.

In \code{p.val.disp}, if \code{p.val.disp = "star"}, the p-values will be
displayed as star symbols (asterisks) as follows: ***, ** or * for p-values
of p < 0.001, p < 0.01 and p < 0.05, respectively. If \code{p.val.disp = "round"},
the p-values are rounded to p < 0.001, p < 0.01 and p < 0.05. If
\code{p.val.disp = "value"}, the actual values are displayed, rounded to two
significant numbers.
}
\examples{
data(LGR_manID)
LGR_flux <- goFlux(LGR_manID, "CO2dry_ppm")
criteria <- c("MAE", "g.factor", "MDF", "SErel")
LGR_res <- best.flux(LGR_flux, criteria)
LGR_plots <- flux.plot(
  flux.results = LGR_res, dataframe = LGR_manID,
  gastype = "CO2dry_ppm", quality.check = TRUE,
  plot.legend = c("MAE", "RMSE", "k.ratio", "g.factor", "SErel"),
  plot.display = c("Ci", "C0", "MDF", "prec", "nb.obs", "flux.term"))

}
\seealso{
See also the functions \code{\link[GoFluxYourself]{goFlux}},
         \code{\link[GoFluxYourself]{best.flux}} and
         \code{\link[GoFluxYourself]{flux2pdf}}
         for more information about usage.

Look up the functions \code{\link[GoFluxYourself]{g.factor}},
         \code{\link[GoFluxYourself]{k.max}},
         \code{\link[GoFluxYourself]{MDF}},
         \code{\link[GoFluxYourself]{flux.term}},
         \code{\link[GoFluxYourself]{LM.flux}} and
         \code{\link[GoFluxYourself]{HM.flux}} for more information about
         these parameters.
}
