% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.term.R
\name{flux.term}
\alias{flux.term}
\title{Flux term}
\usage{
flux.term(V_L, P_kPa, A_cm2, T_C, H2O_mol)
}
\arguments{
\item{V_L}{numerical; total volume inside the chamber, tubes, instruments, etc. (L)}

\item{P_kPa}{numerical; atmospheric pressure (kPa)}

\item{A_cm2}{numerical; area of the soil surface inside the chamber (cm2)}

\item{T_C}{numerical; air temperature before chamber closure (Celsius)}

\item{H2O_mol}{numerical; water vapor concentration in the air before chamber closure (mol/mol)}
}
\value{
a numerical value
}
\description{
The "flux term" corrects the flux estimate for atmospheric pressure, air
temperature, the effect of water vapor, the volume of the chamber, and
the surface area covered by the chamber. This corrections changes the units
of the flux estimate from a concentration (ppm or ppb) per time (seconds)
into a concentration (ppm or ppb) per area (m2) per time (seconds):
e.g. ppm CO2 m-2 s-1.
}
\keyword{internal}
