% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux2pdf.R
\name{flux2pdf}
\alias{flux2pdf}
\title{Save plots to pdf}
\usage{
flux2pdf(plot.list, outfile = NULL, width = 11.6, height = 8.2)
}
\arguments{
\item{plot.list}{a list of plots; output from the function
\code{\link[GoFluxYourself]{flux.plot}}.}

\item{outfile}{a character string indicating the full name of the pdf
output file (including the path). Default is to save the pdf
in the working directory, with the name of the plot.list.}

\item{width}{numerical value; width of the pdf page. Default is 11.6 inches.}

\item{height}{numerical value; height of the pdf page. Default is 8.2 inches.}
}
\value{
a pdf file containing all plots from plot.list, sorted by UniqueID.
}
\description{
Save a list of plots to a pdf file, and combine plots sharing a UniqueID
onto the same page.
}
\examples{
data(example_LGR_manID)
example_LGR_flux <- goFlux(example_LGR_manID, "CO2dry_ppm")
criteria <- c("MAE", "g.factor", "kappa", "MDF", "SE.rel")
example_LGR_res <- best.flux(example_LGR_flux, criteria)
example_LGR_plots <- flux.plot(example_LGR_res, example_LGR_manID, "CO2dry_ppm")
flux2pdf(example_LGR_plots)

}
\seealso{
See also the function \code{\link[GoFluxYourself]{goFlux}},
         \code{\link[GoFluxYourself]{best.flux}} and
         \code{\link[GoFluxYourself]{flux.plot}}
         for more information about usage.
}
