% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.factor.R
\name{g.factor}
\alias{g.factor}
\title{G-factor}
\usage{
g.factor(HM.flux, LM.flux)
}
\arguments{
\item{HM.flux}{a numerical value; flux estimate from a non-linear
regression model (e.g. Hutchinson and Mosier). Calculated with
the `HM.flux()` function of this package.}

\item{LM.flux}{a numerical value; flux estimate from a linear regression
model. Calculated with the `LM.flux()` function of this package.}
}
\value{
a numerical value
}
\description{
Calculates the g-factor: the ratio between the flux estimates of a non-linear
regression model (e.g. Hutchinson and Mosier) and a linear regression model.
The g-factor should be <4 (flexible threshold), <2 (medium), or <1.25 (conservative).
}
\references{
Hüppi et al. (2018). Restricting the nonlinearity parameter in soil greenhouse gas flux calculation for more reliable flux estimates. *PloS one*, 13(7), e0200876.
}
\seealso{
Look up the functions \code{\link[GoFluxYourself]{HM.flux}} and
         \code{\link[GoFluxYourself]{LM.flux}} of this package for more
         information about these parameters.
}
\keyword{internal}
