% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.factor.R
\name{g.factor}
\alias{g.factor}
\title{G-Factor}
\usage{
g.factor(HM.flux, LM.flux)
}
\arguments{
\item{HM.flux}{a numerical value; flux estimate from a non-linear flux
calculation model (e.g. Hutchinson and Mosier; HM)}

\item{LM.flux}{a numerical value; flux estimate from a linear flux
calculation model}
}
\value{
a numerical value
}
\description{
Calculates the G-factor: the ratio between the result of a non-linear flux
calculation model (e.g. Hutchinson and Mosier; HM) and the result of a
linear flux calculation model. The g-factor should be <4 (flexible threshold),
<2 (medium), or <1.25 (conservative).
}
\keyword{internal}
