% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2508_import.R
\name{G2508_import}
\alias{G2508_import}
\title{Import function for Picarro G2508 GHG analyzer}
\usage{
G2508_import(inputfile, date.format = "ymd", timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .dat}

\item{date.format}{character string; chose one of the following: "dmy", "ymd",
or "mdy". Default is "ymd", as it is the date format from
the example data file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if save = TRUE, saves the file as RData in a RData folder
in the current working directory. If save = FALSE, returns the file
in the Console, or load in the Environment if assigned to an object.}
}
\value{
a data frame
}
\description{
Imports single raw gas measurement files from the Picarro G2508 with the
extension .dat (CO2, CH4, N2O, and H2O GHG analyzer)
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "G2508/2022/08/01/example_G2508.dat",
                         package = "GoFluxYourself")

# Run function
G2508.data <- G2508_import(inputfile = file.path)

}
\seealso{
Use the wrapper function \code{\link[GoFluxYourself]{import2RData}}
         to import multiple files from the same folder path using any instrument.

Import functions for individual instruments:
         \code{\link[GoFluxYourself]{GAIA_import}},
         \code{\link[GoFluxYourself]{LGR_import}},
         \code{\link[GoFluxYourself]{LI6400_import}},
         \code{\link[GoFluxYourself]{LI7810_import}},
         \code{\link[GoFluxYourself]{LI7820_import}},
         \code{\link[GoFluxYourself]{LI8100_import}},
         \code{\link[GoFluxYourself]{LI8200_import}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
