% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2508_import.R
\name{G2508_import}
\alias{G2508_import}
\title{Import function for Picarro G2508 GHG analyzer}
\usage{
G2508_import(inputfile, date.format = "ymd", timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{the name of a file with the extension .dat}

\item{date.format}{Date format. Chose one of the following: "dmy", "ymd", or "mdy".
Default is "ymd" as it is the date format from the example
data file provided.}

\item{timezone}{a time zone in which to import the data to POSIXct format.
Default is "UTC". Note about time zone: I recommend using
the time zone "UTC" to avoid any issue related to summer
time and winter time changes.}

\item{save}{logical. If save = TRUE, save the file as Rdata in a Rdata folder
in the current working directory. If save = FALSE, return the file
in the Console, or load in the Environment if assigned to an object.}
}
\value{
a data frame
}
\description{
Imports single raw gas measurement files from the Picarro G2508 with the
extension .dat (CO2, CH4, N2O, and H2O GHG analyzer)
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "G2508/2022/08/01/example_G2508.dat",
                         package = "GoFluxYourself")

# Run function
G2508.data <- G2508_import(inputfile = file.path)

}
\seealso{
[import2Rdata()]

[GAIA_import()]

[LGR_import()]

[LI6400_import()]

[LI7810_import()]

[LI7820_import()]

[LI8100_import()]

[LI8200_import()]
}
