% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAIA_import.R
\name{GAIA_import}
\alias{GAIA_import}
\title{Import function for GAIA synced with multiple chambers and instruments}
\usage{
GAIA_import(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  pivot = "long",
  active = TRUE,
  flag = c(7, 11),
  save = FALSE
)
}
\arguments{
\item{inputfile}{the name of a file with the extension .csv}

\item{date.format}{Date format. Chose one of the following: "dmy", "ymd", or "mdy".
Default is "ymd" as it is the date format from the example
data file provided.}

\item{timezone}{a time zone in which to import the data to POSIXct format.
Default is "UTC". Note about time zone: I recommend using
the time zone "UTC" to avoid any issue related to summer
time and winter time changes.}

\item{pivot}{Character string. Either "long" or "wide". If pivot = "long",
each column containing information about Tsoil, Tcham, SWC, PAR
and operating status (Op.stat) will be saved in a single column
per parameter. If pivot = "wide", the default display of one column
per chamber per parameter will be used.}

\item{active}{Logical. If active = TRUE, preserve data for active chambers only.}

\item{flag}{Numeric vector. Indicates the operating status that should be used
for the flux calculation. Default is flag = c(7,11), where 7 indicates
"Chamber Idle Closed Clear" and 11 indicates "Chamber Idle Closed Dark".}

\item{save}{logical. If save = TRUE, save the file as RData in a RData folder
in the current working directory. If save = FALSE, return the file
in the Console, or load in the Environment if assigned to an object.}
}
\value{
a data frame
}
\description{
Imports single raw gas measurement files from the automated chamber
ECOFlux (GAIA2TECH) with the extension .csv
(LI-7810: CO2, CH4 and H2O / LI7820: N2O and H2O)
}
\examples{
# Examples on how to use:
file.path <- system.file("extdata", "GAIA/example_GAIA.csv", package = "GoFluxYourself")

GAIA.data <- GAIA_import(inputfile = file.path)

}
\seealso{
Use the wraper function \code{\link[GoFluxYourself]{import2RData}}
         to import multiple files from the same folder path using any instrument.

Import functions for individual instruments:
         \code{\link[GoFluxYourself]{G2508_import}},
         \code{\link[GoFluxYourself]{LGR_import}},
         \code{\link[GoFluxYourself]{LI6400_import}},
         \code{\link[GoFluxYourself]{LI7810_import}},
         \code{\link[GoFluxYourself]{LI7820_import}},
         \code{\link[GoFluxYourself]{LI8100_import}},
         \code{\link[GoFluxYourself]{LI8200_import}}
}
