% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAIA_import.R
\name{GAIA_import}
\alias{GAIA_import}
\alias{import.GAIA}
\title{Import function for the automated chamber ECOFlux (GAIA2TECH) synced with
multiple chambers and instruments}
\usage{
GAIA_import(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  pivot = "long",
  active = TRUE,
  flag = c(7, 11),
  background = FALSE,
  save = FALSE,
  CH.col = "COM5A0",
  SWC.col = "1C08_Soil Moisture",
  Tsoil.col = "1C07_Soil Temperature",
  Tcham.col = "2C07_Chamber Temperature",
  PAR.col = "3C07_Sunlight",
  Op.stat.col = "0C06_OperatingStatus",
  inst1 = "XT2C00_Instrument",
  inst2 = "XT3C00_Instrument",
  inst3 = NULL,
  gas1 = c("XT2C04_CH4", "XT2C05_CO2", "XT2C06_H2O"),
  gas2 = c("XT3C04_N2O", "XT3C05_H2O"),
  gas3 = NULL,
  prec1 = c(0.6, 3.5, 45),
  prec2 = c(0.4, 45),
  prec3 = NULL,
  dry1 = T,
  dry2 = T,
  dry3 = NULL,
  manual = FALSE,
  sep = "\\t",
  skip = 1
)

import.GAIA(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  pivot = "long",
  active = TRUE,
  flag = c(7, 11),
  background = FALSE,
  save = FALSE,
  CH.col = "COM5A0",
  SWC.col = "1C08_Soil Moisture",
  Tsoil.col = "1C07_Soil Temperature",
  Tcham.col = "2C07_Chamber Temperature",
  PAR.col = "3C07_Sunlight",
  Op.stat.col = "0C06_OperatingStatus",
  inst1 = "XT2C00_Instrument",
  inst2 = "XT3C00_Instrument",
  inst3 = NULL,
  gas1 = c("XT2C04_CH4", "XT2C05_CO2", "XT2C06_H2O"),
  gas2 = c("XT3C04_N2O", "XT3C05_H2O"),
  gas3 = NULL,
  prec1 = c(0.6, 3.5, 45),
  prec2 = c(0.4, 45),
  prec3 = NULL,
  dry1 = T,
  dry2 = T,
  dry3 = NULL,
  manual = FALSE,
  sep = "\\t",
  skip = 1
)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .csv}

\item{date.format}{character string; specifies the date format found in the
raw data file. Choose one of the following: "dmy", "ymd",
or "mdy". Default is "ymd", as it is the date format from
the example data file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{pivot}{character string; either "long" or "wide". If \code{pivot = "long"},
each column containing information about \code{Tsoil},
\code{Tcham}, \code{SWC}, \code{PAR} and operating status
(\code{Op.stat}) will be saved in a single column per parameter.
If \code{pivot = "wide"}, the default display of one column per
chamber per parameter will be used.}

\item{active}{logical; if \code{active = TRUE}, preserve data for active
chambers only.}

\item{flag}{numeric vector; indicates the operating status that should be used
for the flux calculation. Default is \code{flag = c(7,11)}, where
7 indicates "Chamber Idle Closed Clear" and 11 indicates
"Chamber Idle Closed Dark".}

\item{background}{logical; if \code{background = FALSE}, removes all data from
\code{activ.cham == "Background"}.}

\item{save}{logical; if \code{save = TRUE}, saves the file as an .RData file
in a RData folder in the current working directory. If
\code{save = FALSE}, returns the file in the Console, or load in
the Environment if assigned to an object.}

\item{Op.stat.col, PAR.col, Tcham.col, Tsoil.col, SWC.col, CH.col}{character
string; a pattern to match all columns that fit the corresponding
parameter. For example, all columns containing the pattern "3C07_Sunlight"
will be renamed with the pattern "_PAR". Then, if \code{pivot = "long"},
all columns with the pattern "_PAR" will be merged together.}

\item{inst1, inst2, inst3}{character strings; a pattern to match the columns
containing the name of each instrument. By default,
\code{inst1 = "XT2C00_Instrument"}, \code{inst2 = "XT3C00_Instrument"},
and \code{inst3 = NULL}, to match the example data file provided with
the package.}

\item{gas1, gas2, gas3}{character vectors; a pattern to match the columns
containing each gas measurement. By default,
\code{gas1 = c("XT2C04_CH4", "XT2C05_CO2", "XT2C06_H2O")},
\code{gas2 = c("XT3C04_N2O", "XT3C05_H2O")}, and \code{gas3 = NULL},
to match the example data file provided with the package.}

\item{prec1, prec2, prec3}{numerical vectors; the precision of the instrument
for each gas mentioned in \code{gas1}, \code{gas2} and \code{gas3}.
By default, \code{prec1 = c(0.6, 3.5, 45)}, \code{prec2 = c(0.4, 45)},
and \code{prec3 = NULL}, to match the example data file provided with
the package. Note that the order in the arguments \code{precX} must
match the order of the arguments in \code{gasX}.}

\item{dry1, dry2, dry3}{logical; are the gas measurements compensated for water
vapor (dry fraction)? If \code{dryX = TRUE} (default), the gases are
compensated for water vapor, and will be named accordingly. For example,
the column "XT2C05_CO2" will become "CO2dry_ppm".}

\item{manual}{logical; indicates if measurements were made manually and there
are no records of chamber identification, sequence, etc. If
\code{TRUE}, the arguments \code{Op.stat.col}, \code{CH.col},
\code{active}, \code{background}, \code{flag}, and \code{pivot}
are ignored.}

\item{sep}{character string defining the field separator character. Values on
each line of the file are separated by this character. By default,
\code{sep = "\t"} for tabulation.}

\item{skip}{integer; the number of lines of the data file to skip before
beginning to read data. By default, \code{skip = 1}.}
}
\value{
A data frame containing raw data from the automated chamber
         ECOFlux (GAIA2TECH)
}
\description{
Imports single raw gas measurement files from the automated chamber
ECOFlux (GAIA2TECH) linked with up to three greenhouse gas analyzers.
Default settings are set for two LI-COR instruments: LI-7810
(\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}},
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}) and LI7820
(\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}), to match the
example data file provided with the package.
}
\details{
This function has been designed for the automated chamber ECOFlux (GAIA2TECH)
synced with up to 9 chambers and two GHG analyzers from LI-COR (LI-7810:
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}},
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}} / LI7820:
\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}). If this function
could be useful for you, but does not meet your needs, please contact the
maintainer of this package for potential adaptations.

In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For the automated chamber ECOFlux,
the date is found in the column "Titles:".

The arguments \code{PAR.col}, \code{Tcham.col}, \code{Tsoil.col} and
\code{SWC.col} correspond to different types of probes linked to the ECOFlux
chamber: PAR (sunlight), chamber temperature, soil temperature and soil water
content volumetric, respectively. The argument \code{Op.stat.col} corresponds
to the columns Operating Status of each chamber. The argument \code{CH.col}
indicates a character string preceding the chamber number for each column of
the raw data. For example, the column "COM5A010C06_OperatingStatus" in the
raw data file contains the Operating Status for the chamber 1 if
\code{CH.col = "COM5A0"} and \code{Op.stat.col = "0C06_OperatingStatus"}. If
these columns are absent from the raw data, set arguments to NULL.

Note that this function was designed for the following units in the raw file:
\itemize{
  \item kPa for atmospheric pressure
  \item volumetric water content (\%) for soil moisture
  \item Celsius for air temperature
  \item \ifelse{html}{\out{µmol photons m<sup>-2</sup>s<sup>-1</sup> for PAR}}{\eqn{µmol photons m^{-2}s^{-1} for PAR}{ASCII}}}
For each gas, the units are taken from the second row in the raw file (Units:).
If your instruments use different units, either convert the units after
import, change the settings on your instruments, or contact the maintainer of
this package for support.

Regarding the parameters \code{dryX}, in case of uncertainty, either contact
your technical support or assume that gases are compensated for water vapor,
which is normally the case.

As opposed to the other import functions, there is no option to "keep_all" with
this instrument. If you would like to import additional data using this
function, please contact the maintainer of this package for support.

The precision of the instrument is needed to restrict kappa-max
(\code{\link[goFlux]{k.max}}) in the non-linear flux calculation
(\code{\link[goFlux]{HM.flux}}). Kappa-max is inversely proportional to
instrument precision. If the precision of your instrument is unknown, it is
better to use a low value (e.g. 1 ppm for
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}, or 1 ppb for
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} and
\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}}) to allow for more
curvature, especially for water vapor fluxes, or very long measurements, that
are normally curved. The default values given for instrument precision are
the ones found online for the latest models of the
\href{https://www.licor.com/env/products/trace-gas/LI-7810}{LI-7810} and
\href{https://www.licor.com/env/products/trace-gas/LI-7820}{LI-7820},
available at the time of the creation of this function (11-2023).
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "GAIA/GAIA.csv", package = "goFlux")

# Run function
imp.GAIA <- import.GAIA(inputfile = file.path)
}
\seealso{
Use the wrapper function \code{\link[goFlux]{import2RData}}
         to import multiple files from the same folder path using any instrument.

See also, import functions for other instruments:
         \code{\link[goFlux]{import.DX4015}},
         \code{\link[goFlux]{import.EGM5}},
         \code{\link[goFlux]{import.G2201i}},
         \code{\link[goFlux]{import.G2508}},
         \code{\link[goFlux]{import.G4301}},
         \code{\link[goFlux]{import.GasmetPD}},
         \code{\link[goFlux]{import.GT5000}},
         \code{\link[goFlux]{import.LI6400}},
         \code{\link[goFlux]{import.LI7810}},
         \code{\link[goFlux]{import.LI7820}},
         \code{\link[goFlux]{import.LI8100}},
         \code{\link[goFlux]{import.LI8200}},
         \code{\link[goFlux]{import.LI8250}},
         \code{\link[goFlux]{import.N2OM1}},
         \code{\link[goFlux]{import.uCH4}},
         \code{\link[goFlux]{import.uN2O}},
         \code{\link[goFlux]{import.UGGA}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
