% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAIA_import.R
\name{GAIA_import}
\alias{GAIA_import}
\title{Import function for GAIA synced with multiple chambers and instruments}
\usage{
GAIA_import(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  pivot = "long",
  active = TRUE,
  flag = c(7, 11),
  background = FALSE,
  save = FALSE,
  CH.col = "COM5A0",
  SWC.col = "1C08_Soil.Moisture",
  Tsoil.col = "1C07_Soil.Temperature",
  Tcham.col = "2C07_Chamber.Temperature",
  PAR.col = "3C07_Sunlight",
  Op.stat.col = "0C06_OperatingStatus",
  CO2.col = "XT2C05_CO2",
  CH4.col = "XT2C04_CH4",
  H2O1.col = "XT2C06_H2O",
  N2O.col = "XT3C04_N2O",
  H2O2.col = "XT3C05_H2O"
)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .csv}

\item{date.format}{date format; the date format used in the raw data file.
Chose one of the following: "dmy", "ymd", or "mdy". Default
is "ymd", as it is the date format from the example data
file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{pivot}{character string; either "long" or "wide". If \code{pivot = "long"},
each column containing information about Tsoil, Tcham, SWC, PAR
and operating status (Op.stat) will be saved in a single column
per parameter. If \code{pivot = "wide"}, the default display of
one column per chamber per parameter will be used.}

\item{active}{logical; if \code{active = TRUE}, preserve data for active
chambers only.}

\item{flag}{numeric vector; indicates the operating status that should be used
for the flux calculation. Default is \code{flag = c(7,11)}, where
7 indicates "Chamber Idle Closed Clear" and 11 indicates
"Chamber Idle Closed Dark".}

\item{background}{logical; if \code{background = FALSE}, removes all data from
\code{activ.cham == "Background"}.}

\item{save}{logical; if \code{save = TRUE}, saves the file as RData in a RData
folder in the current working directory. If \code{save = FALSE},
returns the file in the Console, or load in the Environment if
assigned to an object.}

\item{Op.stat.col, PAR.col, Tcham.col, Tsoil.col, SWC.col, CH.col}{character string;
a pattern to match all columns that fit the corresponding parameter. For
example, all columns containing the pattern "3C07_Sunlight" will be
renamed with the pattern "_PAR". Then, if \code{pivot = "long"}, all
columns with the pattern "_PAR" will be merged together.}

\item{CO2.col, CH4.col, H2O1.col, N2O.col, H2O2.col}{character string; a pattern
to match the columns containing the corresponding gas measurements.
\code{H2O1.col} must be the same instrument as \code{CO2.col} and
\code{CH4.col}, and \code{H2O2.col} must be the same instrument as
\code{N2O.col}.}
}
\value{
a data frame containing raw data from the automated chamber
         ECOFlux (GAIA2TECH)
}
\description{
Imports single raw gas measurement files from the automated chamber
ECOFlux (GAIA2TECH) with the extension .csv
(LI-7810: \ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}},
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}} / LI7820:
\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}})
}
\details{
This function has been designed for the automated chamber ECOFlux (GAIA2TECH)
synced with up to 9 chambers and two GHG analyzers from LI-COR (LI-7810:
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}},
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}} / LI7820:
\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}). If this function
could be useful for you, but does not meet your needs, please contact the
maintainer of this package for potential adaptations.

In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For the instrument GAIA the
date is found in the column "Titles:".

Note that this function was designed for the following default units:
\itemize{
  \item ppm for \ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} and
  \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}
  \item ppb for \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} and
  \ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}}
  \item kPa for pressure
  \item volumetric water content (\%) for soil moisture
  \item Celsius for temperature}
If your instrument uses different units, either convert the units after import,
change the settings on your instrument, or contact the maintainer of this
package for support.

As opposed to the other import functions, there is no option to "keep_all" with
this instrument. If you would like to import additional data using this
function, please contact the maintainer of this package for support.
}
\examples{
# Examples on how to use:
file.path <- system.file("extdata", "GAIA/example_GAIA.csv", package = "GoFluxYourself")

GAIA.data <- GAIA_import(inputfile = file.path)

}
\seealso{
Use the wrapper function \code{\link[GoFluxYourself]{import2RData}}
         to import multiple files from the same folder path using any instrument.

Import functions for individual instruments:
         \code{\link[GoFluxYourself]{DX4015_import}},
         \code{\link[GoFluxYourself]{G2508_import}},
         \code{\link[GoFluxYourself]{LGR_import}},
         \code{\link[GoFluxYourself]{LI6400_import}},
         \code{\link[GoFluxYourself]{LI7810_import}},
         \code{\link[GoFluxYourself]{LI7820_import}},
         \code{\link[GoFluxYourself]{LI8100_import}},
         \code{\link[GoFluxYourself]{LI8200_import}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
