% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAIA_import.R
\name{GAIA_import}
\alias{GAIA_import}
\alias{import.GAIA}
\title{Import function for the automated chamber ECOFlux (GAIA2TECH) synced with
multiple chambers and instruments}
\usage{
GAIA_import(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  pivot = "long",
  active = TRUE,
  flag = c(7, 11),
  background = FALSE,
  save = FALSE,
  prec = c(3.5, 0.6, 0.4, 45, 45),
  CH.col = "COM5A0",
  SWC.col = "1C08_Soil.Moisture",
  Tsoil.col = "1C07_Soil.Temperature",
  Tcham.col = "2C07_Chamber.Temperature",
  PAR.col = "3C07_Sunlight",
  Op.stat.col = "0C06_OperatingStatus",
  CO2.col = "XT2C05_CO2",
  CH4.col = "XT2C04_CH4",
  H2O1.col = "XT2C06_H2O",
  H2O1.name = "H2O_LI7810",
  N2O.col = "XT3C04_N2O",
  H2O2.col = "XT3C05_H2O",
  H2O2.name = "H2O_LI7820",
  sep = "\\t",
  skip = 1
)

import.GAIA(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  pivot = "long",
  active = TRUE,
  flag = c(7, 11),
  background = FALSE,
  save = FALSE,
  prec = c(3.5, 0.6, 0.4, 45, 45),
  CH.col = "COM5A0",
  SWC.col = "1C08_Soil.Moisture",
  Tsoil.col = "1C07_Soil.Temperature",
  Tcham.col = "2C07_Chamber.Temperature",
  PAR.col = "3C07_Sunlight",
  Op.stat.col = "0C06_OperatingStatus",
  CO2.col = "XT2C05_CO2",
  CH4.col = "XT2C04_CH4",
  H2O1.col = "XT2C06_H2O",
  H2O1.name = "H2O_LI7810",
  N2O.col = "XT3C04_N2O",
  H2O2.col = "XT3C05_H2O",
  H2O2.name = "H2O_LI7820",
  sep = "\\t",
  skip = 1
)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .csv}

\item{date.format}{character string; specifies the date format found in the
raw data file. Choose one of the following: "dmy", "ymd",
or "mdy". Default is "ymd", as it is the date format from
the example data file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{pivot}{character string; either "long" or "wide". If \code{pivot = "long"},
each column containing information about \code{Tsoil},
\code{Tcham}, \code{SWC}, \code{PAR} and operating status
(\code{Op.stat}) will be saved in a single column per parameter.
If \code{pivot = "wide"}, the default display of one column per
chamber per parameter will be used.}

\item{active}{logical; if \code{active = TRUE}, preserve data for active
chambers only.}

\item{flag}{numeric vector; indicates the operating status that should be used
for the flux calculation. Default is \code{flag = c(7,11)}, where
7 indicates "Chamber Idle Closed Clear" and 11 indicates
"Chamber Idle Closed Dark".}

\item{background}{logical; if \code{background = FALSE}, removes all data from
\code{activ.cham == "Background"}.}

\item{save}{logical; if \code{save = TRUE}, saves the file as an .RData file
in a RData folder in the current working directory. If
\code{save = FALSE}, returns the file in the Console, or load in
the Environment if assigned to an object.}

\item{prec}{numerical vector; the precision of the instrument for each gas,
in the following order: "CO2dry_ppm", "CH4dry_ppb", "N2Odry_ppb"
and "H2O_ppm" from each analyzer. The default is
\code{prec = c(3.5, 0.6, 0.4, 45, 45)}. For each gas missing, the
value is ignored.}

\item{Op.stat.col, PAR.col, Tcham.col, Tsoil.col, SWC.col, CH.col}{character
string; a pattern to match all columns that fit the corresponding
parameter. For example, all columns containing the pattern "3C07_Sunlight"
will be renamed with the pattern "_PAR". Then, if \code{pivot = "long"},
all columns with the pattern "_PAR" will be merged together.}

\item{CO2.col, CH4.col, H2O1.col, N2O.col, H2O2.col}{character string; a pattern
to match the columns containing the corresponding gas measurements.
\code{H2O1.col} must be the same instrument as \code{CO2.col} and
\code{CH4.col}, and \code{H2O2.col} must be the same instrument as
\code{N2O.col}. If columns are absent from the raw data, the arguments
are ignored.}

\item{H2O1.name, H2O2.name}{character string; how to name the water vapor columns
for each instrument. Ideally, use \code{H201.name = "H2O_instrument.name"}.
For example, \code{H201.name = "H2O_LI7810"}.}

\item{sep}{character string defining the field separator character. Values on
each line of the file are separated by this character. By default,
\code{sep = "\t"} for tabulation.}

\item{skip}{integer; the number of lines of the data file to skip before
beginning to read data. By default, \code{skip = 1}.}
}
\value{
A data frame containing raw data from the automated chamber
         ECOFlux (GAIA2TECH)
}
\description{
Imports single raw gas measurement files from the automated chamber
ECOFlux (GAIA2TECH) linked with up to two instruments and up to five gases
(by default used with LI-7810:
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}},
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}} / LI7820:
\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}).
}
\details{
This function has been designed for the automated chamber ECOFlux (GAIA2TECH)
synced with up to 9 chambers and two GHG analyzers from LI-COR (LI-7810:
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}},
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}} / LI7820:
\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}). If this function
could be useful for you, but does not meet your needs, please contact the
maintainer of this package for potential adaptations.

In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For the instrument GAIA the
date is found in the column "Titles:".

The arguments \code{PAR.col}, \code{Tcham.col},
\code{Tsoil.col} and \code{SWC.col} correspond to different types of probes
linked to the ECOFlux chamber: PAR, chamber temperature, soil
temperature and soil water content volumetric, respectively. The argument
\code{Op.stat.col} corresponds to the columns Operating Status of each
chamber. The argument \code{CH.col} indicates a character string preceding
the chamber number for each column of the raw data. For example, the column
"COM5A010C06_OperatingStatus" in the raw data file contains the Operating
Status for the chamber 1 if \code{CH.col = "COM5A0"} and
\code{Op.stat.col = "0C06_OperatingStatus"}. If columns are absent from the
raw data, the arguments are ignored.

Note that this function was designed for the following units in the raw file:
\itemize{
  \item ppm for \ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} and
  \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}
  \item ppb for \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} and
  \ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}}
  \item kPa for pressure
  \item volumetric water content (\%) for soil moisture
  \item Celsius for temperature
  \item \ifelse{html}{\out{µmol photons m<sup>-2</sup>s<sup>-1</sup> for PAR}}{\eqn{µmol photons m^{-2}s^{-1} for PAR}{ASCII}}}
If your instruments use different units, either convert the units after
import, change the settings on your instrument, or contact the maintainer of
this package for support.

As opposed to the other import functions, there is no option to "keep_all" with
this instrument. If you would like to import additional data using this
function, please contact the maintainer of this package for support.

The precision of the instrument is needed to restrict kappa-max
(\code{\link[goFlux]{k.max}}) in the non-linear flux calculation
(\code{\link[goFlux]{HM.flux}}). Kappa-max is inversely proportional to
instrument precision. If the precision of your instrument is unknown, it is
better to use a low value (e.g. 1 ppm for
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}, or 1 ppb for
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} and
\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}}) to allow for more
curvature, especially for water vapor fluxes, or very long measurements, that
are normally curved. The default values given for instrument precision are
the ones found online for the latest models of the
\href{https://www.licor.com/env/products/trace-gas/LI-7810}{LI-7810} and
\href{https://www.licor.com/env/products/trace-gas/LI-7820}{LI-7820},
available at the time of the creation of this function (11-2023). Five values
are required for this argument and their position in the vector is important.
If you are using the function with less than five gases, the other values are
simply ignored.
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "GAIA/GAIA.csv", package = "goFlux")

# Run function
imp.GAIA <- import.GAIA(inputfile = file.path)
}
\seealso{
Use the wrapper function \code{\link[goFlux]{import2RData}}
         to import multiple files from the same folder path using any instrument.

See also, import functions for other instruments:
         \code{\link[goFlux]{import.DX4015}},
         \code{\link[goFlux]{import.EGM5}},
         \code{\link[goFlux]{import.G2508}},
         \code{\link[goFlux]{import.G4301}},
         \code{\link[goFlux]{import.LI6400}},
         \code{\link[goFlux]{import.LI7810}},
         \code{\link[goFlux]{import.LI7820}},
         \code{\link[goFlux]{import.LI8100}},
         \code{\link[goFlux]{import.LI8200}},
         \code{\link[goFlux]{import.N2OM1}},
         \code{\link[goFlux]{import.uCH4}},
         \code{\link[goFlux]{import.uN2O}},
         \code{\link[goFlux]{import.UGGA}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
