% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goFlux.R
\name{goFlux}
\alias{goFlux}
\title{goFlux: a user-friendly GHG fluxes calculation tool}
\usage{
goFlux(
  dataframe,
  gastype,
  H2O_col = "H2O_ppm",
  prec = NULL,
  Area = NULL,
  offset = NULL,
  Vtot = NULL,
  Vcham = NULL,
  Pcham = NULL,
  Tcham = NULL,
  k.mult = 1
)
}
\arguments{
\item{dataframe}{a data.frame containing gas measurements (see \code{gastype}
below), water vapor measurements (see \code{H2O_col} below)
and the following columns: \code{UniqueID}, \code{Etime},
\code{Vtot}, \code{Area}, \code{Pcham}, \code{Tcham} and
\code{flag} (see the parameters \code{Vtot}, \code{Area},
\code{Pcham} and \code{Tcham} below for more details).}

\item{gastype}{character string; specifies which column should be used for the
flux calculations. Must be one of the following: "CO2dry_ppm",
"CH4dry_ppb", "N2Odry_ppb" or "H2O_ppm".}

\item{H2O_col}{character string; specifies which column should be used to
subtract the effect of water vapor in the chamber space.}

\item{prec}{numerical value; precision of the instruments. Units must be the
same as \code{gastype}. Default values for
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}},
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}},
\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}
are based on the LI-COR instruments:
\itemize{
  \item \ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} = 3.5 ppm;
  \item \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} = 0.6 ppb;
  \item \ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}} = 0.4 ppb;
  \item \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}} = 45 ppm.
} \cr
For other instruments, you must manually specify the precision of
the instrument. If using the ultra-portable GGA (GLA132 series):
\itemize{
  \item \ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} = 0.3 ppm;
  \item \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} = 1.4 ppb;
  \item \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}} = 50 ppm.
} \cr
If using the micro  ultra-portable GGA (GLA131 series):
\itemize{
  \item \ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}} = 0.35 ppm;
  \item \ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} = 0.9 ppb;
  \item \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}} = 200 ppm.
}}

\item{Area}{numerical value; area of the soil surface inside the chamber
\ifelse{html}{\out{(cm<sup>2</sup>)}}{\eqn{(cm^2)}{ASCII}}.
Alternatively, provide the column 'Area' in dataframe if \strong{Area} is
different between samples.}

\item{offset}{numerical value; height between the soil surface and the chamber
 (cm). Alternatively, provide the column offset in dataframe if
 offset is different between samples. OPTIONAL: only used if Vtot
is missing.}

\item{Vtot}{numerical value; total volume inside the chamber, tubes, instruments,
etc. (L). Alternatively, provide the column Vtot in dataframe if
Vtot is different between samples. If Vtot is missing, the function
will calculate it from Area, Vcham and offset.}

\item{Vcham}{numerical value; volume inside the chamber, tubes and instruments
(L). Alternatively, provide the column Vcham in dataframe if
Vcham is different between samples. OPTIONAL: only used if Vtot
is missing.}

\item{Pcham}{numerical value; pressure inside the chamber (kPa).
Alternatively, provide the column Pcham in dataframe if
Pcham is different between samples.}

\item{Tcham}{numerical value; temperature inside the chamber (Celcius).
Alternatively, provide the column Tcham in dataframe if
Tcham is different between samples.}

\item{k.mult}{numerical value; a multiplier for the allowed kappa-max.
kappa-max is the maximal curvature (kappa) of the non-linear
regression (Hutchinson and Mosier model) allowed for a each
flux measurement. Default setting is [k.mult = 1].}
}
\value{
a data frame
}
\description{
A wrapper function to calculate GHG fluxes from static chamber measurements.
Calculates linear and non-linear fluxes (Hutchinson and Mosier model; HM).
}
\details{
Flux estimate units are nmol/m^~2~^*s (if initial concentration is ppm,
e.g. CO2dry_ppm) and µmol/m^~2~^*s (if initial concentration is ppb,
e.g. CH4dry_ppm).

The function `k.max()` calculates the maximal curvature (kappa) of the non-linear
model (Hutchinson and Mosier) allowed for each flux measurements. k.max
is calculated based on the minimal detectable flux (MDF), the linear
flux estimate and the measurement time. The unit of the kappa-max is s^~-1~^.

The function `MDF()` calculates the minimal detectable flux (MDF) based on
instrument precision, measurements time, and the number of measurement points.
}
\examples{
data(example_LGR_manID)
goFlux(example_LGR_manID, "CO2dry_ppm")
goFlux(example_LGR_manID, "CH4dry_ppb")
goFlux(example_LGR_manID, "H2O_ppm")

}
\seealso{
Look up the functions \code{\link[GoFluxYourself]{MDF}},
         \code{\link[GoFluxYourself]{flux.term}},
         \code{\link[GoFluxYourself]{g.factor}},
         \code{\link[GoFluxYourself]{k.max}},
         \code{\link[GoFluxYourself]{HM.flux}} and
         \code{\link[GoFluxYourself]{LM.flux}} of this package for more
         information about these parameters.
}
