% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goFlux.R
\name{goFlux}
\alias{goFlux}
\title{goFlux: user-friendly GHG fluxes calculation tool}
\usage{
goFlux(
  dataframe,
  gastype,
  H2O_col = "H2O_ppm",
  prec = NULL,
  Area = NULL,
  offset = NULL,
  Vtot = NULL,
  Vcham = NULL,
  Pcham = NULL,
  Tcham = NULL,
  k.ratio = 1,
  Ci.lim = NULL,
  C0.lim = NULL
)
}
\arguments{
\item{dataframe}{a data.frame containing gas measurements (see gastype below),
water vapor measurements (see H2O_col below) and the following:
UniqueID, Etime, Vtot, Area, Pcham, Tcham, flag}

\item{gastype}{character string. Specify which column should be used for the
flux calculations. Must be one of the following: "CO2dry_ppm",
"CH4dry_ppb", "N2Odry_ppb" or "H2O_ppm".}

\item{H2O_col}{character string. Specify which column should be used to
subtract the effect of water vapor in the chamber space.}

\item{prec}{numerical value; precision of the instruments. Default values for
CO2, CH4, N2O and H2O are based on the LI-COR instruments:
CO2 = 3.5 ppm; CH4 = 0.6 ppb; N2O = 0.4 ppb; H2O = 45 ppm.
Units must be the same as gastype. For Los Gatos Research instruments,
if using the ultra-portable GGA (GLA132 series):
prec = CO2: 0.3 ppm; CH4: 1.4 ppb; H2O: 50 ppm.
If using the micro  ultra-portable GGA (GLA131 series):
prec = CO2: 0.35 ppm; CH4: 0.9 ppb; H2O: 200 ppm.}

\item{Area}{numerical value; area of the soil surface inside the chamber (cm2).
Alternatively, provide the column Area in dataframe if Area is
different between samples.}

\item{offset}{numerical value; height between the soil surface and the chamber
(cm). Alternatively, provide the column offset in dataframe if
offset is different between samples.}

\item{Vtot}{numerical value; total volume inside the chamber, tubes, instruments,
etc. (L). Alternatively, provide the column Vtot in dataframe if
Vtot is different between samples.}

\item{Vcham}{numerical value; volume inside the chamber, tubes and instruments
(L). Alternatively, provide the column Vcham in dataframe if
Vcham is different between samples.}

\item{Pcham}{numerical value; pressure inside the chamber (kPa).
Alternatively, provide the column Pcham in dataframe if
Pcham is different between samples.}

\item{Tcham}{numerical value; temperature inside the chamber (Celcius).
Alternatively, provide the column Tcham in dataframe if
Tcham is different between samples.}

\item{k.ratio}{numerical; a multiplier for the allowed k.max.
Default is k.ratio = 1.}

\item{Ci.lim}{numerical vector of length 2; inferior and superior limits of
the maximal concentration at the end of the measurement.
Same units as gastype. If not provided, calculated based on
Ci of LM.flux().}

\item{C0.lim}{numerical vector of length 2; inferior and superior limits of
the intercept (initial concentration). Same units as gastype.
If not provided, calculated based on C0 of LM.flux().}
}
\value{
a data frame
}
\description{
A wrapper function to calculate GHG fluxes from closed chamber measurements.
Calculates linear and non-linear fluxes (Hutchinson and Mosier model; HM).
}
\details{
Flux estimate units are nmol/m2*s (if initial concentration is ppm, e.g. CO2dry_ppm)
and µmol/m2*s (if initial concentraion is ppb, e.g. CH4dry_ppm).

[k.max()] calculates the maximal curvature (kappa) of the non-linear model
(Hutchinson and Mosier model) allowed for a certain flux measurements. k.max
is calculated based on the minimal detectable flux (MDF), the linear
flux estimate and the measurement time. The units of the kappa-max is s-1.

[MDF()] calculates the minimal detectable flux (MDF) based on instrument
precision, measurements time, and the number of measurement points.
}
\examples{
data(example_LGR_manID)
goFlux(example_LGR_manID, "CO2dry_ppm")
goFlux(example_LGR_manID, "CH4dry_ppb")
goFlux(example_LGR_manID, "H2O_ppm")

}
\seealso{
[flux.term()]

[MDF()]

[LM.flux()]

[HM.flux()]

[g.factor()]

[k.max()]
}
