% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goFlux.R
\name{goFlux}
\alias{goFlux}
\title{goFlux: a user-friendly GHG fluxes calculation tool}
\usage{
goFlux(
  dataframe,
  gastype,
  H2O_col = "H2O_ppm",
  prec = NULL,
  Area = NULL,
  offset = NULL,
  Vtot = NULL,
  Vcham = NULL,
  Pcham = NULL,
  Tcham = NULL
)
}
\arguments{
\item{dataframe}{a data.frame containing gas measurements (see gastype below),
water vapor measurements (see H2O_col below) and the following:
UniqueID, Etime, Vtot, Area, Pcham, Tcham, flag}

\item{gastype}{character string; specifies which column should be used for the
flux calculations. Must be one of the following: "CO2dry_ppm",
"CH4dry_ppb", "N2Odry_ppb" or "H2O_ppm".}

\item{H2O_col}{character string; specifies which column should be used to
subtract the effect of water vapor in the chamber space.}

\item{prec}{numerical value; precision of the instruments. Default values for
CO<sub>2</sub>, CH<sub>4</sub>, N<sub>2</sub>O and H<sub>2</sub>O
are based on the LI-COR instruments: CO<sub>2</sub> = 3.5 ppm;
CH<sub>4</sub> = 0.6 ppb; N<sub>2</sub>O = 0.4 ppb; H<sub>2</sub>O = 45 ppm.
Units must be the same as gastype. For Los Gatos Research instruments,
you must manually specify the precision of the instruments.
If using the ultra-portable GGA (GLA132 series): prec = CO<sub>2</sub>:
0.3 ppm; CH<sub>4</sub>: 1.4 ppb; H<sub>2</sub>O: 50 ppm.
If using the micro  ultra-portable GGA (GLA131 series): prec =
CO<sub>2</sub>: 0.35 ppm; CH<sub>4</sub>: 0.9 ppb; H<sub>2</sub>O: 200 ppm.}

\item{Area}{numerical value; area of the soil surface inside the chamber
(cm~2~). Alternatively, provide the column Area in
dataframe if Area is different between samples.}

\item{offset}{numerical value; height between the soil surface and the chamber
(cm). Alternatively, provide the column offset in dataframe if
offset is different between samples.}

\item{Vtot}{numerical value; total volume inside the chamber, tubes, instruments,
etc. (L). Alternatively, provide the column Vtot in dataframe if
Vtot is different between samples.}

\item{Vcham}{numerical value; volume inside the chamber, tubes and instruments
(L). Alternatively, provide the column Vcham in dataframe if
Vcham is different between samples.}

\item{Pcham}{numerical value; pressure inside the chamber (kPa).
Alternatively, provide the column Pcham in dataframe if
Pcham is different between samples.}

\item{Tcham}{numerical value; temperature inside the chamber (Celcius).
Alternatively, provide the column Tcham in dataframe if
Tcham is different between samples.}
}
\value{
a data frame
}
\description{
A wrapper function to calculate GHG fluxes from static chamber measurements.
Calculates linear and non-linear fluxes (Hutchinson and Mosier model; HM).
}
\details{
Flux estimate units are nmol/m~2~*s (if initial concentration is ppm,
e.g. CO2dry_ppm) and µmol/m~2~*s (if initial concentration is ppb,
e.g. CH4dry_ppm).

The function `k.max()` calculates the maximal curvature (kappa) of the non-linear
model (Hutchinson and Mosier) allowed for each flux measurements. k.max
is calculated based on the minimal detectable flux (MDF), the linear
flux estimate and the measurement time. The unit of the kappa-max is s-1.

The function `MDF()` calculates the minimal detectable flux (MDF) based on
instrument precision, measurements time, and the number of measurement points.
}
\examples{
data(example_LGR_manID)
goFlux(example_LGR_manID, "CO2dry_ppm")
goFlux(example_LGR_manID, "CH4dry_ppb")
goFlux(example_LGR_manID, "H2O_ppm")

}
\seealso{
Look up the functions \code{\link[GoFluxYourself]{MDF}},
         \code{\link[GoFluxYourself]{flux.term}},
         \code{\link[GoFluxYourself]{g.factor}},
         \code{\link[GoFluxYourself]{k.max}},
         \code{\link[GoFluxYourself]{HM.flux}} and
         \code{\link[GoFluxYourself]{LM.flux}} of this package for more
         information about these parameters.
}
