% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HM.flux.R
\name{HM.flux}
\alias{HM.flux}
\title{Non-linear model for flux calculation: the Hutchinson and Mosier model}
\usage{
HM.flux(
  gas.meas,
  time.meas,
  flux.term,
  k.max,
  Ci = NULL,
  C0 = NULL,
  k = 0.005,
  k.ratio = 1,
  Ci.lim = NULL,
  C0.lim = NULL
)
}
\arguments{
\item{gas.meas}{numerical vector containing gas measurements (ppm or ppb)}

\item{time.meas}{numerical vector containing time stamps (seconds)}

\item{flux.term}{numerical value; flux term calculated with the function flux.term}

\item{k.max}{numerical; kappa-max is the maximal curvature allowed in the
Hutchinson and Mosier model. Calculated with the k.max function.}

\item{Ci}{numerical; gas concentration after i seconds (ppm or ppb). For forest
soil conditions, use these estimates for the following gastypes:
"CO2dry_ppm": Ci = 1000 ppm; "CH4dry_ppb": Ci = 1800 ppb;
"N2Odry_ppb": Ci = 500 ppb; "H2O_ppm": Ci = 20000 ppm.}

\item{C0}{numerical; initial gas concentration at time 0 second (ppm or ppb)
For forest soil conditions, use these estimates for the following
gastypes: "CO2dry_ppm": C0 = 380 ppm; "CH4dry_ppb": C0 = 2200 ppb;
"N2Odry_ppb": C0 = 250 ppb; "H2O_ppm": C0 = 10000 ppm.}

\item{k}{numerical; kappa gives the curvature of the regression line (s-1).
The default parameter for kappa is k = 0.005 s-1. k = 0.005 is the
minimal value of kappa if LM.flux = MDF and t = 180 secs.}

\item{k.ratio}{numerical; a multiplier for the allowed k.max. Default is
k.ratio = 1.}

\item{Ci.lim}{numerical vector of length 2; inferior and superior limits of
the maximal concentration at the end of the measurement.
Same units as Ci.}

\item{C0.lim}{numerical vector of length 2; inferior and superior limits of
the intercept (initial concentration). Same units as C0.}
}
\value{
a data.frame
}
\description{
Estimates a gas flux with a non-linear model, and then extracts values from
the model fit, calculates standard error, R2, p-value and RMSE.
}
\keyword{internal}
