% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HM.flux.R
\name{HM.flux}
\alias{HM.flux}
\title{Non-linear model for flux calculation}
\usage{
HM.flux(
  gas.meas,
  time.meas,
  flux.term,
  k.max,
  Ct = NULL,
  C0 = NULL,
  k.mult = 1
)
}
\arguments{
\item{gas.meas}{numerical vector containing gas measurements (ppm or ppb)}

\item{time.meas}{numerical vector containing time stamps (seconds)}

\item{flux.term}{numerical value; flux term calculated with the function
\code{\link[GoFluxYourself]{flux.term}}}

\item{k.max}{numerical; kappa-max is the maximal curvature allowed in the
Hutchinson and Mosier model. Calculated with the
\code{\link[GoFluxYourself]{k.max}} function.}

\item{Ct}{numerical; gas concentration after \emph{t} seconds (ppm or ppb).
For typical forest soil conditions, use these estimates for the
following \code{gastypes}:
\itemize{
    \item "CO2dry_ppm": \code{Ct = 1000 ppm}
    \item "CH4dry_ppb": \code{Ct = 1800 ppb}
    \item "N2Odry_ppb": \code{Ct = 500 ppb}
    \item "H2O_ppm": \code{Ct = 20000 ppm}
  }}

\item{C0}{numerical; initial gas concentration at time 0 second (ppm or ppb)
For typical forest soil conditions, use these estimates for the
following \code{gastypes}:
\itemize{
    \item "CO2dry_ppm": \code{C0 = 380 ppm}
    \item "CH4dry_ppb": \code{C0 = 2200 ppb}
    \item "N2Odry_ppb": \code{C0 = 250 ppb}
    \item "H2O_ppm": \code{C0 = 10000 ppm}
  }}

\item{k.mult}{numerical; a multiplier for the allowed k.max. Default is no
multiplier (\code{k.mult = 1}).}
}
\value{
Returns a data frame with 10 columns: non-linear flux estimate,
         initial gas concentration (C0), final gas concentration (Ct), slope
         at \code{t=0}, mean absolute error (MAE), root mean square error
         (RMSE), standard error (se), relative se (se.rel),
         \ifelse{html}{\out{r<sup>2</sup>}}{\eqn{r^2}{ASCII}},
         and curvature (kappa).
}
\description{
Estimates a gas flux with a non-linear regression, the Hutchinson and Mosier
model, and then extracts values from the model fit.
}
\details{
Flux estimate units are
\ifelse{html}{\out{µmol/m<sup>2</sup>s}}{\eqn{µmol/m^{2}s}{ASCII}}
(if initial concentration is ppm, e.g. CO2dry_ppm) and
\ifelse{html}{\out{nmol/m<sup>2</sup>s}}{\eqn{nmol/m^{2}s}{ASCII}}
(if initial concentration is ppb, e.g. CH4dry_ppb).
}
\references{
Hüppi et al. (2018). Restricting the nonlinearity parameter in
soil greenhouse gas flux calculation for more reliable flux estimates.
\emph{PloS one}, 13(7), e0200876.
}
\seealso{
Look up the functions \code{\link[GoFluxYourself]{flux.term}} and
         \code{\link[GoFluxYourself]{k.max}} of this package for more
         information about these parameters.

See also the function \code{\link[GoFluxYourself]{LM.flux}} for
         information about the linear regression model used in this package.
}
\keyword{internal}
