% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import2RData.R
\name{import2RData}
\alias{import2RData}
\title{Looping function for quick import of multiple raw gas measurement files}
\usage{
import2RData(path, instrument, date.format, timezone = "UTC", keep_all, prec)
}
\arguments{
\item{path}{character string; a folder path containing all files to be imported.
Ideally, to avoid any errors, the folder should only contain the
files to be imported.}

\item{instrument}{character string; specifies which instrument was used to
generate the files contained in the folder path. Choose one
of the following: "DX4015", "EGM5", "G2508", "G4301" "GAIA",
"LI-6400", "LI-7810", "LI-7820", "LI-8100", "LI-8200",
"LGR" or "N2OM1". For more information about an instrument,
see the section "See Also" below.}

\item{date.format}{character string; specifies the date format found in the
raw data file. Choose one of the following: "dmy", "ymd",
or "mdy".}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{keep_all}{logical; if \code{keep_all = TRUE}, keep all columns from the raw
file. The default is \code{keep_all = FALSE}, and columns that
are not necessary for gas flux calculation are removed.}

\item{prec}{numerical vector; the precision of the instrument for each gas.
Look at the examples below, or the help for each import function
of each instrument, to know what values to use.}
}
\value{
A data frame saved as RData in a newly created folder, RData, into
the working directory.
}
\description{
Imports all raw gas measurement files contained in a folder.
Adapted to multiple greenhouse gas analyzers and other instruments:
\itemize{
\item \strong{LI-COR}: LI-6400, LI-7810, LI-7820, LI-8100, LI-8200 (smart chamber)
\item \strong{Los Gatos Research (LGR)}: ultra-portable GGA (GLA132 series),
micro ultra-portable GGA (GLA131 series)
and N2OM1 (GLA151 series)
\item \strong{GAIA2TECH (DMR)} automated chamber ECOFlux
\item \strong{Picarro}: G2508 and G4301
\item \strong{Gasmet}: DX4015
\item \strong{PP-Systems}: EGM-5
}
}
\details{
In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For most instruments, the date is
one of the columns in the raw data file:
\itemize{
\item DX4015: column Date
\item EGM5: column Date
\item G2508: column DATE
\item G4301: column DATE
\item GAIA: column Titles:
\item LGR: column Time
\item LI-6400: (see comment below)
\item LI-7810: column DATE
\item LI-7820: column DATE
\item LI-8100: column Date
\item LI-8200: (see comment below)
\item N2OM1: column Time
}
For the instrument LI-6400, the date is found in one of the first lines in
a format containing abbreviations, for example "Thr Aug 6 2020", which would
be the date format "mdy". For the instrument LI-8200, the date is found under
one of the measurements, next to '"Date":'.
}
\examples{
# Examples on how to use it with all instruments.
# The default time zone "UTC" is used in all cases.

# with the Gasmet instrument DX4015
file.path <- system.file("extdata/DX4015", package = "goFlux")
import2RData(path = file.path, instrument = "DX4015",
             date.format = "ymd", keep_all = FALSE,
             prec = c(1.6, 23, 13, 2, 23, 33))

# with the PP-Systems EGM-5
file.path <- system.file("extdata/EGM5", package = "goFlux")
import2RData(path = file.path, instrument = "EGM5",
             date.format = "dmy", keep_all = FALSE,
             prec = c(3, 1, 500))

# with the Picarro instrument G2508
file.path <- system.file("extdata/G2508", package = "goFlux")
import2RData(path = file.path, instrument = "G2508",
             date.format = "ymd", keep_all = FALSE,
             prec = c(0.24, 0.3, 5, 0.16, 500))

# with the Picarro instrument G4301
file.path <- system.file("extdata/G4301", package = "goFlux")
import2RData(path = file.path, instrument = "G4301",
             date.format = "ymd", keep_all = FALSE,
             prec = c(0.025, 0.1, 10))

# with the automated chamber ECOFlux (GAIA2TECH)
# with this instrument, "keep_all" is not a valid argument.
file.path <- system.file("extdata/GAIA", package = "goFlux")
import2RData(path = file.path, instrument = "GAIA",
             date.format = "ymd", prec = c(3.5, 0.6, 0.4, 45, 45))

# with Los Gatos Research UGGA (GLA132 series)
file.path <- system.file("extdata/LGR", package = "goFlux")
import2RData(path = file.path, instrument = "LGR",
             date.format = "dmy", keep_all = FALSE,
             prec = c(0.2, 1.4, 50))

# with Los Gatos Research m-GGA (GLA131 series)
file.path <- system.file("extdata/LGR", package = "goFlux")
import2RData(path = file.path, instrument = "LGR",
             date.format = "dmy", keep_all = FALSE,
             prec = c(0.35, 0.9, 200))

# with Los Gatos Research N2OM1 (GLA151 series)
file.path <- system.file("extdata/N2OM1", package = "goFlux")
import2RData(path = file.path, instrument = "N2OM1",
             date.format = "dmy", keep_all = FALSE,
             prec = c(0.5, 2, 50))

# with the LI-COR LI-6400 gas analyzer
file.path <- system.file("extdata/LI6400", package = "goFlux")
import2RData(path = file.path, instrument = "LI-6400",
             date.format = "mdy", keep_all = FALSE,
             prec = c(0.15, 20))

# with the LI-COR LI-7810 gas analyzer
file.path <- system.file("extdata/LI7810", package = "goFlux")
import2RData(path = file.path, instrument = "LI-7810",
             date.format = "ymd", keep_all = FALSE,
             prec = c(3.5, 0.6, 45))

# with the LI-COR LI-7820 gas analyzer
file.path <- system.file("extdata/LI7820", package = "goFlux")
import2RData(path = file.path, instrument = "LI-7820",
             date.format = "ymd", keep_all = FALSE,
             prec = c(0.4, 45))

# with the LI-COR LI-8100 gas analyzer
file.path <- system.file("extdata/LI8100", package = "goFlux")
import2RData(path = file.path, instrument = "LI-8100",
             date.format = "ymd", keep_all = FALSE,
             prec = c(1, 10))

# with the LI-COR smart chamber (LI-8200)
# with this instrument, "keep_all" and "prec" are not valid arguments.
file.path <- system.file("extdata/LI8200", package = "goFlux")
import2RData(path = file.path, instrument = "LI-8200",
             date.format = "ymd")

}
\seealso{
Use the wrapper function \code{\link[goFlux]{import2RData}}
to import multiple files from the same folder path using any instrument.

Import functions for individual instruments:
\code{\link[goFlux]{DX4015_import}},
\code{\link[goFlux]{EGM5_import}},
\code{\link[goFlux]{G2508_import}},
\code{\link[goFlux]{G4301_import}},
\code{\link[goFlux]{GAIA_import}},
\code{\link[goFlux]{LGR_import}},
\code{\link[goFlux]{LI6400_import}},
\code{\link[goFlux]{LI7810_import}},
\code{\link[goFlux]{LI7820_import}},
\code{\link[goFlux]{LI8100_import}},
\code{\link[goFlux]{LI8200_import}},
\code{\link[goFlux]{N2OM1_import}}
}
