% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import2Rdata.R
\name{import2Rdata}
\alias{import2Rdata}
\title{Looping function for quick import of multiple raw gas measurement files}
\usage{
import2Rdata(path, instrument, date.format, timezone = "UTC")
}
\arguments{
\item{path}{Character string. A folder path containing all files to be imported.
Ideally, the folder should only contain the files to be imported.}

\item{instrument}{Character string. Specify which instrument was used to generate
the files contained in the folder path. Chose one of the
following: "LGR", "G2508", "GAIA", "LI-6400", "LI-7810",
"LI-7820", "LI-8100", or "LI-8200".}

\item{date.format}{Date format. Chose one of the following: "dmy", "ymd", or "mdy".}

\item{timezone}{a time zone in which to import the data to POSIXct format.
Default is "UTC". Note about time zone: I recommend using
the time zone "UTC" to avoid any issue related to summer
time and winter time changes.}
}
\value{
a data frame saved as Rdata in a newly created folder, Rdata, into
your working directory.
}
\description{
Imports all raw gas measurement files contained in a folder.
Adapted to multiple greenhouse gas analyzers:
\itemize{
\item LI-COR: LI-6400, LI-7810, LI-7820, LI-8100, LI-8200 (smart chamber)
\item Los Gatos Research instruments: (e.g. UGGA and m-GGA)
\item GAIA2TECH (DMR) automated chamber ECOFlux
\item Picarro: G2508
}
}
\examples{
# Examples on how to use it with all instruments.
# The default time zone "UTC" is used in all cases.

# with Los Gatos Research instruments (e.g. UGGA or m-GGA)
import2Rdata(path = "inst/extdata/LGR", instrument = "LGR", date.format = "dmy")

# with the Picarro instrument G2508
import2Rdata(path = "inst/extdata/G2508", instrument = "G2508", date.format = "ymd")

# with the automated chamber ECOFlux (GAIA2TECH)
import2Rdata(path = "inst/extdata/GAIA", instrument = "GAIA", date.format = "ymd")

# with LI-COR instruments
import2Rdata(path = "inst/extdata/LI6400", instrument = "LI-6400", date.format = "ymd")
import2Rdata(path = "inst/extdata/LI7810", instrument = "LI-7810", date.format = "ymd")
import2Rdata(path = "inst/extdata/LI7820", instrument = "LI-7820", date.format = "ymd")
import2Rdata(path = "inst/extdata/LI8100", instrument = "LI-8100", date.format = "ymd")

# with the LI-COR smart chamber (LI-8200)
# with this instrument, date.format is not a useful parameter.
import2Rdata(path = "inst/extdata/LI8200", instrument = "LI-8200")

}
\seealso{
Use the wraper function \code{\link[GoFluxYourself]{import2Rdata}}
to import multiple files from the same folder path using any instrument.

Import functions for individual instruments:
\code{\link[GoFluxYourself]{G2508_import}},
\code{\link[GoFluxYourself]{GAIA_import}},
\code{\link[GoFluxYourself]{LGR_import}},
\code{\link[GoFluxYourself]{LI6400_import}},
\code{\link[GoFluxYourself]{LI7810_import}},
\code{\link[GoFluxYourself]{LI7820_import}},
\code{\link[GoFluxYourself]{LI8100_import}},
\code{\link[GoFluxYourself]{LI8200_import}}
}
