% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import2RData.R
\name{import2RData}
\alias{import2RData}
\title{Looping function for quick import of multiple raw gas measurement files}
\usage{
import2RData(
  path,
  instrument,
  date.format,
  timezone = "UTC",
  keep_all = FALSE,
  prec,
  proc.data.field = NULL,
  pivot = "long",
  active = TRUE,
  flag = c(7, 11),
  background = FALSE,
  CH.col,
  SWC.col,
  Tsoil.col,
  Tcham.col,
  PAR.col,
  Op.stat.col,
  inst1,
  inst2 = NULL,
  inst3 = NULL,
  gas1,
  gas2 = NULL,
  gas3 = NULL,
  prec1,
  prec2 = NULL,
  prec3 = NULL,
  dry1 = T,
  dry2 = T,
  dry3 = NULL,
  sep = "\\t",
  skip = 1
)
}
\arguments{
\item{path}{character string; a folder path containing all files to be imported.
Ideally, to avoid any errors, the folder should only contain the
files to be imported.}

\item{instrument}{character string; specifies which instrument was used to
generate the files contained in the folder path. Choose one
of the following: "DX4015", "EGM5", "G2508", "G4301" "GAIA",
"GasmetPD", "GT5000", "LI-6400", "LI-7810", "LI-7820",
"LI-8100", "LI-8200", "LI-8250", "N2OM1", "uCH4", "uN2O",
"UGGA". For more information about an instrument, see the
section "See Also" below.}

\item{date.format}{character string; specifies the date format found in the
raw data file. Choose one of the following: "dmy", "ymd",
or "mdy".}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{keep_all}{logical; if \code{keep_all = TRUE}, keep all columns from the raw
file. The default is \code{keep_all = FALSE}, and columns that
are not necessary for gas flux calculation are removed.}

\item{prec}{numerical vector; the precision of the instrument for each gas.
Look at the examples below, or the help for each import function
of each instrument, to know what values to use.}

\item{proc.data.field}{numeric value; select the process data field used with
the EGM-5. See \code{\link[goFlux]{import.EGM5}} for
more details.}

\item{pivot}{character string; either "long" or "wide". If \code{pivot = "long"},
each column containing information about \code{Tsoil},
\code{Tcham}, \code{SWC}, \code{PAR} and operating status
(\code{Op.stat}) will be saved in a single column per parameter.
If \code{pivot = "wide"}, the default display of one column per
chamber per parameter will be used.}

\item{active}{logical; if \code{active = TRUE}, preserve data for active
chambers only.}

\item{flag}{numeric vector; indicates the operating status that should be used
for the flux calculation. Default is \code{flag = c(7,11)}, where
7 indicates "Chamber Idle Closed Clear" and 11 indicates
"Chamber Idle Closed Dark".}

\item{background}{logical; if \code{background = FALSE}, removes all data from
\code{activ.cham == "Background"}.}

\item{Op.stat.col, PAR.col, Tcham.col, Tsoil.col, SWC.col, CH.col}{character string;
a pattern to match the columns that fit the corresponding parameter. See
\code{\link[goFlux]{import.GAIA}} or \code{\link[goFlux]{import.LI8250}}
for more details.}

\item{inst1, inst2, inst3}{character strings; a pattern to match the columns
containing the name of each instrument. See
\code{\link[goFlux]{import.GAIA}} or \code{\link[goFlux]{import.LI8250}}
for more details.}

\item{gas1, gas2, gas3}{character vectors; a pattern to match the columns
containing each gas measurement. See
\code{\link[goFlux]{import.GAIA}} or \code{\link[goFlux]{import.LI8250}}
for more details.}

\item{prec1, prec2, prec3}{numerical vectors; the precision of the instrument
for each gas mentioned in \code{gas1} and \code{gas2}. See
\code{\link[goFlux]{import.GAIA}} or \code{\link[goFlux]{import.LI8250}}
for more details. Note that the order in the arguments \code{precX} must
match the order of the arguments in \code{gasX}.}

\item{dry1, dry2, dry3}{logical; are the gas measurements compensated for water
vapor (dry fraction)? If \code{dryX = TRUE} (default), the gases are
compensated for water vapor, and will be named accordingly. See
\code{\link[goFlux]{import.GAIA}} or \code{\link[goFlux]{import.LI8250}}
for more details.}

\item{sep}{character string defining the field separator character. Values on
each line of the file are separated by this character. By default,
\code{sep = "\t"} for tabulation.}

\item{skip}{integer; the number of lines of the data file to skip before
beginning to read data. By default, \code{skip = 1}.}
}
\value{
A data frame saved as RData in a newly created folder, RData, into
the working directory.
}
\description{
Imports all raw gas measurement files contained in a folder.
Adapted to multiple greenhouse gas analyzers and other instruments:
\itemize{
\item \strong{LI-COR}: LI-6400, LI-7810, LI-7820, LI-8100, LI-8200 (smart chamber),
LI-8250 (Multiplexer)
\item \strong{Los Gatos Research (LGR)}: ultra-portable GGA (GLA132 series),
micro ultra-portable GGA (GLA131 series)
and N2OM1 (GLA151 series)
\item \strong{GAIA2TECH (DMR)}: automated chamber ECOFlux
\item \strong{Picarro}: G2508 and G4301
\item \strong{Gasmet}: DX4015 and GT5000
\item \strong{PP-Systems}: EGM-5
\item \strong{Aeris Technologies}: \ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}}/
\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}}
high accuracy analyzer and
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}}/
\ifelse{html}{\out{C<sub>2</sub>H<sub>6</sub>}}{\eqn{C[2]H[6]}{ASCII}}
natural gas detection system
\item \strong{GasmetPD}: Custom multiplexer from the University of Padova, Italy
}
}
\details{
In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For most instruments, the date is
one of the columns in the raw data file:
\itemize{
\item DX4015: column Date
\item EGM5: column Date
\item G2508: column DATE
\item G4301: column DATE
\item GAIA: column Titles:
\item GasmetPD: column Date
\item GT5000: column Date
\item LI-6400: (see comment below)
\item LI-7810: column DATE
\item LI-7820: column DATE
\item LI-8100: column Date
\item LI-8200: (see comment below)
\item N2OM1: column Time
\item uCH4: column Time Stamp
\item uN2O: column Time Stamp
\item UGGA: column Time
}
For the instrument LI-6400, the date is found in one of the first lines in
a format containing abbreviations, for example "Thr Aug 6 2020", which would
be the date format "mdy". For the instrument LI-8200, the date is found under
one of the measurements, next to '"Date":'. For the instrument LI-8250, the
date format is in the file name.

The argument \code{proc.data.field} is used with the PP-Sytems EGM-5
instrument only. Look up the
\href{https://ppsystems.com/download/technical_manuals/80109-1-EGM-5_Operation_V103.pdf}{EGM-5 Operation Manual}
at page 90 for more details about the different Process Data Fields
(\code{proc.data.field}).

The arguments \code{active}, \code{pivot}, \code{flag}, \code{background},
\code{CH.col}, \code{PAR.col}, \code{Tcham.col}, \code{Op.stat.col},
\code{inst3}, \code{gas3}, \code{prec3}, \code{dry3}, \code{sep} and \code{skip}
are used with the function \code{\link[goFlux]{import.GAIA}} only. The arguments
\code{Tsoil.col}, \code{SWC.col}, \code{inst1}, \code{inst2}, \code{gas1},
\code{gas2}, \code{prec1}, \code{prec2}, \code{dry1}, \code{dry2}, are used
with both the function \code{\link[goFlux]{import.GAIA}} and the function
\code{\link[goFlux]{import.LI8250}}.
}
\examples{

# Examples on how to use it with all instruments.
# The default time zone "UTC" is used in all cases.

# with the Gasmet instrument DX4015
file.path <- system.file("extdata/DX4015", package = "goFlux")
import2RData(path = file.path, instrument = "DX4015",
             date.format = "ymd", prec = c(1.6, 23, 13, 2, 23, 33))

# with the Gasmet instrument GT5000
file.path <- system.file("extdata/GT5000", package = "goFlux")
import2RData(path = file.path, instrument = "GT5000",
             date.format = "ymd", prec = c(1.6, 23, 13, 2, 23, 33))

# with the PP-Systems EGM-5
# The argument proc.data.field is used with the PP-System EGM-5 instrument only.
file.path <- system.file("extdata/EGM5", package = "goFlux")
import2RData(path = file.path, instrument = "EGM5",
             date.format = "dmy", prec = c(3, 1, 500), proc.data.field = 2)

# with the Picarro instrument G2508
file.path <- system.file("extdata/G2508", package = "goFlux")
import2RData(path = file.path, instrument = "G2508",
             date.format = "ymd", prec = c(0.24, 0.3, 5, 0.16, 500))

# with the Picarro instrument G4301
file.path <- system.file("extdata/G4301", package = "goFlux")
import2RData(path = file.path, instrument = "G4301",
             date.format = "ymd", prec = c(0.025, 0.1, 10))

# with the automated chamber ECOFlux (GAIA2TECH)
# with this instrument, "keep_all" is not a valid argument.
file.path <- system.file("extdata/GAIA", package = "goFlux")
import2RData(path = file.path, instrument = "GAIA", date.format = "ymd",
             pivot = "long", active = TRUE, flag = c(7,11),
             background = FALSE,
             CH.col = "COM5A0",
             SWC.col = "1C08_Soil Moisture",
             Tsoil.col = "1C07_Soil Temperature",
             Tcham.col = "2C07_Chamber Temperature",
             PAR.col = "3C07_Sunlight",
             Op.stat.col = "0C06_OperatingStatus",
             inst1 = "XT2C00_Instrument",
             inst2 = "XT3C00_Instrument",
             gas1 = c("XT2C04_CH4", "XT2C05_CO2", "XT2C06_H2O"),
             gas2 = c("XT3C04_N2O", "XT3C05_H2O"),
             prec1 = c(0.6, 3.5, 45),
             prec2 = c(0.4, 45))

# with the custom multiplexer from the University of Padova (GasmetPD)
file.path <- system.file("extdata/GasmetPD", package = "goFlux")
import2RData(path = file.path, instrument = "GasmetPD",
             date.format = "ymd", prec = c(1.6, 13, 2, 23, 33))

# with Los Gatos Research UGGA (GLA132 series)
file.path <- system.file("extdata/UGGA", package = "goFlux")
import2RData(path = file.path, instrument = "UGGA",
             date.format = "dmy", prec = c(0.2, 1.4, 50))

# with Los Gatos Research m-GGA (GLA131 series)
file.path <- system.file("extdata/UGGA", package = "goFlux")
import2RData(path = file.path, instrument = "UGGA",
             date.format = "dmy", prec = c(0.35, 0.9, 200))

# with Los Gatos Research N2OM1 (GLA151 series)
file.path <- system.file("extdata/N2OM1", package = "goFlux")
import2RData(path = file.path, instrument = "N2OM1",
             date.format = "dmy", prec = c(0.5, 2, 50))

# with the LI-COR LI-6400 gas analyzer
file.path <- system.file("extdata/LI6400", package = "goFlux")
import2RData(path = file.path, instrument = "LI-6400",
             date.format = "mdy", prec = c(0.15, 20))

# with the LI-COR LI-7810 gas analyzer
file.path <- system.file("extdata/LI7810", package = "goFlux")
import2RData(path = file.path, instrument = "LI-7810",
             date.format = "ymd", prec = c(3.5, 0.6, 45))

# with the LI-COR LI-7820 gas analyzer
file.path <- system.file("extdata/LI7820", package = "goFlux")
import2RData(path = file.path, instrument = "LI-7820",
             date.format = "ymd", prec = c(0.4, 45))

# with the LI-COR LI-8100 gas analyzer
file.path <- system.file("extdata/LI8100", package = "goFlux")
import2RData(path = file.path, instrument = "LI-8100",
             date.format = "ymd", prec = c(1, 10))

# with the LI-COR smart chamber (LI-8200)
# with this instrument, "keep_all" and "prec" are not valid arguments.
file.path <- system.file("extdata/LI8200", package = "goFlux")
import2RData(path = file.path, instrument = "LI-8200", date.format = "ymd")

# with the LI-COR Multiplexer (LI-8250)
file.path <- system.file("extdata/LI8250", package = "goFlux")
import2RData(path = file.path, instrument = "LI-8250", date.format = "ymd",
             inst1 = "LI-7810", gas1 = c("CO2_DRY", "CH4_DRY", "H2O"),
             prec1 = c(3.5, 0.6, 45), SWC.col = "SWC_1", Tsoil.col = "TS_1")

# with the Aeris Technologies MIRA Ultra CH4/C2H6 analyzer
file.path <- system.file("extdata/uCH4", package = "goFlux")
import2RData(path = file.path, instrument = "uCH4",
             date.format = "mdy", prec = c(1, 0.5, 15))

# with the Aeris Technologies MIRA Ultra N2O/CO2 analyzer
file.path <- system.file("extdata/uN2O", package = "goFlux")
import2RData(path = file.path, instrument = "uN2O",
             date.format = "mdy", prec = c(0.2, 0.2, 15))

}
\seealso{
Use the wrapper function \code{\link[goFlux]{import2RData}}
to import multiple files from the same folder path using any instrument.

Import functions for individual instruments:
\code{\link[goFlux]{import.DX4015}},
\code{\link[goFlux]{import.EGM5}},
\code{\link[goFlux]{import.G2508}},
\code{\link[goFlux]{import.G4301}},
\code{\link[goFlux]{import.GAIA}},
\code{\link[goFlux]{import.GasmetPD}},
\code{\link[goFlux]{import.GT5000}},
\code{\link[goFlux]{import.LI6400}},
\code{\link[goFlux]{import.LI7810}},
\code{\link[goFlux]{import.LI7820}},
\code{\link[goFlux]{import.LI8100}},
\code{\link[goFlux]{import.LI8200}},
\code{\link[goFlux]{import.LI8250}},
\code{\link[goFlux]{import.N2OM1}},
\code{\link[goFlux]{import.uCH4}},
\code{\link[goFlux]{import.uN2O}},
\code{\link[goFlux]{import.UGGA}}
}
