% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.max.R
\name{k.max}
\alias{k.max}
\title{KAPPA MAX}
\usage{
k.max(MDF, LM.flux, t)
}
\arguments{
\item{MDF}{numerical value; minimal detectable flux, calculated with the function
\code{\link[GoFluxYourself]{MDF}}}

\item{LM.flux}{numerical value; flux estimate from a linear flux calculation model,
calculated with the function \code{\link[GoFluxYourself]{LM.flux}}}

\item{t}{numerical value; measurement time (enclosure time) (seconds)}
}
\value{
A numerical value: the kappa-max.
}
\description{
The parameter kappa determines the curvature of the non-linear regression in the
Hutchinson and Mosier model. A maximum threshold for this parameter, kappa-max,
can be calculated based on the minimal detectable flux (MDF), the linear
flux estimate and the measurement time. The unit of the kappa-max is
\ifelse{html}{\out{s<sup>-1</sup>}}{\eqn{s^{-1}}{ASCII}}.
}
\references{
Hüppi et al. (2018). Restricting the nonlinearity parameter in
soil greenhouse gas flux calculation for more reliable flux estimates.
\emph{PloS one}, 13(7), e0200876.
}
\seealso{
Look up the functions \code{\link[GoFluxYourself]{MDF}} and
         \code{\link[GoFluxYourself]{LM.flux}} of this package
         for more information about these parameters.

Kappa max is used in the function \code{\link[GoFluxYourself]{HM.flux}}
         to restrict the maximum allowed curvature in the Hutchinson and
         Mosier model.
}
\keyword{internal}
