% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGR_imp-data.R
\docType{data}
\name{LGR_imp}
\alias{LGR_imp}
\title{Example gas measurements data from LGR instruments}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(LGR_imp)
}
\description{
Raw data from the micro-GGA Los Gatos Research greenhouse gas analyzer that
has been processed by the function \code{\link[GoFluxYourself]{LGR_import}}
of this package.
}
\keyword{datasets}
