% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGR_import.R
\name{LGR_import}
\alias{LGR_import}
\title{Import function for Los Gatos Research GHG analyzers}
\usage{
LGR_import(inputfile, date.format = "dmy", timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .txt}

\item{date.format}{date format; the date format used in the raw data file.
Chose one of the following: "dmy", "ymd", or "mdy". Default
is "dmy", as it is the date format from the example data
file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if save = TRUE, saves the file as RData in a RData folder
in the current working directory. If save = FALSE, returns the file
in the Console, or load in the Environment if assigned to an object.}
}
\value{
a data frame containing raw data from LGR GHG analyzer.
}
\description{
Imports single raw gas measurement files from the ultra-portable GHG analyzers
(UGGA and m-GGA) from Los Gatos Research
(\ifelse{html}{\out{CO<sub>2</sub>}}{\eqn{CO[2]}{ASCII}},
\ifelse{html}{\out{CH<sub>4</sub>}}{\eqn{CH[4]}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}) with the extension .txt
}
\details{
In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For the instrument G2508, the date
is found in the column "Time".
}
\examples{
# Examples on how to use:
file.path <- system.file("extdata", "LGR/example_LGR.txt", package = "GoFluxYourself")

LGR.data <- LGR_import(inputfile = file.path)

}
\seealso{
Use the wrapper function \code{\link[GoFluxYourself]{import2RData}}
         to import multiple files from the same folder path using any instrument.

Import functions for individual instruments:
         \code{\link[GoFluxYourself]{DX4015_import}},
         \code{\link[GoFluxYourself]{G2508_import}},
         \code{\link[GoFluxYourself]{GAIA_import}},
         \code{\link[GoFluxYourself]{LI6400_import}},
         \code{\link[GoFluxYourself]{LI7810_import}},
         \code{\link[GoFluxYourself]{LI7820_import}},
         \code{\link[GoFluxYourself]{LI8100_import}},
         \code{\link[GoFluxYourself]{LI8200_import}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
