% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGR_import.R
\name{LGR_import}
\alias{LGR_import}
\title{Import function for Los Gatos Research GHG analyzers}
\usage{
LGR_import(inputfile, date.format = "dmy", timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{the name of a file with the extension .txt}

\item{date.format}{Date format. Chose one of the following: "dmy", "ymd", or "mdy".
Default is "dmy" as it is the date format from the example
data file provided.}

\item{timezone}{a time zone in which to import the data to POSIXct format.
Default is "UTC". Note about time zone: I recommend using
the time zone "UTC" to avoid any issue related to summer
time and winter time changes.}

\item{save}{logical. If save = TRUE, save the file as Rdata in the current
working directory. If save = FALSE, return the file in the Console,
or load in the Environment if assigned to an object.}
}
\value{
a data frame
}
\description{
Imports single raw gas measurement files from the ultra-portable GHG analyzers
(UGGA and m-GGA) from Los Gatos Research (CO2, CH4 and H2O) with the extension .txt
}
\examples{
# Examples on how to use:
file.path <- system.file("extdata", "LGR/example_LGR.txt", package = "GoFluxYourself")

LGR.data <- LGR_import(inputfile = file.path)

}
