% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI6400_import.R
\name{LI6400_import}
\alias{LI6400_import}
\title{Import function for LI-COR GHG analyzer LI-6400}
\usage{
LI6400_import(inputfile, date.format = "mdy", timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{the name of a file with the extension .txt}

\item{date.format}{Date format. Chose one of the following: "dmy", "ymd", or "mdy".
Default is "mdy" as it is the date format from the example
data file provided.}

\item{timezone}{a time zone in which to import the data to POSIXct format.
Default is "UTC". Note about time zone: I recommend using
the time zone "UTC" to avoid any issue related to summer
time and winter time changes.}

\item{save}{logical. If save = TRUE, save the file as Rdata in a Rdata folder
in the current working directory. If save = FALSE, return the file
in the Console, or load in the Environment if assigned to an object.}
}
\value{
a data frame
}
\description{
Imports single raw gas measurement files from the LI-COR 6400
(CO2 and H2O GHG analyzer)
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "LI6400/example_LI6400.txt", package = "GoFluxYourself")

# Run function
LI6400.data <- LI6400_import(inputfile = file.path)

}
