% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI7810_import.R
\name{LI7810_import}
\alias{LI7810_import}
\title{Import function for LI-COR GHG analyzer LI-7810}
\usage{
LI7810_import(inputfile, date.format = "ymd", timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{the name of a file with the extension .data}

\item{date.format}{Date format. Chose one of the following: "dmy", "ymd", or "mdy".
Default is "ymd" as it is the date format from the example
data file provided.}

\item{timezone}{a time zone in which to import the data to POSIXct format.
Default is "UTC". Note about time zone: I recommend using
the time zone "UTC" to avoid any issue related to summer
time and winter time changes.}

\item{save}{logical. If save = TRUE, save the file as Rdata in a Rdata folder
in the current working directory. If save = FALSE, return the file
in the Console, or load in the Environment if assigned to an object.}
}
\value{
a data frame
}
\description{
Imports single raw gas measurement files from the LI-COR 7810
(CO2, CH4 and H2O GHG analyzer)
}
\examples{
# Examples on how to use:
file.path <- system.file("extdata", "LI7810/example_LI7810.data", package = "GoFluxYourself")

LI7810.data <- LI7810_import(inputfile = file.path)

}
