% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI7820_import.R
\name{LI7820_import}
\alias{LI7820_import}
\title{Import function for LI-COR GHG analyzer LI-7820}
\usage{
LI7820_import(
  inputfile,
  date.format = "ymd",
  timezone = "UTC",
  save = FALSE,
  keep_all = FALSE
)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .data}

\item{date.format}{date format; the date format used in the raw data file.
Chose one of the following: "dmy", "ymd", or "mdy". Default
is "ymd", as it is the date format from the example data
file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if save = TRUE, saves the file as RData in a RData folder
in the current working directory. If save = FALSE, returns the file
in the Console, or load in the Environment if assigned to an object.}

\item{keep_all}{logical; if \code{keep_all = TRUE}, keep all columns from raw
file. The default is \code{keep_all = FALSE}, and columns that
are not necessary for gas flux calculation are removed.}
}
\value{
a data frame containing raw data from LI-COR GHG analyzer LI-7820.
}
\description{
Imports single raw gas measurement files from the LI-COR 7820
(\ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}} and
\ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}} GHG analyzer)
}
\details{
In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For the instrument LI-7820, the
date is found in the column "DATE".

Note that this function was designed for the following default units:
\itemize{
  \item ppm for \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}
  \item ppb for \ifelse{html}{\out{N<sub>2</sub>O}}{\eqn{N[2]O}{ASCII}}
  \item kPa for pressure
  \item Celsius for temperature}
If your instrument uses different units, either convert the units after import,
change the settings on your instrument, or contact the maintainer of this
package for support.
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "LI7820/LI7820.data", package = "GoFluxYourself")

# Run function
LI7820.data <- LI7820_import(inputfile = file.path)

}
\seealso{
Use the wrapper function \code{\link[GoFluxYourself]{import2RData}}
         to import multiple files from the same folder path using any instrument.

Import functions for individual instruments:
         \code{\link[GoFluxYourself]{DX4015_import}},
         \code{\link[GoFluxYourself]{G2508_import}},
         \code{\link[GoFluxYourself]{GAIA_import}},
         \code{\link[GoFluxYourself]{LGR_import}},
         \code{\link[GoFluxYourself]{LI6400_import}},
         \code{\link[GoFluxYourself]{LI7810_import}},
         \code{\link[GoFluxYourself]{LI8100_import}},
         \code{\link[GoFluxYourself]{LI8200_import}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
