% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI8100_import.R
\name{LI8100_import}
\alias{LI8100_import}
\title{Import function for LI-COR GHG analyzer LI-8100}
\usage{
LI8100_import(inputfile, date.format = "ymd", timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .81x}

\item{date.format}{character string; chose one of the following: "dmy", "ymd",
or "mdy". Default is "ymd", as it is the date format from
the example data file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if save = TRUE, saves the file as RData in a RData folder
in the current working directory. If save = FALSE, returns the file
in the Console, or load in the Environment if assigned to an object.}
}
\value{
a data frame
}
\description{
Imports single raw gas measurement files from the LI-COR 8100
(CO2 and H2O GHG analyzer)
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "LI8100/example_LI8100.81x", package = "GoFluxYourself")

# Run function
LI8100.data <- LI8100_import(inputfile = file.path)

}
\seealso{
Use the wrapper function \code{\link[GoFluxYourself]{import2RData}}
         to import multiple files from the same folder path using any instrument.

See also, import functions for other instruments:
         \code{\link[GoFluxYourself]{G2508_import}},
         \code{\link[GoFluxYourself]{GAIA_import}},
         \code{\link[GoFluxYourself]{LGR_import}},
         \code{\link[GoFluxYourself]{LI6400_import}},
         \code{\link[GoFluxYourself]{LI7810_import}},
         \code{\link[GoFluxYourself]{LI7820_import}},
         \code{\link[GoFluxYourself]{LI8200_import}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
