% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI8200_import.R
\name{LI8200_import}
\alias{LI8200_import}
\title{Import function for LI-COR smart chamber (LI-8200)}
\usage{
LI8200_import(inputfile, timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .json}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if save = TRUE, saves the file as RData in a RData folder
in the current working directory. If save = FALSE, returns the file
in the Console, or load in the Environment if assigned to an object.}
}
\value{
a data frame containing raw data from the LI-COR Smart Chamber LI-8200.
}
\description{
Imports single raw data files from the LI-COR smart chamber (LI-8200)
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "LI8200/example_LI8200.json", package = "GoFluxYourself")

# Run function
LI8200.data <- LI8200_import(inputfile = file.path)

}
\seealso{
Use the wrapper function \code{\link[GoFluxYourself]{import2RData}}
         to import multiple files from the same folder path using any instrument.

Import functions for individual instruments:
         \code{\link[GoFluxYourself]{DX4015_import}},
         \code{\link[GoFluxYourself]{G2508_import}},
         \code{\link[GoFluxYourself]{GAIA_import}},
         \code{\link[GoFluxYourself]{LGR_import}},
         \code{\link[GoFluxYourself]{LI6400_import}},
         \code{\link[GoFluxYourself]{LI7810_import}},
         \code{\link[GoFluxYourself]{LI7820_import}},
         \code{\link[GoFluxYourself]{LI8100_import}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
