% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI8200_import.R
\name{LI8200_import}
\alias{LI8200_import}
\title{Import function for LI-COR smart chamber (LI-8200)}
\usage{
LI8200_import(inputfile, timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{the name of a file with the extension .json}

\item{timezone}{a time zone in which to import the data to POSIXct format.
Default is "UTC". Note about time zone: I recommend using
the time zone "UTC" to avoid any issue related to summer
time and winter time changes.}

\item{save}{logical. If save = TRUE, save the file as Rdata in a Rdata folder
in the current working directory. If save = FALSE, return the file
in the Console, or load in the Environment if assigned to an object.}
}
\value{
a data frame
}
\description{
Imports single raw data files from the LI-COR smart chamber (LI-8200)
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "LI8200/example_LI8200.json", package = "GoFluxYourself")

# Run function
LI8200.data <- LI8200_import(inputfile = file.path)

}
\seealso{
Use the wraper function \code{\link[GoFluxYourself]{import2Rdata}}
         to import multiple files from the same folder path using any instrument.

Import functions for individual instruments:
         \code{\link[GoFluxYourself]{G2508_import}},
         \code{\link[GoFluxYourself]{GAIA_import}},
         \code{\link[GoFluxYourself]{LGR_import}},
         \code{\link[GoFluxYourself]{LI6400_import}},
         \code{\link[GoFluxYourself]{LI7810_import}},
         \code{\link[GoFluxYourself]{LI7820_import}},
         \code{\link[GoFluxYourself]{LI8100_import}}
}
