% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LI8200_import.R
\name{LI8200_import}
\alias{LI8200_import}
\title{Import function for LI-COR smart chamber (LI-8200)}
\usage{
LI8200_import(inputfile, date.format = "ymd", timezone = "UTC", save = FALSE)
}
\arguments{
\item{inputfile}{character string; the name of a file with the extension .json}

\item{date.format}{date format; the date format used in the raw data file.
Choose one of the following: "dmy", "ymd", or "mdy".
Default is "ymd", as it is the date format from the
example data file provided.}

\item{timezone}{character string; a time zone in which to import the data to
POSIXct format. Default is "UTC". Note about time zone: it is
recommended to use the time zone "UTC" to avoid any issue
related to summer time and winter time changes.}

\item{save}{logical; if \code{save = TRUE}, saves the file as RData in a
RData folder in the current working directory. If
\code{save = FALSE}, returns the file in the Console, or load in
the Environment if assigned to an object.}
}
\value{
A data frame containing raw data from the LI-COR Smart Chamber LI-8200.
}
\description{
Imports single raw data files from the LI-COR smart chamber (LI-8200)
}
\details{
In \code{date.format}, the date format refers to a date found in the raw data
file, not the date format in the file name. For the instrument LI-8200, the
date is found under one of the measurement, next to "Date":.

Note that this function was designed for the following units in the raw file:
\itemize{
  \item seconds for DeadBand
  \item cm for Area and Offset
  \item mmol/mol for \ifelse{html}{\out{H<sub>2</sub>O}}{\eqn{H[2]O}{ASCII}}
  \item litters for volumes
  \item kPa for pressure
  \item Celsius for temperature}
If your instrument uses different units, either convert the units after import,
change the settings on your instrument, or contact the maintainer of this
package for support.

In addition, soil temperature and moisture are read with a HydraProbe
connected to the Smart Chamber. If you are using a different setup, please
contact the maintainer of this package for support.

As opposed to the other import functions, there is no option to "keep_all" with
the Smart Chamber. If you would like to import additional data using this
function, please contact the maintainer of this package for support.
}
\examples{
# Load file from downloaded package
file.path <- system.file("extdata", "LI8200/LI8200.json", package = "GoFluxYourself")

# Run function
LI8200_imp <- LI8200_import(inputfile = file.path)

}
\seealso{
Use the wrapper function \code{\link[GoFluxYourself]{import2RData}}
         to import multiple files from the same folder path using any instrument.

See also, import functions for other instruments:
         \code{\link[GoFluxYourself]{DX4015_import}},
         \code{\link[GoFluxYourself]{G2508_import}},
         \code{\link[GoFluxYourself]{GAIA_import}},
         \code{\link[GoFluxYourself]{LGR_import}},
         \code{\link[GoFluxYourself]{LI6400_import}},
         \code{\link[GoFluxYourself]{LI7810_import}},
         \code{\link[GoFluxYourself]{LI7820_import}},
         \code{\link[GoFluxYourself]{LI8100_import}}

See \code{\link[base]{timezones}} for a description of the underlying
         timezone attribute.
}
