% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM.flux.R
\name{LM.flux}
\alias{LM.flux}
\title{Linear model for flux calculation}
\usage{
LM.flux(gas.meas, time.meas, flux.term)
}
\arguments{
\item{gas.meas}{numerical vector containing gas measurements (ppm or ppb)}

\item{time.meas}{numerical vector containing time stamps (seconds)}

\item{flux.term}{numerical value; flux term calculated with the function `flux.term()`}
}
\value{
a data.frame
}
\description{
Estimates a gas flux with a linear regression, and then extracts values from
the linear fit, calculates standard error, r~2~, p-value and Root Mean Squared
Error (RMSE).
}
\seealso{
Look up the function \code{\link[GoFluxYourself]{flux.term}} of this
         package for more information about this parameter.

See also the function \code{\link[GoFluxYourself]{HM.flux}}
         for information about the non-linear regression model used in this package.
}
\keyword{internal}
