% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM.flux.R
\name{LM.flux}
\alias{LM.flux}
\title{Linear model for flux calculation}
\usage{
LM.flux(gas.meas, time.meas, flux.term)
}
\arguments{
\item{gas.meas}{numerical vector containing gas measurements (ppm or ppb)}

\item{time.meas}{numerical vector containing time stamps (seconds)}

\item{flux.term}{numerical value; flux term calculated with the function
\code{\link[goFlux]{flux.term}}}
}
\value{
Returns a data frame with 11 columns: linear flux estimate, initial
         gas concentration (\code{LM.C0}), final gas concentration
         (\code{LM.Ct}), slope of linear regression (\code{LM.slope}), mean
         absolute error (\code{LM.MAE}), root mean square error
         (\code{LM.RMSE}), Akaike information criterion corrected for small
         sample size (\code{LM.AICc}), standard error (\code{LM.SE}),
         relative standard error (\code{LM.se.rel}),
         \ifelse{html}{\out{r<sup>2</sup>}}{\eqn{r^2}{ASCII}} (\code{LM.r2}),
         and p-value (\code{LM.p.val}).
}
\description{
Estimates a gas flux with a linear regression, and then extracts values from
the linear fit.
}
\details{
Flux estimate units are
\ifelse{html}{\out{µmol/m<sup>2</sup>s}}{\eqn{µmol/m^{2}s}{ASCII}}
(if initial concentration is ppm, e.g. CO2dry_ppm) and
\ifelse{html}{\out{nmol/m<sup>2</sup>s}}{\eqn{nmol/m^{2}s}{ASCII}}
(if initial concentration is ppb, e.g. CH4dry_ppb).
}
\seealso{
Look up the function \code{\link[goFlux]{flux.term}} of this
         package for more information about this parameter.

See also the function \code{\link[goFlux]{HM.flux}}
         for information about the non-linear regression model used in this package.
}
\keyword{internal}
