% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDF.R
\name{MDF}
\alias{MDF}
\title{Minimal Detectable Flux (MDF)}
\usage{
MDF(p, t, n, flux.term)
}
\arguments{
\item{p}{numerical; precision of the instrument (same units as measured gas; ex. ppm)}

\item{t}{numerical; measurement time (enclosure time; seconds)}

\item{n}{numerical; number of measurements during measurement time}

\item{flux.term}{numerical; flux term calculated with the function flux.term}
}
\value{
a numerical value
}
\description{
The minimal detectable flux based on instrument precision, measurements time,
and the number of measurement points.
}
\keyword{internal}
