% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDF.R
\name{MDF}
\alias{MDF}
\title{Minimal Detectable Flux (MDF)}
\usage{
MDF(p, t, flux.term)
}
\arguments{
\item{p}{numerical; precision of the instrument (same units as measured gas; ex. ppm)}

\item{t}{numerical; measurement time (enclosure time; seconds)}

\item{flux.term}{numerical; flux term calculated with the function `flux.term()`}
}
\value{
a numerical value
}
\description{
The minimal detectable flux based on instrument precision and the
measurement time.
}
\references{
Christiansen et al. (2015). Comparison of CO<sub>2</sub>, CH<sub>4</sub> and N<sub>2</sub>O soil-atmosphere exchange measured in static chambers with cavity ring-down spectroscopy and gas chromatography. *Agricultural and Forest Meteorology*, 211, 48-57.
}
\seealso{
\code{\link[GoFluxYourself]{flux.term}}
}
\keyword{internal}
