% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.win.R
\name{obs.win}
\alias{obs.win}
\title{Define measurements' window for manual identification}
\usage{
obs.win(
  inputfile,
  auxfile = NULL,
  gastype = "CO2dry_ppm",
  obs.length = NULL,
  shoulder = 120
)
}
\arguments{
\item{inputfile}{Character string. The name of a file with the extension .RData
(output from import or align functions).}

\item{auxfile}{Auxiliary data frame containing the columns "start.time" and
"UniqueID". start.time must contain a date and be in POSIXct
format. The time zone must be the same as the POSIX.time in
the inputfile. The default time zone for the import functions
is "UTC". A file from the Smart Chamber (LI-8200) can be
used as an auxiliary file in this case.}

\item{gastype}{Character string. Specify which gas should be displayed on the
plot to manually select start time and end time of measurements.
Must be one of the following: "CO2dry_ppm", "CH4dry_ppb",
"N2Odry_ppb" or "H2O_ppm". Default is "CO2dry_ppm".}

\item{obs.length}{Numerical. Chamber closure time (seconds). Default is NULL.
If obs.length is not provided, a column "obs.length" should
be contained in the auxiliary file.}

\item{shoulder}{Numerical. Time before and after measurement in observation
window (seconds). Default is 120 seconds.}
}
\value{
a list of data frame, split by UniqueID.
}
\description{
Split gas data into one observation window per gas measurement
and create a list of data frame (one data frame per UniqueID).
}
\examples{
# Examples on how to use it in multiple situations:
# Note that gastype = "CO2dry_ppm" is the default setting
library(dplyr)

## with a LGR instrument and an auxiliary file (.txt)
aux.path <- system.file("extdata", "LGR/example_LGR_aux.txt",
                        package = "GoFluxYourself")
auxfile <- read.delim(aux.path) \%>\%
  mutate(start.time = as.POSIXct(start.time, tz = "UTC"))
data(example_LGR_imp)
example_LGR_ow <- obs.win(inputfile = example_LGR_imp, auxfile = auxfile,
                          obs.length = 180, shoulder = 60)

## with a LI-COR instrument and the Smart Chamber as auxiliary file
data(example_LI8200_imp)
data(example_LI7810_imp)
example_LI7810_ow <- obs.win(inputfile = example_LI7810_imp,
                             auxfile = example_LI8200_imp,
                             shoulder = 30)

## with the LI-6400 and no auxiliary file
data(example_LI6400_imp)
example_LI6400_ow <- obs.win(inputfile = example_LI6400_imp, shoulder = 0)

}
